% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_or_delete_rules_request.R
\docType{class}
\name{AddOrDeleteRulesRequest}
\alias{AddOrDeleteRulesRequest}
\title{AddOrDeleteRulesRequest}
\format{
An \code{R6Class} generator object
}
\description{
AddOrDeleteRulesRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of object types defined in the oneOf schema.
Initialize a new AddOrDeleteRulesRequest.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AddOrDeleteRulesRequest-new}{\code{AddOrDeleteRulesRequest$new()}}
\item \href{#method-AddOrDeleteRulesRequest-fromJSON}{\code{AddOrDeleteRulesRequest$fromJSON()}}
\item \href{#method-AddOrDeleteRulesRequest-toJSONString}{\code{AddOrDeleteRulesRequest$toJSONString()}}
\item \href{#method-AddOrDeleteRulesRequest-toJSON}{\code{AddOrDeleteRulesRequest$toJSON()}}
\item \href{#method-AddOrDeleteRulesRequest-validateJSON}{\code{AddOrDeleteRulesRequest$validateJSON()}}
\item \href{#method-AddOrDeleteRulesRequest-toString}{\code{AddOrDeleteRulesRequest$toString()}}
\item \href{#method-AddOrDeleteRulesRequest-clone}{\code{AddOrDeleteRulesRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddOrDeleteRulesRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-AddOrDeleteRulesRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AddOrDeleteRulesRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddOrDeleteRulesRequest$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "AddRulesRequest", "DeleteRulesRequest"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddOrDeleteRulesRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AddOrDeleteRulesRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of AddOrDeleteRulesRequest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddOrDeleteRulesRequest$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of AddOrDeleteRulesRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddOrDeleteRulesRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AddOrDeleteRulesRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize AddOrDeleteRulesRequest to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddOrDeleteRulesRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the AddOrDeleteRulesRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddOrDeleteRulesRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AddOrDeleteRulesRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize AddOrDeleteRulesRequest to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddOrDeleteRulesRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the AddOrDeleteRulesRequest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddOrDeleteRulesRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AddOrDeleteRulesRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to AddOrDeleteRulesRequest and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddOrDeleteRulesRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddOrDeleteRulesRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-AddOrDeleteRulesRequest-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddOrDeleteRulesRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddOrDeleteRulesRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AddOrDeleteRulesRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddOrDeleteRulesRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
