% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rules_request.R
\docType{class}
\name{AddRulesRequest}
\alias{AddRulesRequest}
\title{AddRulesRequest}
\format{
An \code{R6Class} generator object
}
\description{
AddRulesRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{add}}{list( \link{RuleNoId} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AddRulesRequest-new}{\code{AddRulesRequest$new()}}
\item \href{#method-AddRulesRequest-toJSON}{\code{AddRulesRequest$toJSON()}}
\item \href{#method-AddRulesRequest-fromJSON}{\code{AddRulesRequest$fromJSON()}}
\item \href{#method-AddRulesRequest-toJSONString}{\code{AddRulesRequest$toJSONString()}}
\item \href{#method-AddRulesRequest-fromJSONString}{\code{AddRulesRequest$fromJSONString()}}
\item \href{#method-AddRulesRequest-validateJSON}{\code{AddRulesRequest$validateJSON()}}
\item \href{#method-AddRulesRequest-toString}{\code{AddRulesRequest$toString()}}
\item \href{#method-AddRulesRequest-isValid}{\code{AddRulesRequest$isValid()}}
\item \href{#method-AddRulesRequest-getInvalidFields}{\code{AddRulesRequest$getInvalidFields()}}
\item \href{#method-AddRulesRequest-clone}{\code{AddRulesRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddRulesRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-AddRulesRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AddRulesRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddRulesRequest$new(add, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add}}{add}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddRulesRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AddRulesRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddRulesRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AddRulesRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddRulesRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AddRulesRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of AddRulesRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddRulesRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AddRulesRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddRulesRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AddRulesRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddRulesRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AddRulesRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddRulesRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AddRulesRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of AddRulesRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddRulesRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AddRulesRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddRulesRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AddRulesRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to AddRulesRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddRulesRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddRulesRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-AddRulesRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddRulesRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of AddRulesRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddRulesRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-AddRulesRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddRulesRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddRulesRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-AddRulesRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddRulesRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddRulesRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AddRulesRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddRulesRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
