% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animated_gif_all_of.R
\docType{class}
\name{AnimatedGifAllOf}
\alias{AnimatedGifAllOf}
\title{AnimatedGifAllOf}
\format{
An \code{R6Class} generator object
}
\description{
AnimatedGifAllOf Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{preview_image_url}}{character [optional]}

\item{\code{variants}}{list( \link{Variant} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnimatedGifAllOf-new}{\code{AnimatedGifAllOf$new()}}
\item \href{#method-AnimatedGifAllOf-toJSON}{\code{AnimatedGifAllOf$toJSON()}}
\item \href{#method-AnimatedGifAllOf-fromJSON}{\code{AnimatedGifAllOf$fromJSON()}}
\item \href{#method-AnimatedGifAllOf-toJSONString}{\code{AnimatedGifAllOf$toJSONString()}}
\item \href{#method-AnimatedGifAllOf-fromJSONString}{\code{AnimatedGifAllOf$fromJSONString()}}
\item \href{#method-AnimatedGifAllOf-validateJSON}{\code{AnimatedGifAllOf$validateJSON()}}
\item \href{#method-AnimatedGifAllOf-toString}{\code{AnimatedGifAllOf$toString()}}
\item \href{#method-AnimatedGifAllOf-isValid}{\code{AnimatedGifAllOf$isValid()}}
\item \href{#method-AnimatedGifAllOf-getInvalidFields}{\code{AnimatedGifAllOf$getInvalidFields()}}
\item \href{#method-AnimatedGifAllOf-clone}{\code{AnimatedGifAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnimatedGifAllOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnimatedGifAllOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AnimatedGifAllOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnimatedGifAllOf$new(preview_image_url = NULL, variants = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{preview_image_url}}{preview_image_url}

\item{\code{variants}}{An array of all available variants of the media.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnimatedGifAllOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnimatedGifAllOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnimatedGifAllOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AnimatedGifAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnimatedGifAllOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnimatedGifAllOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of AnimatedGifAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnimatedGifAllOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AnimatedGifAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnimatedGifAllOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnimatedGifAllOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnimatedGifAllOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AnimatedGifAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnimatedGifAllOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnimatedGifAllOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of AnimatedGifAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnimatedGifAllOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AnimatedGifAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnimatedGifAllOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnimatedGifAllOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to AnimatedGifAllOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnimatedGifAllOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnimatedGifAllOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-AnimatedGifAllOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnimatedGifAllOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of AnimatedGifAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnimatedGifAllOf-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-AnimatedGifAllOf-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnimatedGifAllOf$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnimatedGifAllOf-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-AnimatedGifAllOf-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnimatedGifAllOf$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnimatedGifAllOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnimatedGifAllOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnimatedGifAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
