% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_user_mutation_response.R
\docType{class}
\name{BlockUserMutationResponse}
\alias{BlockUserMutationResponse}
\title{BlockUserMutationResponse}
\format{
An \code{R6Class} generator object
}
\description{
BlockUserMutationResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{BlockUserMutationResponseData} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BlockUserMutationResponse-new}{\code{BlockUserMutationResponse$new()}}
\item \href{#method-BlockUserMutationResponse-toJSON}{\code{BlockUserMutationResponse$toJSON()}}
\item \href{#method-BlockUserMutationResponse-fromJSON}{\code{BlockUserMutationResponse$fromJSON()}}
\item \href{#method-BlockUserMutationResponse-toJSONString}{\code{BlockUserMutationResponse$toJSONString()}}
\item \href{#method-BlockUserMutationResponse-fromJSONString}{\code{BlockUserMutationResponse$fromJSONString()}}
\item \href{#method-BlockUserMutationResponse-validateJSON}{\code{BlockUserMutationResponse$validateJSON()}}
\item \href{#method-BlockUserMutationResponse-toString}{\code{BlockUserMutationResponse$toString()}}
\item \href{#method-BlockUserMutationResponse-isValid}{\code{BlockUserMutationResponse$isValid()}}
\item \href{#method-BlockUserMutationResponse-getInvalidFields}{\code{BlockUserMutationResponse$getInvalidFields()}}
\item \href{#method-BlockUserMutationResponse-clone}{\code{BlockUserMutationResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new BlockUserMutationResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BlockUserMutationResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of BlockUserMutationResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BlockUserMutationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BlockUserMutationResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of BlockUserMutationResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BlockUserMutationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to BlockUserMutationResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of BlockUserMutationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
