% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_user_mutation_response_data.R
\docType{class}
\name{BlockUserMutationResponseData}
\alias{BlockUserMutationResponseData}
\title{BlockUserMutationResponseData}
\format{
An \code{R6Class} generator object
}
\description{
BlockUserMutationResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{blocking}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BlockUserMutationResponseData-new}{\code{BlockUserMutationResponseData$new()}}
\item \href{#method-BlockUserMutationResponseData-toJSON}{\code{BlockUserMutationResponseData$toJSON()}}
\item \href{#method-BlockUserMutationResponseData-fromJSON}{\code{BlockUserMutationResponseData$fromJSON()}}
\item \href{#method-BlockUserMutationResponseData-toJSONString}{\code{BlockUserMutationResponseData$toJSONString()}}
\item \href{#method-BlockUserMutationResponseData-fromJSONString}{\code{BlockUserMutationResponseData$fromJSONString()}}
\item \href{#method-BlockUserMutationResponseData-validateJSON}{\code{BlockUserMutationResponseData$validateJSON()}}
\item \href{#method-BlockUserMutationResponseData-toString}{\code{BlockUserMutationResponseData$toString()}}
\item \href{#method-BlockUserMutationResponseData-isValid}{\code{BlockUserMutationResponseData$isValid()}}
\item \href{#method-BlockUserMutationResponseData-getInvalidFields}{\code{BlockUserMutationResponseData$getInvalidFields()}}
\item \href{#method-BlockUserMutationResponseData-clone}{\code{BlockUserMutationResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new BlockUserMutationResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponseData$new(blocking = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blocking}}{blocking}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BlockUserMutationResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of BlockUserMutationResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BlockUserMutationResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BlockUserMutationResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of BlockUserMutationResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BlockUserMutationResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to BlockUserMutationResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of BlockUserMutationResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserMutationResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserMutationResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserMutationResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
