% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_user_request.R
\docType{class}
\name{BlockUserRequest}
\alias{BlockUserRequest}
\title{BlockUserRequest}
\format{
An \code{R6Class} generator object
}
\description{
BlockUserRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{target_user_id}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BlockUserRequest-new}{\code{BlockUserRequest$new()}}
\item \href{#method-BlockUserRequest-toJSON}{\code{BlockUserRequest$toJSON()}}
\item \href{#method-BlockUserRequest-fromJSON}{\code{BlockUserRequest$fromJSON()}}
\item \href{#method-BlockUserRequest-toJSONString}{\code{BlockUserRequest$toJSONString()}}
\item \href{#method-BlockUserRequest-fromJSONString}{\code{BlockUserRequest$fromJSONString()}}
\item \href{#method-BlockUserRequest-validateJSON}{\code{BlockUserRequest$validateJSON()}}
\item \href{#method-BlockUserRequest-toString}{\code{BlockUserRequest$toString()}}
\item \href{#method-BlockUserRequest-isValid}{\code{BlockUserRequest$isValid()}}
\item \href{#method-BlockUserRequest-getInvalidFields}{\code{BlockUserRequest$getInvalidFields()}}
\item \href{#method-BlockUserRequest-clone}{\code{BlockUserRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new BlockUserRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserRequest$new(target_user_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_user_id}}{Unique identifier of this User. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BlockUserRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of BlockUserRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BlockUserRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BlockUserRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of BlockUserRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BlockUserRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to BlockUserRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of BlockUserRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BlockUserRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BlockUserRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BlockUserRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
