% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark_add_request.R
\docType{class}
\name{BookmarkAddRequest}
\alias{BookmarkAddRequest}
\title{BookmarkAddRequest}
\format{
An \code{R6Class} generator object
}
\description{
BookmarkAddRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tweet_id}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BookmarkAddRequest-new}{\code{BookmarkAddRequest$new()}}
\item \href{#method-BookmarkAddRequest-toJSON}{\code{BookmarkAddRequest$toJSON()}}
\item \href{#method-BookmarkAddRequest-fromJSON}{\code{BookmarkAddRequest$fromJSON()}}
\item \href{#method-BookmarkAddRequest-toJSONString}{\code{BookmarkAddRequest$toJSONString()}}
\item \href{#method-BookmarkAddRequest-fromJSONString}{\code{BookmarkAddRequest$fromJSONString()}}
\item \href{#method-BookmarkAddRequest-validateJSON}{\code{BookmarkAddRequest$validateJSON()}}
\item \href{#method-BookmarkAddRequest-toString}{\code{BookmarkAddRequest$toString()}}
\item \href{#method-BookmarkAddRequest-isValid}{\code{BookmarkAddRequest$isValid()}}
\item \href{#method-BookmarkAddRequest-getInvalidFields}{\code{BookmarkAddRequest$getInvalidFields()}}
\item \href{#method-BookmarkAddRequest-clone}{\code{BookmarkAddRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkAddRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkAddRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new BookmarkAddRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkAddRequest$new(tweet_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tweet_id}}{Unique identifier of this Tweet. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkAddRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkAddRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkAddRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BookmarkAddRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkAddRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkAddRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of BookmarkAddRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkAddRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BookmarkAddRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkAddRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkAddRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkAddRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BookmarkAddRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkAddRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkAddRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of BookmarkAddRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkAddRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BookmarkAddRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkAddRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkAddRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to BookmarkAddRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkAddRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkAddRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkAddRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkAddRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of BookmarkAddRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkAddRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkAddRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkAddRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkAddRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkAddRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkAddRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkAddRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkAddRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkAddRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
