% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark_mutation_response.R
\docType{class}
\name{BookmarkMutationResponse}
\alias{BookmarkMutationResponse}
\title{BookmarkMutationResponse}
\format{
An \code{R6Class} generator object
}
\description{
BookmarkMutationResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{BookmarkMutationResponseData} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BookmarkMutationResponse-new}{\code{BookmarkMutationResponse$new()}}
\item \href{#method-BookmarkMutationResponse-toJSON}{\code{BookmarkMutationResponse$toJSON()}}
\item \href{#method-BookmarkMutationResponse-fromJSON}{\code{BookmarkMutationResponse$fromJSON()}}
\item \href{#method-BookmarkMutationResponse-toJSONString}{\code{BookmarkMutationResponse$toJSONString()}}
\item \href{#method-BookmarkMutationResponse-fromJSONString}{\code{BookmarkMutationResponse$fromJSONString()}}
\item \href{#method-BookmarkMutationResponse-validateJSON}{\code{BookmarkMutationResponse$validateJSON()}}
\item \href{#method-BookmarkMutationResponse-toString}{\code{BookmarkMutationResponse$toString()}}
\item \href{#method-BookmarkMutationResponse-isValid}{\code{BookmarkMutationResponse$isValid()}}
\item \href{#method-BookmarkMutationResponse-getInvalidFields}{\code{BookmarkMutationResponse$getInvalidFields()}}
\item \href{#method-BookmarkMutationResponse-clone}{\code{BookmarkMutationResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new BookmarkMutationResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BookmarkMutationResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of BookmarkMutationResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BookmarkMutationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BookmarkMutationResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of BookmarkMutationResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BookmarkMutationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to BookmarkMutationResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of BookmarkMutationResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
