% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark_mutation_response_data.R
\docType{class}
\name{BookmarkMutationResponseData}
\alias{BookmarkMutationResponseData}
\title{BookmarkMutationResponseData}
\format{
An \code{R6Class} generator object
}
\description{
BookmarkMutationResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bookmarked}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BookmarkMutationResponseData-new}{\code{BookmarkMutationResponseData$new()}}
\item \href{#method-BookmarkMutationResponseData-toJSON}{\code{BookmarkMutationResponseData$toJSON()}}
\item \href{#method-BookmarkMutationResponseData-fromJSON}{\code{BookmarkMutationResponseData$fromJSON()}}
\item \href{#method-BookmarkMutationResponseData-toJSONString}{\code{BookmarkMutationResponseData$toJSONString()}}
\item \href{#method-BookmarkMutationResponseData-fromJSONString}{\code{BookmarkMutationResponseData$fromJSONString()}}
\item \href{#method-BookmarkMutationResponseData-validateJSON}{\code{BookmarkMutationResponseData$validateJSON()}}
\item \href{#method-BookmarkMutationResponseData-toString}{\code{BookmarkMutationResponseData$toString()}}
\item \href{#method-BookmarkMutationResponseData-isValid}{\code{BookmarkMutationResponseData$isValid()}}
\item \href{#method-BookmarkMutationResponseData-getInvalidFields}{\code{BookmarkMutationResponseData$getInvalidFields()}}
\item \href{#method-BookmarkMutationResponseData-clone}{\code{BookmarkMutationResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new BookmarkMutationResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponseData$new(bookmarked = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bookmarked}}{bookmarked}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BookmarkMutationResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of BookmarkMutationResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BookmarkMutationResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BookmarkMutationResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of BookmarkMutationResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BookmarkMutationResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to BookmarkMutationResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of BookmarkMutationResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BookmarkMutationResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BookmarkMutationResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BookmarkMutationResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
