% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashtag_entity.R
\docType{class}
\name{CashtagEntity}
\alias{CashtagEntity}
\title{CashtagEntity}
\format{
An \code{R6Class} generator object
}
\description{
CashtagEntity Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{end}}{integer}

\item{\code{start}}{integer}

\item{\code{tag}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CashtagEntity-new}{\code{CashtagEntity$new()}}
\item \href{#method-CashtagEntity-toJSON}{\code{CashtagEntity$toJSON()}}
\item \href{#method-CashtagEntity-fromJSON}{\code{CashtagEntity$fromJSON()}}
\item \href{#method-CashtagEntity-toJSONString}{\code{CashtagEntity$toJSONString()}}
\item \href{#method-CashtagEntity-fromJSONString}{\code{CashtagEntity$fromJSONString()}}
\item \href{#method-CashtagEntity-validateJSON}{\code{CashtagEntity$validateJSON()}}
\item \href{#method-CashtagEntity-toString}{\code{CashtagEntity$toString()}}
\item \href{#method-CashtagEntity-isValid}{\code{CashtagEntity$isValid()}}
\item \href{#method-CashtagEntity-getInvalidFields}{\code{CashtagEntity$getInvalidFields()}}
\item \href{#method-CashtagEntity-clone}{\code{CashtagEntity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagEntity-new"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagEntity-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CashtagEntity class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagEntity$new(end, start, tag, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{end}}{Index (zero-based) at which position this entity ends.  The index is exclusive.}

\item{\code{start}}{Index (zero-based) at which position this entity starts.  The index is inclusive.}

\item{\code{tag}}{tag}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagEntity-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagEntity-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagEntity$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CashtagEntity in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagEntity-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagEntity-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of CashtagEntity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagEntity$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CashtagEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagEntity-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagEntity-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagEntity$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CashtagEntity in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagEntity-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagEntity-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of CashtagEntity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagEntity$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CashtagEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagEntity-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagEntity-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to CashtagEntity and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagEntity$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagEntity-toString"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagEntity-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagEntity$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of CashtagEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagEntity-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagEntity-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagEntity$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagEntity-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagEntity-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagEntity$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagEntity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagEntity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagEntity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
