% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashtag_fields.R
\docType{class}
\name{CashtagFields}
\alias{CashtagFields}
\title{CashtagFields}
\format{
An \code{R6Class} generator object
}
\description{
CashtagFields Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tag}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CashtagFields-new}{\code{CashtagFields$new()}}
\item \href{#method-CashtagFields-toJSON}{\code{CashtagFields$toJSON()}}
\item \href{#method-CashtagFields-fromJSON}{\code{CashtagFields$fromJSON()}}
\item \href{#method-CashtagFields-toJSONString}{\code{CashtagFields$toJSONString()}}
\item \href{#method-CashtagFields-fromJSONString}{\code{CashtagFields$fromJSONString()}}
\item \href{#method-CashtagFields-validateJSON}{\code{CashtagFields$validateJSON()}}
\item \href{#method-CashtagFields-toString}{\code{CashtagFields$toString()}}
\item \href{#method-CashtagFields-isValid}{\code{CashtagFields$isValid()}}
\item \href{#method-CashtagFields-getInvalidFields}{\code{CashtagFields$getInvalidFields()}}
\item \href{#method-CashtagFields-clone}{\code{CashtagFields$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagFields-new"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagFields-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CashtagFields class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagFields$new(tag, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag}}{tag}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagFields-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagFields-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagFields$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CashtagFields in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagFields-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagFields-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of CashtagFields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagFields$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CashtagFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagFields-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagFields-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagFields$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CashtagFields in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagFields-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagFields-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of CashtagFields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagFields$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CashtagFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagFields-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagFields-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to CashtagFields and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagFields$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagFields-toString"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagFields-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagFields$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of CashtagFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagFields-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagFields-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagFields$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagFields-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagFields-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagFields$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CashtagFields-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CashtagFields-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CashtagFields$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
