% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_forbidden_problem_all_of.R
\docType{class}
\name{ClientForbiddenProblemAllOf}
\alias{ClientForbiddenProblemAllOf}
\title{ClientForbiddenProblemAllOf}
\format{
An \code{R6Class} generator object
}
\description{
ClientForbiddenProblemAllOf Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{reason}}{character [optional]}

\item{\code{registration_url}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClientForbiddenProblemAllOf-new}{\code{ClientForbiddenProblemAllOf$new()}}
\item \href{#method-ClientForbiddenProblemAllOf-toJSON}{\code{ClientForbiddenProblemAllOf$toJSON()}}
\item \href{#method-ClientForbiddenProblemAllOf-fromJSON}{\code{ClientForbiddenProblemAllOf$fromJSON()}}
\item \href{#method-ClientForbiddenProblemAllOf-toJSONString}{\code{ClientForbiddenProblemAllOf$toJSONString()}}
\item \href{#method-ClientForbiddenProblemAllOf-fromJSONString}{\code{ClientForbiddenProblemAllOf$fromJSONString()}}
\item \href{#method-ClientForbiddenProblemAllOf-validateJSON}{\code{ClientForbiddenProblemAllOf$validateJSON()}}
\item \href{#method-ClientForbiddenProblemAllOf-toString}{\code{ClientForbiddenProblemAllOf$toString()}}
\item \href{#method-ClientForbiddenProblemAllOf-isValid}{\code{ClientForbiddenProblemAllOf$isValid()}}
\item \href{#method-ClientForbiddenProblemAllOf-getInvalidFields}{\code{ClientForbiddenProblemAllOf$getInvalidFields()}}
\item \href{#method-ClientForbiddenProblemAllOf-clone}{\code{ClientForbiddenProblemAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientForbiddenProblemAllOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClientForbiddenProblemAllOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ClientForbiddenProblemAllOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientForbiddenProblemAllOf$new(reason = NULL, registration_url = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reason}}{reason}

\item{\code{registration_url}}{registration_url}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientForbiddenProblemAllOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ClientForbiddenProblemAllOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientForbiddenProblemAllOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ClientForbiddenProblemAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientForbiddenProblemAllOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ClientForbiddenProblemAllOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ClientForbiddenProblemAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientForbiddenProblemAllOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ClientForbiddenProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientForbiddenProblemAllOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ClientForbiddenProblemAllOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientForbiddenProblemAllOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ClientForbiddenProblemAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientForbiddenProblemAllOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ClientForbiddenProblemAllOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ClientForbiddenProblemAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientForbiddenProblemAllOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ClientForbiddenProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientForbiddenProblemAllOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ClientForbiddenProblemAllOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ClientForbiddenProblemAllOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientForbiddenProblemAllOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientForbiddenProblemAllOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ClientForbiddenProblemAllOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientForbiddenProblemAllOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ClientForbiddenProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientForbiddenProblemAllOf-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ClientForbiddenProblemAllOf-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientForbiddenProblemAllOf$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientForbiddenProblemAllOf-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ClientForbiddenProblemAllOf-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientForbiddenProblemAllOf$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientForbiddenProblemAllOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClientForbiddenProblemAllOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientForbiddenProblemAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
