% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_job.R
\docType{class}
\name{ComplianceJob}
\alias{ComplianceJob}
\title{ComplianceJob}
\format{
An \code{R6Class} generator object
}
\description{
ComplianceJob Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{created_at}}{character}

\item{\code{download_expires_at}}{character}

\item{\code{download_url}}{character}

\item{\code{id}}{character}

\item{\code{name}}{character [optional]}

\item{\code{status}}{\link{ComplianceJobStatus}}

\item{\code{type}}{\link{ComplianceJobType}}

\item{\code{upload_expires_at}}{character}

\item{\code{upload_url}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ComplianceJob-new}{\code{ComplianceJob$new()}}
\item \href{#method-ComplianceJob-toJSON}{\code{ComplianceJob$toJSON()}}
\item \href{#method-ComplianceJob-fromJSON}{\code{ComplianceJob$fromJSON()}}
\item \href{#method-ComplianceJob-toJSONString}{\code{ComplianceJob$toJSONString()}}
\item \href{#method-ComplianceJob-fromJSONString}{\code{ComplianceJob$fromJSONString()}}
\item \href{#method-ComplianceJob-validateJSON}{\code{ComplianceJob$validateJSON()}}
\item \href{#method-ComplianceJob-toString}{\code{ComplianceJob$toString()}}
\item \href{#method-ComplianceJob-isValid}{\code{ComplianceJob$isValid()}}
\item \href{#method-ComplianceJob-getInvalidFields}{\code{ComplianceJob$getInvalidFields()}}
\item \href{#method-ComplianceJob-clone}{\code{ComplianceJob$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJob-new"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJob-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ComplianceJob class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJob$new(
  created_at,
  download_expires_at,
  download_url,
  id,
  status,
  type,
  upload_expires_at,
  upload_url,
  name = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{created_at}}{Creation time of the compliance job.}

\item{\code{download_expires_at}}{Expiration time of the download URL.}

\item{\code{download_url}}{URL from which the user will retrieve their compliance results.}

\item{\code{id}}{Compliance Job ID.}

\item{\code{status}}{status}

\item{\code{type}}{type}

\item{\code{upload_expires_at}}{Expiration time of the upload URL.}

\item{\code{upload_url}}{URL to which the user will upload their Tweet or user IDs.}

\item{\code{name}}{User-provided name for a compliance job.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJob-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJob-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJob$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ComplianceJob in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJob-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJob-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ComplianceJob
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJob$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ComplianceJob
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJob-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJob-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJob$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ComplianceJob in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJob-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJob-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ComplianceJob
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJob$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ComplianceJob
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJob-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJob-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ComplianceJob and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJob$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJob-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJob-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJob$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ComplianceJob
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJob-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJob-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJob$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJob-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJob-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJob$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJob-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJob-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJob$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
