% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_job_status.R
\docType{class}
\name{ComplianceJobStatus}
\alias{ComplianceJobStatus}
\title{ComplianceJobStatus}
\format{
An \code{R6Class} generator object
}
\description{
ComplianceJobStatus Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ComplianceJobStatus-new}{\code{ComplianceJobStatus$new()}}
\item \href{#method-ComplianceJobStatus-toJSON}{\code{ComplianceJobStatus$toJSON()}}
\item \href{#method-ComplianceJobStatus-fromJSON}{\code{ComplianceJobStatus$fromJSON()}}
\item \href{#method-ComplianceJobStatus-toJSONString}{\code{ComplianceJobStatus$toJSONString()}}
\item \href{#method-ComplianceJobStatus-fromJSONString}{\code{ComplianceJobStatus$fromJSONString()}}
\item \href{#method-ComplianceJobStatus-clone}{\code{ComplianceJobStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJobStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJobStatus-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ComplianceJobStatus class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJobStatus$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJobStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJobStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJobStatus$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ComplianceJobStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJobStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJobStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ComplianceJobStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJobStatus$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ComplianceJobStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJobStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJobStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJobStatus$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ComplianceJobStatus in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJobStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJobStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ComplianceJobStatus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJobStatus$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ComplianceJobStatus
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJobStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJobStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJobStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
