% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_job_type.R
\docType{class}
\name{ComplianceJobType}
\alias{ComplianceJobType}
\title{ComplianceJobType}
\format{
An \code{R6Class} generator object
}
\description{
ComplianceJobType Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ComplianceJobType-new}{\code{ComplianceJobType$new()}}
\item \href{#method-ComplianceJobType-toJSON}{\code{ComplianceJobType$toJSON()}}
\item \href{#method-ComplianceJobType-fromJSON}{\code{ComplianceJobType$fromJSON()}}
\item \href{#method-ComplianceJobType-toJSONString}{\code{ComplianceJobType$toJSONString()}}
\item \href{#method-ComplianceJobType-fromJSONString}{\code{ComplianceJobType$fromJSONString()}}
\item \href{#method-ComplianceJobType-clone}{\code{ComplianceJobType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJobType-new"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJobType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ComplianceJobType class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJobType$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJobType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJobType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJobType$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ComplianceJobType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJobType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJobType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ComplianceJobType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJobType$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ComplianceJobType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJobType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJobType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJobType$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ComplianceJobType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJobType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJobType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ComplianceJobType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJobType$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ComplianceJobType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceJobType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceJobType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceJobType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
