% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflict_problem.R
\docType{class}
\name{ConflictProblem}
\alias{ConflictProblem}
\title{ConflictProblem}
\format{
An \code{R6Class} generator object
}
\description{
ConflictProblem Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Super class}{
\code{\link[twitter:Problem]{twitter::Problem}} -> \code{ConflictProblem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detail}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character}

\item{\code{type}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConflictProblem-new}{\code{ConflictProblem$new()}}
\item \href{#method-ConflictProblem-toJSON}{\code{ConflictProblem$toJSON()}}
\item \href{#method-ConflictProblem-fromJSON}{\code{ConflictProblem$fromJSON()}}
\item \href{#method-ConflictProblem-toJSONString}{\code{ConflictProblem$toJSONString()}}
\item \href{#method-ConflictProblem-fromJSONString}{\code{ConflictProblem$fromJSONString()}}
\item \href{#method-ConflictProblem-validateJSON}{\code{ConflictProblem$validateJSON()}}
\item \href{#method-ConflictProblem-toString}{\code{ConflictProblem$toString()}}
\item \href{#method-ConflictProblem-isValid}{\code{ConflictProblem$isValid()}}
\item \href{#method-ConflictProblem-getInvalidFields}{\code{ConflictProblem$getInvalidFields()}}
\item \href{#method-ConflictProblem-clone}{\code{ConflictProblem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConflictProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConflictProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ConflictProblem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConflictProblem$new(title, type, detail = NULL, status = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{type}}{type}

\item{\code{detail}}{detail}

\item{\code{status}}{status}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConflictProblem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConflictProblem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConflictProblem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ConflictProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConflictProblem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConflictProblem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ConflictProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConflictProblem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ConflictProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConflictProblem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ConflictProblem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConflictProblem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ConflictProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConflictProblem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ConflictProblem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ConflictProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConflictProblem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ConflictProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConflictProblem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConflictProblem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ConflictProblem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConflictProblem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConflictProblem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ConflictProblem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConflictProblem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ConflictProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConflictProblem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ConflictProblem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConflictProblem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConflictProblem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ConflictProblem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConflictProblem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConflictProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConflictProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConflictProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
