% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_exception_problem.R
\docType{class}
\name{ConnectionExceptionProblem}
\alias{ConnectionExceptionProblem}
\title{ConnectionExceptionProblem}
\format{
An \code{R6Class} generator object
}
\description{
ConnectionExceptionProblem Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Super class}{
\code{\link[twitter:Problem]{twitter::Problem}} -> \code{ConnectionExceptionProblem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detail}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character}

\item{\code{type}}{character}

\item{\code{connection_issue}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConnectionExceptionProblem-new}{\code{ConnectionExceptionProblem$new()}}
\item \href{#method-ConnectionExceptionProblem-toJSON}{\code{ConnectionExceptionProblem$toJSON()}}
\item \href{#method-ConnectionExceptionProblem-fromJSON}{\code{ConnectionExceptionProblem$fromJSON()}}
\item \href{#method-ConnectionExceptionProblem-toJSONString}{\code{ConnectionExceptionProblem$toJSONString()}}
\item \href{#method-ConnectionExceptionProblem-fromJSONString}{\code{ConnectionExceptionProblem$fromJSONString()}}
\item \href{#method-ConnectionExceptionProblem-validateJSON}{\code{ConnectionExceptionProblem$validateJSON()}}
\item \href{#method-ConnectionExceptionProblem-toString}{\code{ConnectionExceptionProblem$toString()}}
\item \href{#method-ConnectionExceptionProblem-isValid}{\code{ConnectionExceptionProblem$isValid()}}
\item \href{#method-ConnectionExceptionProblem-getInvalidFields}{\code{ConnectionExceptionProblem$getInvalidFields()}}
\item \href{#method-ConnectionExceptionProblem-clone}{\code{ConnectionExceptionProblem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionExceptionProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionExceptionProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ConnectionExceptionProblem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionExceptionProblem$new(
  title,
  type,
  detail = NULL,
  status = NULL,
  connection_issue = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{type}}{type}

\item{\code{detail}}{detail}

\item{\code{status}}{status}

\item{\code{connection_issue}}{connection_issue}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionExceptionProblem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionExceptionProblem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionExceptionProblem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ConnectionExceptionProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionExceptionProblem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionExceptionProblem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ConnectionExceptionProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionExceptionProblem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ConnectionExceptionProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionExceptionProblem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionExceptionProblem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionExceptionProblem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ConnectionExceptionProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionExceptionProblem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionExceptionProblem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ConnectionExceptionProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionExceptionProblem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ConnectionExceptionProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionExceptionProblem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionExceptionProblem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ConnectionExceptionProblem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionExceptionProblem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionExceptionProblem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionExceptionProblem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionExceptionProblem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ConnectionExceptionProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionExceptionProblem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionExceptionProblem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionExceptionProblem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionExceptionProblem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionExceptionProblem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionExceptionProblem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionExceptionProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionExceptionProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionExceptionProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
