% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_annotation.R
\docType{class}
\name{ContextAnnotation}
\alias{ContextAnnotation}
\title{ContextAnnotation}
\format{
An \code{R6Class} generator object
}
\description{
ContextAnnotation Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{domain}}{\link{ContextAnnotationDomainFields}}

\item{\code{entity}}{\link{ContextAnnotationEntityFields}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ContextAnnotation-new}{\code{ContextAnnotation$new()}}
\item \href{#method-ContextAnnotation-toJSON}{\code{ContextAnnotation$toJSON()}}
\item \href{#method-ContextAnnotation-fromJSON}{\code{ContextAnnotation$fromJSON()}}
\item \href{#method-ContextAnnotation-toJSONString}{\code{ContextAnnotation$toJSONString()}}
\item \href{#method-ContextAnnotation-fromJSONString}{\code{ContextAnnotation$fromJSONString()}}
\item \href{#method-ContextAnnotation-validateJSON}{\code{ContextAnnotation$validateJSON()}}
\item \href{#method-ContextAnnotation-toString}{\code{ContextAnnotation$toString()}}
\item \href{#method-ContextAnnotation-isValid}{\code{ContextAnnotation$isValid()}}
\item \href{#method-ContextAnnotation-getInvalidFields}{\code{ContextAnnotation$getInvalidFields()}}
\item \href{#method-ContextAnnotation-clone}{\code{ContextAnnotation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnnotation-new"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnnotation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ContextAnnotation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnnotation$new(domain, entity, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{domain}}{domain}

\item{\code{entity}}{entity}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnnotation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnnotation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnnotation$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ContextAnnotation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnnotation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnnotation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ContextAnnotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnnotation$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ContextAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnnotation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnnotation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnnotation$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ContextAnnotation in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnnotation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnnotation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ContextAnnotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnnotation$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ContextAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnnotation-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnnotation-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ContextAnnotation and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnnotation$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnnotation-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnnotation-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnnotation$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ContextAnnotation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnnotation-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnnotation-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnnotation$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnnotation-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnnotation-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnnotation$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextAnnotation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ContextAnnotation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextAnnotation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
