% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_compliance_job_request.R
\docType{class}
\name{CreateComplianceJobRequest}
\alias{CreateComplianceJobRequest}
\title{CreateComplianceJobRequest}
\format{
An \code{R6Class} generator object
}
\description{
CreateComplianceJobRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{resumable}}{character [optional]}

\item{\code{type}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CreateComplianceJobRequest-new}{\code{CreateComplianceJobRequest$new()}}
\item \href{#method-CreateComplianceJobRequest-toJSON}{\code{CreateComplianceJobRequest$toJSON()}}
\item \href{#method-CreateComplianceJobRequest-fromJSON}{\code{CreateComplianceJobRequest$fromJSON()}}
\item \href{#method-CreateComplianceJobRequest-toJSONString}{\code{CreateComplianceJobRequest$toJSONString()}}
\item \href{#method-CreateComplianceJobRequest-fromJSONString}{\code{CreateComplianceJobRequest$fromJSONString()}}
\item \href{#method-CreateComplianceJobRequest-validateJSON}{\code{CreateComplianceJobRequest$validateJSON()}}
\item \href{#method-CreateComplianceJobRequest-toString}{\code{CreateComplianceJobRequest$toString()}}
\item \href{#method-CreateComplianceJobRequest-isValid}{\code{CreateComplianceJobRequest$isValid()}}
\item \href{#method-CreateComplianceJobRequest-getInvalidFields}{\code{CreateComplianceJobRequest$getInvalidFields()}}
\item \href{#method-CreateComplianceJobRequest-clone}{\code{CreateComplianceJobRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CreateComplianceJobRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobRequest$new(type, name = NULL, resumable = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Type of compliance job to list.}

\item{\code{name}}{User-provided name for a compliance job.}

\item{\code{resumable}}{If true, this endpoint will return a pre-signed URL with resumable uploads enabled.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CreateComplianceJobRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of CreateComplianceJobRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CreateComplianceJobRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CreateComplianceJobRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of CreateComplianceJobRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CreateComplianceJobRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to CreateComplianceJobRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of CreateComplianceJobRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
