% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_compliance_job_response.R
\docType{class}
\name{CreateComplianceJobResponse}
\alias{CreateComplianceJobResponse}
\title{CreateComplianceJobResponse}
\format{
An \code{R6Class} generator object
}
\description{
CreateComplianceJobResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{ComplianceJob} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CreateComplianceJobResponse-new}{\code{CreateComplianceJobResponse$new()}}
\item \href{#method-CreateComplianceJobResponse-toJSON}{\code{CreateComplianceJobResponse$toJSON()}}
\item \href{#method-CreateComplianceJobResponse-fromJSON}{\code{CreateComplianceJobResponse$fromJSON()}}
\item \href{#method-CreateComplianceJobResponse-toJSONString}{\code{CreateComplianceJobResponse$toJSONString()}}
\item \href{#method-CreateComplianceJobResponse-fromJSONString}{\code{CreateComplianceJobResponse$fromJSONString()}}
\item \href{#method-CreateComplianceJobResponse-validateJSON}{\code{CreateComplianceJobResponse$validateJSON()}}
\item \href{#method-CreateComplianceJobResponse-toString}{\code{CreateComplianceJobResponse$toString()}}
\item \href{#method-CreateComplianceJobResponse-isValid}{\code{CreateComplianceJobResponse$isValid()}}
\item \href{#method-CreateComplianceJobResponse-getInvalidFields}{\code{CreateComplianceJobResponse$getInvalidFields()}}
\item \href{#method-CreateComplianceJobResponse-clone}{\code{CreateComplianceJobResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CreateComplianceJobResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CreateComplianceJobResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of CreateComplianceJobResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CreateComplianceJobResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CreateComplianceJobResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of CreateComplianceJobResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CreateComplianceJobResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to CreateComplianceJobResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of CreateComplianceJobResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateComplianceJobResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CreateComplianceJobResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CreateComplianceJobResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
