% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_rules_request.R
\docType{class}
\name{DeleteRulesRequest}
\alias{DeleteRulesRequest}
\title{DeleteRulesRequest}
\format{
An \code{R6Class} generator object
}
\description{
DeleteRulesRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{delete}}{\link{DeleteRulesRequestDelete}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeleteRulesRequest-new}{\code{DeleteRulesRequest$new()}}
\item \href{#method-DeleteRulesRequest-toJSON}{\code{DeleteRulesRequest$toJSON()}}
\item \href{#method-DeleteRulesRequest-fromJSON}{\code{DeleteRulesRequest$fromJSON()}}
\item \href{#method-DeleteRulesRequest-toJSONString}{\code{DeleteRulesRequest$toJSONString()}}
\item \href{#method-DeleteRulesRequest-fromJSONString}{\code{DeleteRulesRequest$fromJSONString()}}
\item \href{#method-DeleteRulesRequest-validateJSON}{\code{DeleteRulesRequest$validateJSON()}}
\item \href{#method-DeleteRulesRequest-toString}{\code{DeleteRulesRequest$toString()}}
\item \href{#method-DeleteRulesRequest-isValid}{\code{DeleteRulesRequest$isValid()}}
\item \href{#method-DeleteRulesRequest-getInvalidFields}{\code{DeleteRulesRequest$getInvalidFields()}}
\item \href{#method-DeleteRulesRequest-clone}{\code{DeleteRulesRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteRulesRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteRulesRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DeleteRulesRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteRulesRequest$new(delete, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delete}}{delete}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteRulesRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteRulesRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteRulesRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DeleteRulesRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteRulesRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteRulesRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of DeleteRulesRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteRulesRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DeleteRulesRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteRulesRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteRulesRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteRulesRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DeleteRulesRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteRulesRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteRulesRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of DeleteRulesRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteRulesRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DeleteRulesRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteRulesRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteRulesRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to DeleteRulesRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteRulesRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteRulesRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteRulesRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteRulesRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of DeleteRulesRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteRulesRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteRulesRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteRulesRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteRulesRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteRulesRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteRulesRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteRulesRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteRulesRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeleteRulesRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
