% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disallowed_resource_problem.R
\docType{class}
\name{DisallowedResourceProblem}
\alias{DisallowedResourceProblem}
\title{DisallowedResourceProblem}
\format{
An \code{R6Class} generator object
}
\description{
DisallowedResourceProblem Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Super class}{
\code{\link[twitter:Problem]{twitter::Problem}} -> \code{DisallowedResourceProblem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detail}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character}

\item{\code{type}}{character}

\item{\code{resource_id}}{character}

\item{\code{resource_type}}{character}

\item{\code{section}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DisallowedResourceProblem-new}{\code{DisallowedResourceProblem$new()}}
\item \href{#method-DisallowedResourceProblem-toJSON}{\code{DisallowedResourceProblem$toJSON()}}
\item \href{#method-DisallowedResourceProblem-fromJSON}{\code{DisallowedResourceProblem$fromJSON()}}
\item \href{#method-DisallowedResourceProblem-toJSONString}{\code{DisallowedResourceProblem$toJSONString()}}
\item \href{#method-DisallowedResourceProblem-fromJSONString}{\code{DisallowedResourceProblem$fromJSONString()}}
\item \href{#method-DisallowedResourceProblem-validateJSON}{\code{DisallowedResourceProblem$validateJSON()}}
\item \href{#method-DisallowedResourceProblem-toString}{\code{DisallowedResourceProblem$toString()}}
\item \href{#method-DisallowedResourceProblem-isValid}{\code{DisallowedResourceProblem$isValid()}}
\item \href{#method-DisallowedResourceProblem-getInvalidFields}{\code{DisallowedResourceProblem$getInvalidFields()}}
\item \href{#method-DisallowedResourceProblem-clone}{\code{DisallowedResourceProblem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DisallowedResourceProblem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblem$new(
  title,
  type,
  resource_id,
  resource_type,
  section,
  detail = NULL,
  status = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{type}}{type}

\item{\code{resource_id}}{resource_id}

\item{\code{resource_type}}{resource_type}

\item{\code{section}}{section}

\item{\code{detail}}{detail}

\item{\code{status}}{status}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DisallowedResourceProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of DisallowedResourceProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DisallowedResourceProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DisallowedResourceProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of DisallowedResourceProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DisallowedResourceProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to DisallowedResourceProblem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of DisallowedResourceProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
