% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disallowed_resource_problem_all_of.R
\docType{class}
\name{DisallowedResourceProblemAllOf}
\alias{DisallowedResourceProblemAllOf}
\title{DisallowedResourceProblemAllOf}
\format{
An \code{R6Class} generator object
}
\description{
DisallowedResourceProblemAllOf Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{resource_id}}{character}

\item{\code{resource_type}}{character}

\item{\code{section}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DisallowedResourceProblemAllOf-new}{\code{DisallowedResourceProblemAllOf$new()}}
\item \href{#method-DisallowedResourceProblemAllOf-toJSON}{\code{DisallowedResourceProblemAllOf$toJSON()}}
\item \href{#method-DisallowedResourceProblemAllOf-fromJSON}{\code{DisallowedResourceProblemAllOf$fromJSON()}}
\item \href{#method-DisallowedResourceProblemAllOf-toJSONString}{\code{DisallowedResourceProblemAllOf$toJSONString()}}
\item \href{#method-DisallowedResourceProblemAllOf-fromJSONString}{\code{DisallowedResourceProblemAllOf$fromJSONString()}}
\item \href{#method-DisallowedResourceProblemAllOf-validateJSON}{\code{DisallowedResourceProblemAllOf$validateJSON()}}
\item \href{#method-DisallowedResourceProblemAllOf-toString}{\code{DisallowedResourceProblemAllOf$toString()}}
\item \href{#method-DisallowedResourceProblemAllOf-isValid}{\code{DisallowedResourceProblemAllOf$isValid()}}
\item \href{#method-DisallowedResourceProblemAllOf-getInvalidFields}{\code{DisallowedResourceProblemAllOf$getInvalidFields()}}
\item \href{#method-DisallowedResourceProblemAllOf-clone}{\code{DisallowedResourceProblemAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblemAllOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblemAllOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DisallowedResourceProblemAllOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblemAllOf$new(resource_id, resource_type, section, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource_id}}{resource_id}

\item{\code{resource_type}}{resource_type}

\item{\code{section}}{section}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblemAllOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblemAllOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblemAllOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DisallowedResourceProblemAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblemAllOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblemAllOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of DisallowedResourceProblemAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblemAllOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DisallowedResourceProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblemAllOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblemAllOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblemAllOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DisallowedResourceProblemAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblemAllOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblemAllOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of DisallowedResourceProblemAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblemAllOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DisallowedResourceProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblemAllOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblemAllOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to DisallowedResourceProblemAllOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblemAllOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblemAllOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblemAllOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblemAllOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of DisallowedResourceProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblemAllOf-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblemAllOf-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblemAllOf$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblemAllOf-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblemAllOf-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblemAllOf$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DisallowedResourceProblemAllOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DisallowedResourceProblemAllOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DisallowedResourceProblemAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
