% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_rule_problem.R
\docType{class}
\name{DuplicateRuleProblem}
\alias{DuplicateRuleProblem}
\title{DuplicateRuleProblem}
\format{
An \code{R6Class} generator object
}
\description{
DuplicateRuleProblem Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Super class}{
\code{\link[twitter:Problem]{twitter::Problem}} -> \code{DuplicateRuleProblem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detail}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character}

\item{\code{type}}{character}

\item{\code{id}}{character [optional]}

\item{\code{value}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DuplicateRuleProblem-new}{\code{DuplicateRuleProblem$new()}}
\item \href{#method-DuplicateRuleProblem-toJSON}{\code{DuplicateRuleProblem$toJSON()}}
\item \href{#method-DuplicateRuleProblem-fromJSON}{\code{DuplicateRuleProblem$fromJSON()}}
\item \href{#method-DuplicateRuleProblem-toJSONString}{\code{DuplicateRuleProblem$toJSONString()}}
\item \href{#method-DuplicateRuleProblem-fromJSONString}{\code{DuplicateRuleProblem$fromJSONString()}}
\item \href{#method-DuplicateRuleProblem-validateJSON}{\code{DuplicateRuleProblem$validateJSON()}}
\item \href{#method-DuplicateRuleProblem-toString}{\code{DuplicateRuleProblem$toString()}}
\item \href{#method-DuplicateRuleProblem-isValid}{\code{DuplicateRuleProblem$isValid()}}
\item \href{#method-DuplicateRuleProblem-getInvalidFields}{\code{DuplicateRuleProblem$getInvalidFields()}}
\item \href{#method-DuplicateRuleProblem-clone}{\code{DuplicateRuleProblem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuplicateRuleProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-DuplicateRuleProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DuplicateRuleProblem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuplicateRuleProblem$new(
  title,
  type,
  detail = NULL,
  status = NULL,
  id = NULL,
  value = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{type}}{type}

\item{\code{detail}}{detail}

\item{\code{status}}{status}

\item{\code{id}}{id}

\item{\code{value}}{value}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuplicateRuleProblem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DuplicateRuleProblem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuplicateRuleProblem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DuplicateRuleProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuplicateRuleProblem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DuplicateRuleProblem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of DuplicateRuleProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuplicateRuleProblem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DuplicateRuleProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuplicateRuleProblem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DuplicateRuleProblem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuplicateRuleProblem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DuplicateRuleProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuplicateRuleProblem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DuplicateRuleProblem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of DuplicateRuleProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuplicateRuleProblem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DuplicateRuleProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuplicateRuleProblem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DuplicateRuleProblem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to DuplicateRuleProblem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuplicateRuleProblem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuplicateRuleProblem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-DuplicateRuleProblem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuplicateRuleProblem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of DuplicateRuleProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuplicateRuleProblem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-DuplicateRuleProblem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuplicateRuleProblem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuplicateRuleProblem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-DuplicateRuleProblem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuplicateRuleProblem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuplicateRuleProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DuplicateRuleProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuplicateRuleProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
