% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_indices_inclusive_inclusive.R
\docType{class}
\name{EntityIndicesInclusiveInclusive}
\alias{EntityIndicesInclusiveInclusive}
\title{EntityIndicesInclusiveInclusive}
\format{
An \code{R6Class} generator object
}
\description{
EntityIndicesInclusiveInclusive Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{end}}{integer}

\item{\code{start}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EntityIndicesInclusiveInclusive-new}{\code{EntityIndicesInclusiveInclusive$new()}}
\item \href{#method-EntityIndicesInclusiveInclusive-toJSON}{\code{EntityIndicesInclusiveInclusive$toJSON()}}
\item \href{#method-EntityIndicesInclusiveInclusive-fromJSON}{\code{EntityIndicesInclusiveInclusive$fromJSON()}}
\item \href{#method-EntityIndicesInclusiveInclusive-toJSONString}{\code{EntityIndicesInclusiveInclusive$toJSONString()}}
\item \href{#method-EntityIndicesInclusiveInclusive-fromJSONString}{\code{EntityIndicesInclusiveInclusive$fromJSONString()}}
\item \href{#method-EntityIndicesInclusiveInclusive-validateJSON}{\code{EntityIndicesInclusiveInclusive$validateJSON()}}
\item \href{#method-EntityIndicesInclusiveInclusive-toString}{\code{EntityIndicesInclusiveInclusive$toString()}}
\item \href{#method-EntityIndicesInclusiveInclusive-isValid}{\code{EntityIndicesInclusiveInclusive$isValid()}}
\item \href{#method-EntityIndicesInclusiveInclusive-getInvalidFields}{\code{EntityIndicesInclusiveInclusive$getInvalidFields()}}
\item \href{#method-EntityIndicesInclusiveInclusive-clone}{\code{EntityIndicesInclusiveInclusive$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityIndicesInclusiveInclusive-new"></a>}}
\if{latex}{\out{\hypertarget{method-EntityIndicesInclusiveInclusive-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new EntityIndicesInclusiveInclusive class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityIndicesInclusiveInclusive$new(end, start, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{end}}{Index (zero-based) at which position this entity ends.  The index is inclusive.}

\item{\code{start}}{Index (zero-based) at which position this entity starts.  The index is inclusive.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityIndicesInclusiveInclusive-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-EntityIndicesInclusiveInclusive-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityIndicesInclusiveInclusive$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
EntityIndicesInclusiveInclusive in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityIndicesInclusiveInclusive-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-EntityIndicesInclusiveInclusive-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of EntityIndicesInclusiveInclusive
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityIndicesInclusiveInclusive$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of EntityIndicesInclusiveInclusive
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityIndicesInclusiveInclusive-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-EntityIndicesInclusiveInclusive-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityIndicesInclusiveInclusive$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
EntityIndicesInclusiveInclusive in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityIndicesInclusiveInclusive-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-EntityIndicesInclusiveInclusive-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of EntityIndicesInclusiveInclusive
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityIndicesInclusiveInclusive$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of EntityIndicesInclusiveInclusive
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityIndicesInclusiveInclusive-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-EntityIndicesInclusiveInclusive-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to EntityIndicesInclusiveInclusive and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityIndicesInclusiveInclusive$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityIndicesInclusiveInclusive-toString"></a>}}
\if{latex}{\out{\hypertarget{method-EntityIndicesInclusiveInclusive-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityIndicesInclusiveInclusive$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of EntityIndicesInclusiveInclusive
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityIndicesInclusiveInclusive-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-EntityIndicesInclusiveInclusive-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityIndicesInclusiveInclusive$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityIndicesInclusiveInclusive-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-EntityIndicesInclusiveInclusive-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityIndicesInclusiveInclusive$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EntityIndicesInclusiveInclusive-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EntityIndicesInclusiveInclusive-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EntityIndicesInclusiveInclusive$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
