% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\docType{class}
\name{Error}
\alias{Error}
\title{Error}
\format{
An \code{R6Class} generator object
}
\description{
Error Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{code}}{integer}

\item{\code{message}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Error-new}{\code{Error$new()}}
\item \href{#method-Error-toJSON}{\code{Error$toJSON()}}
\item \href{#method-Error-fromJSON}{\code{Error$fromJSON()}}
\item \href{#method-Error-toJSONString}{\code{Error$toJSONString()}}
\item \href{#method-Error-fromJSONString}{\code{Error$fromJSONString()}}
\item \href{#method-Error-validateJSON}{\code{Error$validateJSON()}}
\item \href{#method-Error-toString}{\code{Error$toString()}}
\item \href{#method-Error-isValid}{\code{Error$isValid()}}
\item \href{#method-Error-getInvalidFields}{\code{Error$getInvalidFields()}}
\item \href{#method-Error-clone}{\code{Error$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-new"></a>}}
\if{latex}{\out{\hypertarget{method-Error-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Error class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$new(code, message, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{code}

\item{\code{message}}{message}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Error-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Error in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Error-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Error
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Error
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Error-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Error in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Error-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Error
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Error
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Error-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Error and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Error-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Error
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Error-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Error-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Error-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
