% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_unauthorized_problem.R
\docType{class}
\name{FieldUnauthorizedProblem}
\alias{FieldUnauthorizedProblem}
\title{FieldUnauthorizedProblem}
\format{
An \code{R6Class} generator object
}
\description{
FieldUnauthorizedProblem Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Super class}{
\code{\link[twitter:Problem]{twitter::Problem}} -> \code{FieldUnauthorizedProblem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detail}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character}

\item{\code{type}}{character}

\item{\code{field}}{character}

\item{\code{resource_type}}{character}

\item{\code{section}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FieldUnauthorizedProblem-new}{\code{FieldUnauthorizedProblem$new()}}
\item \href{#method-FieldUnauthorizedProblem-toJSON}{\code{FieldUnauthorizedProblem$toJSON()}}
\item \href{#method-FieldUnauthorizedProblem-fromJSON}{\code{FieldUnauthorizedProblem$fromJSON()}}
\item \href{#method-FieldUnauthorizedProblem-toJSONString}{\code{FieldUnauthorizedProblem$toJSONString()}}
\item \href{#method-FieldUnauthorizedProblem-fromJSONString}{\code{FieldUnauthorizedProblem$fromJSONString()}}
\item \href{#method-FieldUnauthorizedProblem-validateJSON}{\code{FieldUnauthorizedProblem$validateJSON()}}
\item \href{#method-FieldUnauthorizedProblem-toString}{\code{FieldUnauthorizedProblem$toString()}}
\item \href{#method-FieldUnauthorizedProblem-isValid}{\code{FieldUnauthorizedProblem$isValid()}}
\item \href{#method-FieldUnauthorizedProblem-getInvalidFields}{\code{FieldUnauthorizedProblem$getInvalidFields()}}
\item \href{#method-FieldUnauthorizedProblem-clone}{\code{FieldUnauthorizedProblem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldUnauthorizedProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-FieldUnauthorizedProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FieldUnauthorizedProblem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldUnauthorizedProblem$new(
  title,
  type,
  field,
  resource_type,
  section,
  detail = NULL,
  status = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{type}}{type}

\item{\code{field}}{field}

\item{\code{resource_type}}{resource_type}

\item{\code{section}}{section}

\item{\code{detail}}{detail}

\item{\code{status}}{status}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldUnauthorizedProblem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FieldUnauthorizedProblem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldUnauthorizedProblem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FieldUnauthorizedProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldUnauthorizedProblem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FieldUnauthorizedProblem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FieldUnauthorizedProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldUnauthorizedProblem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FieldUnauthorizedProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldUnauthorizedProblem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FieldUnauthorizedProblem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldUnauthorizedProblem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FieldUnauthorizedProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldUnauthorizedProblem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FieldUnauthorizedProblem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FieldUnauthorizedProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldUnauthorizedProblem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FieldUnauthorizedProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldUnauthorizedProblem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FieldUnauthorizedProblem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FieldUnauthorizedProblem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldUnauthorizedProblem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldUnauthorizedProblem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FieldUnauthorizedProblem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldUnauthorizedProblem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FieldUnauthorizedProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldUnauthorizedProblem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FieldUnauthorizedProblem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldUnauthorizedProblem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldUnauthorizedProblem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FieldUnauthorizedProblem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldUnauthorizedProblem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldUnauthorizedProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FieldUnauthorizedProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldUnauthorizedProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
