% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_streaming_tweet_response.R
\docType{class}
\name{FilteredStreamingTweetResponse}
\alias{FilteredStreamingTweetResponse}
\title{FilteredStreamingTweetResponse}
\format{
An \code{R6Class} generator object
}
\description{
FilteredStreamingTweetResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{Tweet} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{matching_rules}}{list( \link{FilteredStreamingTweetResponseMatchingRulesInner} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilteredStreamingTweetResponse-new}{\code{FilteredStreamingTweetResponse$new()}}
\item \href{#method-FilteredStreamingTweetResponse-toJSON}{\code{FilteredStreamingTweetResponse$toJSON()}}
\item \href{#method-FilteredStreamingTweetResponse-fromJSON}{\code{FilteredStreamingTweetResponse$fromJSON()}}
\item \href{#method-FilteredStreamingTweetResponse-toJSONString}{\code{FilteredStreamingTweetResponse$toJSONString()}}
\item \href{#method-FilteredStreamingTweetResponse-fromJSONString}{\code{FilteredStreamingTweetResponse$fromJSONString()}}
\item \href{#method-FilteredStreamingTweetResponse-validateJSON}{\code{FilteredStreamingTweetResponse$validateJSON()}}
\item \href{#method-FilteredStreamingTweetResponse-toString}{\code{FilteredStreamingTweetResponse$toString()}}
\item \href{#method-FilteredStreamingTweetResponse-isValid}{\code{FilteredStreamingTweetResponse$isValid()}}
\item \href{#method-FilteredStreamingTweetResponse-getInvalidFields}{\code{FilteredStreamingTweetResponse$getInvalidFields()}}
\item \href{#method-FilteredStreamingTweetResponse-clone}{\code{FilteredStreamingTweetResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FilteredStreamingTweetResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  matching_rules = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{matching_rules}}{The list of rules which matched the Tweet}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FilteredStreamingTweetResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FilteredStreamingTweetResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FilteredStreamingTweetResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FilteredStreamingTweetResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FilteredStreamingTweetResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FilteredStreamingTweetResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FilteredStreamingTweetResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FilteredStreamingTweetResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
