% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/filtered_streaming_tweet_response_matching_rules_inner.R
\docType{class}
\name{FilteredStreamingTweetResponseMatchingRulesInner}
\alias{FilteredStreamingTweetResponseMatchingRulesInner}
\title{FilteredStreamingTweetResponseMatchingRulesInner}
\format{
An \code{R6Class} generator object
}
\description{
FilteredStreamingTweetResponseMatchingRulesInner Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character}

\item{\code{tag}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilteredStreamingTweetResponseMatchingRulesInner-new}{\code{FilteredStreamingTweetResponseMatchingRulesInner$new()}}
\item \href{#method-FilteredStreamingTweetResponseMatchingRulesInner-toJSON}{\code{FilteredStreamingTweetResponseMatchingRulesInner$toJSON()}}
\item \href{#method-FilteredStreamingTweetResponseMatchingRulesInner-fromJSON}{\code{FilteredStreamingTweetResponseMatchingRulesInner$fromJSON()}}
\item \href{#method-FilteredStreamingTweetResponseMatchingRulesInner-toJSONString}{\code{FilteredStreamingTweetResponseMatchingRulesInner$toJSONString()}}
\item \href{#method-FilteredStreamingTweetResponseMatchingRulesInner-fromJSONString}{\code{FilteredStreamingTweetResponseMatchingRulesInner$fromJSONString()}}
\item \href{#method-FilteredStreamingTweetResponseMatchingRulesInner-validateJSON}{\code{FilteredStreamingTweetResponseMatchingRulesInner$validateJSON()}}
\item \href{#method-FilteredStreamingTweetResponseMatchingRulesInner-toString}{\code{FilteredStreamingTweetResponseMatchingRulesInner$toString()}}
\item \href{#method-FilteredStreamingTweetResponseMatchingRulesInner-isValid}{\code{FilteredStreamingTweetResponseMatchingRulesInner$isValid()}}
\item \href{#method-FilteredStreamingTweetResponseMatchingRulesInner-getInvalidFields}{\code{FilteredStreamingTweetResponseMatchingRulesInner$getInvalidFields()}}
\item \href{#method-FilteredStreamingTweetResponseMatchingRulesInner-clone}{\code{FilteredStreamingTweetResponseMatchingRulesInner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponseMatchingRulesInner-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponseMatchingRulesInner-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FilteredStreamingTweetResponseMatchingRulesInner class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponseMatchingRulesInner$new(id, tag = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Unique identifier of this rule.}

\item{\code{tag}}{A tag meant for the labeling of user provided rules.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponseMatchingRulesInner-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponseMatchingRulesInner-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponseMatchingRulesInner$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FilteredStreamingTweetResponseMatchingRulesInner in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponseMatchingRulesInner-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponseMatchingRulesInner-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FilteredStreamingTweetResponseMatchingRulesInner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponseMatchingRulesInner$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FilteredStreamingTweetResponseMatchingRulesInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponseMatchingRulesInner-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponseMatchingRulesInner-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponseMatchingRulesInner$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FilteredStreamingTweetResponseMatchingRulesInner in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponseMatchingRulesInner-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponseMatchingRulesInner-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FilteredStreamingTweetResponseMatchingRulesInner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponseMatchingRulesInner$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FilteredStreamingTweetResponseMatchingRulesInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponseMatchingRulesInner-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponseMatchingRulesInner-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FilteredStreamingTweetResponseMatchingRulesInner and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponseMatchingRulesInner$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponseMatchingRulesInner-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponseMatchingRulesInner-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponseMatchingRulesInner$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FilteredStreamingTweetResponseMatchingRulesInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponseMatchingRulesInner-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponseMatchingRulesInner-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponseMatchingRulesInner$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponseMatchingRulesInner-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponseMatchingRulesInner-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponseMatchingRulesInner$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilteredStreamingTweetResponseMatchingRulesInner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilteredStreamingTweetResponseMatchingRulesInner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredStreamingTweetResponseMatchingRulesInner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
