% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_text_entities_annotations_inner.R
\docType{class}
\name{FullTextEntitiesAnnotationsInner}
\alias{FullTextEntitiesAnnotationsInner}
\title{FullTextEntitiesAnnotationsInner}
\format{
An \code{R6Class} generator object
}
\description{
FullTextEntitiesAnnotationsInner Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{end}}{integer}

\item{\code{start}}{integer}

\item{\code{normalized_text}}{character [optional]}

\item{\code{probability}}{numeric [optional]}

\item{\code{type}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FullTextEntitiesAnnotationsInner-new}{\code{FullTextEntitiesAnnotationsInner$new()}}
\item \href{#method-FullTextEntitiesAnnotationsInner-toJSON}{\code{FullTextEntitiesAnnotationsInner$toJSON()}}
\item \href{#method-FullTextEntitiesAnnotationsInner-fromJSON}{\code{FullTextEntitiesAnnotationsInner$fromJSON()}}
\item \href{#method-FullTextEntitiesAnnotationsInner-toJSONString}{\code{FullTextEntitiesAnnotationsInner$toJSONString()}}
\item \href{#method-FullTextEntitiesAnnotationsInner-fromJSONString}{\code{FullTextEntitiesAnnotationsInner$fromJSONString()}}
\item \href{#method-FullTextEntitiesAnnotationsInner-validateJSON}{\code{FullTextEntitiesAnnotationsInner$validateJSON()}}
\item \href{#method-FullTextEntitiesAnnotationsInner-toString}{\code{FullTextEntitiesAnnotationsInner$toString()}}
\item \href{#method-FullTextEntitiesAnnotationsInner-isValid}{\code{FullTextEntitiesAnnotationsInner$isValid()}}
\item \href{#method-FullTextEntitiesAnnotationsInner-getInvalidFields}{\code{FullTextEntitiesAnnotationsInner$getInvalidFields()}}
\item \href{#method-FullTextEntitiesAnnotationsInner-clone}{\code{FullTextEntitiesAnnotationsInner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInner-new"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInner-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FullTextEntitiesAnnotationsInner class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInner$new(
  end,
  start,
  normalized_text = NULL,
  probability = NULL,
  type = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{end}}{Index (zero-based) at which position this entity ends.  The index is inclusive.}

\item{\code{start}}{Index (zero-based) at which position this entity starts.  The index is inclusive.}

\item{\code{normalized_text}}{Text used to determine annotation.}

\item{\code{probability}}{Confidence factor for annotation type.}

\item{\code{type}}{Annotation type.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInner-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInner-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInner$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FullTextEntitiesAnnotationsInner in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInner-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInner-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FullTextEntitiesAnnotationsInner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInner$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FullTextEntitiesAnnotationsInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInner-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInner-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInner$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FullTextEntitiesAnnotationsInner in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInner-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInner-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FullTextEntitiesAnnotationsInner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInner$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FullTextEntitiesAnnotationsInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInner-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInner-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FullTextEntitiesAnnotationsInner and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInner$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInner-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInner-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInner$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FullTextEntitiesAnnotationsInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInner-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInner-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInner$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInner-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInner-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInner$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
