% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_text_entities_annotations_inner_all_of.R
\docType{class}
\name{FullTextEntitiesAnnotationsInnerAllOf}
\alias{FullTextEntitiesAnnotationsInnerAllOf}
\title{FullTextEntitiesAnnotationsInnerAllOf}
\format{
An \code{R6Class} generator object
}
\description{
FullTextEntitiesAnnotationsInnerAllOf Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{normalized_text}}{character [optional]}

\item{\code{probability}}{numeric [optional]}

\item{\code{type}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FullTextEntitiesAnnotationsInnerAllOf-new}{\code{FullTextEntitiesAnnotationsInnerAllOf$new()}}
\item \href{#method-FullTextEntitiesAnnotationsInnerAllOf-toJSON}{\code{FullTextEntitiesAnnotationsInnerAllOf$toJSON()}}
\item \href{#method-FullTextEntitiesAnnotationsInnerAllOf-fromJSON}{\code{FullTextEntitiesAnnotationsInnerAllOf$fromJSON()}}
\item \href{#method-FullTextEntitiesAnnotationsInnerAllOf-toJSONString}{\code{FullTextEntitiesAnnotationsInnerAllOf$toJSONString()}}
\item \href{#method-FullTextEntitiesAnnotationsInnerAllOf-fromJSONString}{\code{FullTextEntitiesAnnotationsInnerAllOf$fromJSONString()}}
\item \href{#method-FullTextEntitiesAnnotationsInnerAllOf-validateJSON}{\code{FullTextEntitiesAnnotationsInnerAllOf$validateJSON()}}
\item \href{#method-FullTextEntitiesAnnotationsInnerAllOf-toString}{\code{FullTextEntitiesAnnotationsInnerAllOf$toString()}}
\item \href{#method-FullTextEntitiesAnnotationsInnerAllOf-isValid}{\code{FullTextEntitiesAnnotationsInnerAllOf$isValid()}}
\item \href{#method-FullTextEntitiesAnnotationsInnerAllOf-getInvalidFields}{\code{FullTextEntitiesAnnotationsInnerAllOf$getInvalidFields()}}
\item \href{#method-FullTextEntitiesAnnotationsInnerAllOf-clone}{\code{FullTextEntitiesAnnotationsInnerAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInnerAllOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInnerAllOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new FullTextEntitiesAnnotationsInnerAllOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInnerAllOf$new(
  normalized_text = NULL,
  probability = NULL,
  type = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{normalized_text}}{Text used to determine annotation.}

\item{\code{probability}}{Confidence factor for annotation type.}

\item{\code{type}}{Annotation type.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInnerAllOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInnerAllOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInnerAllOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FullTextEntitiesAnnotationsInnerAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInnerAllOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInnerAllOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of FullTextEntitiesAnnotationsInnerAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInnerAllOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FullTextEntitiesAnnotationsInnerAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInnerAllOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInnerAllOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInnerAllOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FullTextEntitiesAnnotationsInnerAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInnerAllOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInnerAllOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of FullTextEntitiesAnnotationsInnerAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInnerAllOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of FullTextEntitiesAnnotationsInnerAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInnerAllOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInnerAllOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to FullTextEntitiesAnnotationsInnerAllOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInnerAllOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInnerAllOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInnerAllOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInnerAllOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of FullTextEntitiesAnnotationsInnerAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInnerAllOf-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInnerAllOf-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInnerAllOf$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInnerAllOf-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInnerAllOf-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInnerAllOf$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FullTextEntitiesAnnotationsInnerAllOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FullTextEntitiesAnnotationsInnerAllOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FullTextEntitiesAnnotationsInnerAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
