% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_problem.R
\docType{class}
\name{GenericProblem}
\alias{GenericProblem}
\title{GenericProblem}
\format{
An \code{R6Class} generator object
}
\description{
GenericProblem Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Super class}{
\code{\link[twitter:Problem]{twitter::Problem}} -> \code{GenericProblem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detail}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character}

\item{\code{type}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GenericProblem-new}{\code{GenericProblem$new()}}
\item \href{#method-GenericProblem-toJSON}{\code{GenericProblem$toJSON()}}
\item \href{#method-GenericProblem-fromJSON}{\code{GenericProblem$fromJSON()}}
\item \href{#method-GenericProblem-toJSONString}{\code{GenericProblem$toJSONString()}}
\item \href{#method-GenericProblem-fromJSONString}{\code{GenericProblem$fromJSONString()}}
\item \href{#method-GenericProblem-validateJSON}{\code{GenericProblem$validateJSON()}}
\item \href{#method-GenericProblem-toString}{\code{GenericProblem$toString()}}
\item \href{#method-GenericProblem-isValid}{\code{GenericProblem$isValid()}}
\item \href{#method-GenericProblem-getInvalidFields}{\code{GenericProblem$getInvalidFields()}}
\item \href{#method-GenericProblem-clone}{\code{GenericProblem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-GenericProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new GenericProblem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericProblem$new(title, type, detail = NULL, status = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{type}}{type}

\item{\code{detail}}{detail}

\item{\code{status}}{status}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericProblem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GenericProblem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericProblem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
GenericProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericProblem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GenericProblem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of GenericProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericProblem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of GenericProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericProblem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GenericProblem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericProblem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
GenericProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericProblem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GenericProblem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of GenericProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericProblem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of GenericProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericProblem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GenericProblem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to GenericProblem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericProblem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericProblem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-GenericProblem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericProblem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of GenericProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericProblem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-GenericProblem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericProblem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericProblem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-GenericProblem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericProblem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GenericProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
