% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_compliance_jobs_id_response.R
\docType{class}
\name{Get2ComplianceJobsIdResponse}
\alias{Get2ComplianceJobsIdResponse}
\title{Get2ComplianceJobsIdResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2ComplianceJobsIdResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{ComplianceJob} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2ComplianceJobsIdResponse-new}{\code{Get2ComplianceJobsIdResponse$new()}}
\item \href{#method-Get2ComplianceJobsIdResponse-toJSON}{\code{Get2ComplianceJobsIdResponse$toJSON()}}
\item \href{#method-Get2ComplianceJobsIdResponse-fromJSON}{\code{Get2ComplianceJobsIdResponse$fromJSON()}}
\item \href{#method-Get2ComplianceJobsIdResponse-toJSONString}{\code{Get2ComplianceJobsIdResponse$toJSONString()}}
\item \href{#method-Get2ComplianceJobsIdResponse-fromJSONString}{\code{Get2ComplianceJobsIdResponse$fromJSONString()}}
\item \href{#method-Get2ComplianceJobsIdResponse-validateJSON}{\code{Get2ComplianceJobsIdResponse$validateJSON()}}
\item \href{#method-Get2ComplianceJobsIdResponse-toString}{\code{Get2ComplianceJobsIdResponse$toString()}}
\item \href{#method-Get2ComplianceJobsIdResponse-isValid}{\code{Get2ComplianceJobsIdResponse$isValid()}}
\item \href{#method-Get2ComplianceJobsIdResponse-getInvalidFields}{\code{Get2ComplianceJobsIdResponse$getInvalidFields()}}
\item \href{#method-Get2ComplianceJobsIdResponse-clone}{\code{Get2ComplianceJobsIdResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsIdResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsIdResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2ComplianceJobsIdResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsIdResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsIdResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsIdResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsIdResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2ComplianceJobsIdResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsIdResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsIdResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2ComplianceJobsIdResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsIdResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2ComplianceJobsIdResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsIdResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsIdResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsIdResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2ComplianceJobsIdResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsIdResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsIdResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2ComplianceJobsIdResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsIdResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2ComplianceJobsIdResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsIdResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsIdResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2ComplianceJobsIdResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsIdResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsIdResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsIdResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsIdResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2ComplianceJobsIdResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsIdResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsIdResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsIdResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsIdResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsIdResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsIdResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsIdResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsIdResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsIdResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
