% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_compliance_jobs_response.R
\docType{class}
\name{Get2ComplianceJobsResponse}
\alias{Get2ComplianceJobsResponse}
\title{Get2ComplianceJobsResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2ComplianceJobsResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{ComplianceJob} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{meta}}{\link{Get2ComplianceJobsResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2ComplianceJobsResponse-new}{\code{Get2ComplianceJobsResponse$new()}}
\item \href{#method-Get2ComplianceJobsResponse-toJSON}{\code{Get2ComplianceJobsResponse$toJSON()}}
\item \href{#method-Get2ComplianceJobsResponse-fromJSON}{\code{Get2ComplianceJobsResponse$fromJSON()}}
\item \href{#method-Get2ComplianceJobsResponse-toJSONString}{\code{Get2ComplianceJobsResponse$toJSONString()}}
\item \href{#method-Get2ComplianceJobsResponse-fromJSONString}{\code{Get2ComplianceJobsResponse$fromJSONString()}}
\item \href{#method-Get2ComplianceJobsResponse-validateJSON}{\code{Get2ComplianceJobsResponse$validateJSON()}}
\item \href{#method-Get2ComplianceJobsResponse-toString}{\code{Get2ComplianceJobsResponse$toString()}}
\item \href{#method-Get2ComplianceJobsResponse-isValid}{\code{Get2ComplianceJobsResponse$isValid()}}
\item \href{#method-Get2ComplianceJobsResponse-getInvalidFields}{\code{Get2ComplianceJobsResponse$getInvalidFields()}}
\item \href{#method-Get2ComplianceJobsResponse-clone}{\code{Get2ComplianceJobsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2ComplianceJobsResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponse$new(data = NULL, errors = NULL, meta = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2ComplianceJobsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2ComplianceJobsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2ComplianceJobsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2ComplianceJobsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2ComplianceJobsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2ComplianceJobsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2ComplianceJobsResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2ComplianceJobsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
