% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_compliance_jobs_response_meta.R
\docType{class}
\name{Get2ComplianceJobsResponseMeta}
\alias{Get2ComplianceJobsResponseMeta}
\title{Get2ComplianceJobsResponseMeta}
\format{
An \code{R6Class} generator object
}
\description{
Get2ComplianceJobsResponseMeta Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{result_count}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2ComplianceJobsResponseMeta-new}{\code{Get2ComplianceJobsResponseMeta$new()}}
\item \href{#method-Get2ComplianceJobsResponseMeta-toJSON}{\code{Get2ComplianceJobsResponseMeta$toJSON()}}
\item \href{#method-Get2ComplianceJobsResponseMeta-fromJSON}{\code{Get2ComplianceJobsResponseMeta$fromJSON()}}
\item \href{#method-Get2ComplianceJobsResponseMeta-toJSONString}{\code{Get2ComplianceJobsResponseMeta$toJSONString()}}
\item \href{#method-Get2ComplianceJobsResponseMeta-fromJSONString}{\code{Get2ComplianceJobsResponseMeta$fromJSONString()}}
\item \href{#method-Get2ComplianceJobsResponseMeta-validateJSON}{\code{Get2ComplianceJobsResponseMeta$validateJSON()}}
\item \href{#method-Get2ComplianceJobsResponseMeta-toString}{\code{Get2ComplianceJobsResponseMeta$toString()}}
\item \href{#method-Get2ComplianceJobsResponseMeta-isValid}{\code{Get2ComplianceJobsResponseMeta$isValid()}}
\item \href{#method-Get2ComplianceJobsResponseMeta-getInvalidFields}{\code{Get2ComplianceJobsResponseMeta$getInvalidFields()}}
\item \href{#method-Get2ComplianceJobsResponseMeta-clone}{\code{Get2ComplianceJobsResponseMeta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponseMeta-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponseMeta-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2ComplianceJobsResponseMeta class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponseMeta$new(result_count = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result_count}}{The number of results returned in this response.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponseMeta-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponseMeta-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponseMeta$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2ComplianceJobsResponseMeta in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponseMeta-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponseMeta-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2ComplianceJobsResponseMeta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponseMeta$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2ComplianceJobsResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponseMeta-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponseMeta-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponseMeta$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2ComplianceJobsResponseMeta in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponseMeta-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponseMeta-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2ComplianceJobsResponseMeta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponseMeta$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2ComplianceJobsResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponseMeta-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponseMeta-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2ComplianceJobsResponseMeta and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponseMeta$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponseMeta-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponseMeta-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponseMeta$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2ComplianceJobsResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponseMeta-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponseMeta-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponseMeta$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponseMeta-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponseMeta-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponseMeta$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ComplianceJobsResponseMeta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ComplianceJobsResponseMeta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ComplianceJobsResponseMeta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
