% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_lists_id_followers_response.R
\docType{class}
\name{Get2ListsIdFollowersResponse}
\alias{Get2ListsIdFollowersResponse}
\title{Get2ListsIdFollowersResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2ListsIdFollowersResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{User} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ListsIdFollowersResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2ListsIdFollowersResponse-new}{\code{Get2ListsIdFollowersResponse$new()}}
\item \href{#method-Get2ListsIdFollowersResponse-toJSON}{\code{Get2ListsIdFollowersResponse$toJSON()}}
\item \href{#method-Get2ListsIdFollowersResponse-fromJSON}{\code{Get2ListsIdFollowersResponse$fromJSON()}}
\item \href{#method-Get2ListsIdFollowersResponse-toJSONString}{\code{Get2ListsIdFollowersResponse$toJSONString()}}
\item \href{#method-Get2ListsIdFollowersResponse-fromJSONString}{\code{Get2ListsIdFollowersResponse$fromJSONString()}}
\item \href{#method-Get2ListsIdFollowersResponse-validateJSON}{\code{Get2ListsIdFollowersResponse$validateJSON()}}
\item \href{#method-Get2ListsIdFollowersResponse-toString}{\code{Get2ListsIdFollowersResponse$toString()}}
\item \href{#method-Get2ListsIdFollowersResponse-isValid}{\code{Get2ListsIdFollowersResponse$isValid()}}
\item \href{#method-Get2ListsIdFollowersResponse-getInvalidFields}{\code{Get2ListsIdFollowersResponse$getInvalidFields()}}
\item \href{#method-Get2ListsIdFollowersResponse-clone}{\code{Get2ListsIdFollowersResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdFollowersResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdFollowersResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2ListsIdFollowersResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdFollowersResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdFollowersResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdFollowersResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdFollowersResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2ListsIdFollowersResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdFollowersResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdFollowersResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2ListsIdFollowersResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdFollowersResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2ListsIdFollowersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdFollowersResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdFollowersResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdFollowersResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2ListsIdFollowersResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdFollowersResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdFollowersResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2ListsIdFollowersResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdFollowersResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2ListsIdFollowersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdFollowersResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdFollowersResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2ListsIdFollowersResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdFollowersResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdFollowersResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdFollowersResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdFollowersResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2ListsIdFollowersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdFollowersResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdFollowersResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdFollowersResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdFollowersResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdFollowersResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdFollowersResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdFollowersResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdFollowersResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdFollowersResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
