% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_lists_id_response.R
\docType{class}
\name{Get2ListsIdResponse}
\alias{Get2ListsIdResponse}
\title{Get2ListsIdResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2ListsIdResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{List} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2ListsIdResponse-new}{\code{Get2ListsIdResponse$new()}}
\item \href{#method-Get2ListsIdResponse-toJSON}{\code{Get2ListsIdResponse$toJSON()}}
\item \href{#method-Get2ListsIdResponse-fromJSON}{\code{Get2ListsIdResponse$fromJSON()}}
\item \href{#method-Get2ListsIdResponse-toJSONString}{\code{Get2ListsIdResponse$toJSONString()}}
\item \href{#method-Get2ListsIdResponse-fromJSONString}{\code{Get2ListsIdResponse$fromJSONString()}}
\item \href{#method-Get2ListsIdResponse-validateJSON}{\code{Get2ListsIdResponse$validateJSON()}}
\item \href{#method-Get2ListsIdResponse-toString}{\code{Get2ListsIdResponse$toString()}}
\item \href{#method-Get2ListsIdResponse-isValid}{\code{Get2ListsIdResponse$isValid()}}
\item \href{#method-Get2ListsIdResponse-getInvalidFields}{\code{Get2ListsIdResponse$getInvalidFields()}}
\item \href{#method-Get2ListsIdResponse-clone}{\code{Get2ListsIdResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2ListsIdResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdResponse$new(data = NULL, errors = NULL, includes = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2ListsIdResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2ListsIdResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2ListsIdResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2ListsIdResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2ListsIdResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2ListsIdResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2ListsIdResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2ListsIdResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
