% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_lists_id_tweets_response.R
\docType{class}
\name{Get2ListsIdTweetsResponse}
\alias{Get2ListsIdTweetsResponse}
\title{Get2ListsIdTweetsResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2ListsIdTweetsResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{Tweet} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ListsIdFollowersResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2ListsIdTweetsResponse-new}{\code{Get2ListsIdTweetsResponse$new()}}
\item \href{#method-Get2ListsIdTweetsResponse-toJSON}{\code{Get2ListsIdTweetsResponse$toJSON()}}
\item \href{#method-Get2ListsIdTweetsResponse-fromJSON}{\code{Get2ListsIdTweetsResponse$fromJSON()}}
\item \href{#method-Get2ListsIdTweetsResponse-toJSONString}{\code{Get2ListsIdTweetsResponse$toJSONString()}}
\item \href{#method-Get2ListsIdTweetsResponse-fromJSONString}{\code{Get2ListsIdTweetsResponse$fromJSONString()}}
\item \href{#method-Get2ListsIdTweetsResponse-validateJSON}{\code{Get2ListsIdTweetsResponse$validateJSON()}}
\item \href{#method-Get2ListsIdTweetsResponse-toString}{\code{Get2ListsIdTweetsResponse$toString()}}
\item \href{#method-Get2ListsIdTweetsResponse-isValid}{\code{Get2ListsIdTweetsResponse$isValid()}}
\item \href{#method-Get2ListsIdTweetsResponse-getInvalidFields}{\code{Get2ListsIdTweetsResponse$getInvalidFields()}}
\item \href{#method-Get2ListsIdTweetsResponse-clone}{\code{Get2ListsIdTweetsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdTweetsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdTweetsResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2ListsIdTweetsResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdTweetsResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdTweetsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdTweetsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdTweetsResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2ListsIdTweetsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdTweetsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdTweetsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2ListsIdTweetsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdTweetsResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2ListsIdTweetsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdTweetsResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdTweetsResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdTweetsResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2ListsIdTweetsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdTweetsResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdTweetsResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2ListsIdTweetsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdTweetsResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2ListsIdTweetsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdTweetsResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdTweetsResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2ListsIdTweetsResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdTweetsResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdTweetsResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdTweetsResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdTweetsResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2ListsIdTweetsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdTweetsResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdTweetsResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdTweetsResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdTweetsResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdTweetsResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdTweetsResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2ListsIdTweetsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2ListsIdTweetsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2ListsIdTweetsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
