% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_spaces_by_creator_ids_response.R
\docType{class}
\name{Get2SpacesByCreatorIdsResponse}
\alias{Get2SpacesByCreatorIdsResponse}
\title{Get2SpacesByCreatorIdsResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2SpacesByCreatorIdsResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{Space} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ComplianceJobsResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2SpacesByCreatorIdsResponse-new}{\code{Get2SpacesByCreatorIdsResponse$new()}}
\item \href{#method-Get2SpacesByCreatorIdsResponse-toJSON}{\code{Get2SpacesByCreatorIdsResponse$toJSON()}}
\item \href{#method-Get2SpacesByCreatorIdsResponse-fromJSON}{\code{Get2SpacesByCreatorIdsResponse$fromJSON()}}
\item \href{#method-Get2SpacesByCreatorIdsResponse-toJSONString}{\code{Get2SpacesByCreatorIdsResponse$toJSONString()}}
\item \href{#method-Get2SpacesByCreatorIdsResponse-fromJSONString}{\code{Get2SpacesByCreatorIdsResponse$fromJSONString()}}
\item \href{#method-Get2SpacesByCreatorIdsResponse-validateJSON}{\code{Get2SpacesByCreatorIdsResponse$validateJSON()}}
\item \href{#method-Get2SpacesByCreatorIdsResponse-toString}{\code{Get2SpacesByCreatorIdsResponse$toString()}}
\item \href{#method-Get2SpacesByCreatorIdsResponse-isValid}{\code{Get2SpacesByCreatorIdsResponse$isValid()}}
\item \href{#method-Get2SpacesByCreatorIdsResponse-getInvalidFields}{\code{Get2SpacesByCreatorIdsResponse$getInvalidFields()}}
\item \href{#method-Get2SpacesByCreatorIdsResponse-clone}{\code{Get2SpacesByCreatorIdsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesByCreatorIdsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesByCreatorIdsResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2SpacesByCreatorIdsResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesByCreatorIdsResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesByCreatorIdsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesByCreatorIdsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesByCreatorIdsResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2SpacesByCreatorIdsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesByCreatorIdsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesByCreatorIdsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2SpacesByCreatorIdsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesByCreatorIdsResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2SpacesByCreatorIdsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesByCreatorIdsResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesByCreatorIdsResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesByCreatorIdsResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2SpacesByCreatorIdsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesByCreatorIdsResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesByCreatorIdsResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2SpacesByCreatorIdsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesByCreatorIdsResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2SpacesByCreatorIdsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesByCreatorIdsResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesByCreatorIdsResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2SpacesByCreatorIdsResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesByCreatorIdsResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesByCreatorIdsResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesByCreatorIdsResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesByCreatorIdsResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2SpacesByCreatorIdsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesByCreatorIdsResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesByCreatorIdsResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesByCreatorIdsResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesByCreatorIdsResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesByCreatorIdsResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesByCreatorIdsResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesByCreatorIdsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesByCreatorIdsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesByCreatorIdsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
