% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_spaces_id_buyers_response.R
\docType{class}
\name{Get2SpacesIdBuyersResponse}
\alias{Get2SpacesIdBuyersResponse}
\title{Get2SpacesIdBuyersResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2SpacesIdBuyersResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{User} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ListsIdFollowersResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2SpacesIdBuyersResponse-new}{\code{Get2SpacesIdBuyersResponse$new()}}
\item \href{#method-Get2SpacesIdBuyersResponse-toJSON}{\code{Get2SpacesIdBuyersResponse$toJSON()}}
\item \href{#method-Get2SpacesIdBuyersResponse-fromJSON}{\code{Get2SpacesIdBuyersResponse$fromJSON()}}
\item \href{#method-Get2SpacesIdBuyersResponse-toJSONString}{\code{Get2SpacesIdBuyersResponse$toJSONString()}}
\item \href{#method-Get2SpacesIdBuyersResponse-fromJSONString}{\code{Get2SpacesIdBuyersResponse$fromJSONString()}}
\item \href{#method-Get2SpacesIdBuyersResponse-validateJSON}{\code{Get2SpacesIdBuyersResponse$validateJSON()}}
\item \href{#method-Get2SpacesIdBuyersResponse-toString}{\code{Get2SpacesIdBuyersResponse$toString()}}
\item \href{#method-Get2SpacesIdBuyersResponse-isValid}{\code{Get2SpacesIdBuyersResponse$isValid()}}
\item \href{#method-Get2SpacesIdBuyersResponse-getInvalidFields}{\code{Get2SpacesIdBuyersResponse$getInvalidFields()}}
\item \href{#method-Get2SpacesIdBuyersResponse-clone}{\code{Get2SpacesIdBuyersResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdBuyersResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdBuyersResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2SpacesIdBuyersResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdBuyersResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdBuyersResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdBuyersResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdBuyersResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2SpacesIdBuyersResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdBuyersResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdBuyersResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2SpacesIdBuyersResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdBuyersResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2SpacesIdBuyersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdBuyersResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdBuyersResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdBuyersResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2SpacesIdBuyersResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdBuyersResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdBuyersResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2SpacesIdBuyersResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdBuyersResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2SpacesIdBuyersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdBuyersResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdBuyersResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2SpacesIdBuyersResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdBuyersResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdBuyersResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdBuyersResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdBuyersResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2SpacesIdBuyersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdBuyersResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdBuyersResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdBuyersResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdBuyersResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdBuyersResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdBuyersResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdBuyersResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdBuyersResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdBuyersResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
