% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_spaces_id_response.R
\docType{class}
\name{Get2SpacesIdResponse}
\alias{Get2SpacesIdResponse}
\title{Get2SpacesIdResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2SpacesIdResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{Space} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2SpacesIdResponse-new}{\code{Get2SpacesIdResponse$new()}}
\item \href{#method-Get2SpacesIdResponse-toJSON}{\code{Get2SpacesIdResponse$toJSON()}}
\item \href{#method-Get2SpacesIdResponse-fromJSON}{\code{Get2SpacesIdResponse$fromJSON()}}
\item \href{#method-Get2SpacesIdResponse-toJSONString}{\code{Get2SpacesIdResponse$toJSONString()}}
\item \href{#method-Get2SpacesIdResponse-fromJSONString}{\code{Get2SpacesIdResponse$fromJSONString()}}
\item \href{#method-Get2SpacesIdResponse-validateJSON}{\code{Get2SpacesIdResponse$validateJSON()}}
\item \href{#method-Get2SpacesIdResponse-toString}{\code{Get2SpacesIdResponse$toString()}}
\item \href{#method-Get2SpacesIdResponse-isValid}{\code{Get2SpacesIdResponse$isValid()}}
\item \href{#method-Get2SpacesIdResponse-getInvalidFields}{\code{Get2SpacesIdResponse$getInvalidFields()}}
\item \href{#method-Get2SpacesIdResponse-clone}{\code{Get2SpacesIdResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2SpacesIdResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdResponse$new(data = NULL, errors = NULL, includes = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2SpacesIdResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2SpacesIdResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2SpacesIdResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2SpacesIdResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2SpacesIdResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2SpacesIdResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2SpacesIdResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2SpacesIdResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesIdResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesIdResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesIdResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
