% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_spaces_response.R
\docType{class}
\name{Get2SpacesResponse}
\alias{Get2SpacesResponse}
\title{Get2SpacesResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2SpacesResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{Space} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2SpacesResponse-new}{\code{Get2SpacesResponse$new()}}
\item \href{#method-Get2SpacesResponse-toJSON}{\code{Get2SpacesResponse$toJSON()}}
\item \href{#method-Get2SpacesResponse-fromJSON}{\code{Get2SpacesResponse$fromJSON()}}
\item \href{#method-Get2SpacesResponse-toJSONString}{\code{Get2SpacesResponse$toJSONString()}}
\item \href{#method-Get2SpacesResponse-fromJSONString}{\code{Get2SpacesResponse$fromJSONString()}}
\item \href{#method-Get2SpacesResponse-validateJSON}{\code{Get2SpacesResponse$validateJSON()}}
\item \href{#method-Get2SpacesResponse-toString}{\code{Get2SpacesResponse$toString()}}
\item \href{#method-Get2SpacesResponse-isValid}{\code{Get2SpacesResponse$isValid()}}
\item \href{#method-Get2SpacesResponse-getInvalidFields}{\code{Get2SpacesResponse$getInvalidFields()}}
\item \href{#method-Get2SpacesResponse-clone}{\code{Get2SpacesResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2SpacesResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesResponse$new(data = NULL, errors = NULL, includes = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2SpacesResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2SpacesResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2SpacesResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2SpacesResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2SpacesResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2SpacesResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2SpacesResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2SpacesResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
