% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_spaces_search_response.R
\docType{class}
\name{Get2SpacesSearchResponse}
\alias{Get2SpacesSearchResponse}
\title{Get2SpacesSearchResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2SpacesSearchResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{Space} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ComplianceJobsResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2SpacesSearchResponse-new}{\code{Get2SpacesSearchResponse$new()}}
\item \href{#method-Get2SpacesSearchResponse-toJSON}{\code{Get2SpacesSearchResponse$toJSON()}}
\item \href{#method-Get2SpacesSearchResponse-fromJSON}{\code{Get2SpacesSearchResponse$fromJSON()}}
\item \href{#method-Get2SpacesSearchResponse-toJSONString}{\code{Get2SpacesSearchResponse$toJSONString()}}
\item \href{#method-Get2SpacesSearchResponse-fromJSONString}{\code{Get2SpacesSearchResponse$fromJSONString()}}
\item \href{#method-Get2SpacesSearchResponse-validateJSON}{\code{Get2SpacesSearchResponse$validateJSON()}}
\item \href{#method-Get2SpacesSearchResponse-toString}{\code{Get2SpacesSearchResponse$toString()}}
\item \href{#method-Get2SpacesSearchResponse-isValid}{\code{Get2SpacesSearchResponse$isValid()}}
\item \href{#method-Get2SpacesSearchResponse-getInvalidFields}{\code{Get2SpacesSearchResponse$getInvalidFields()}}
\item \href{#method-Get2SpacesSearchResponse-clone}{\code{Get2SpacesSearchResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesSearchResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesSearchResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2SpacesSearchResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesSearchResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesSearchResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesSearchResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesSearchResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2SpacesSearchResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesSearchResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesSearchResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2SpacesSearchResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesSearchResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2SpacesSearchResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesSearchResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesSearchResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesSearchResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2SpacesSearchResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesSearchResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesSearchResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2SpacesSearchResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesSearchResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2SpacesSearchResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesSearchResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesSearchResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2SpacesSearchResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesSearchResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesSearchResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesSearchResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesSearchResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2SpacesSearchResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesSearchResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesSearchResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesSearchResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesSearchResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesSearchResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesSearchResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2SpacesSearchResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2SpacesSearchResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2SpacesSearchResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
