% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_counts_all_response.R
\docType{class}
\name{Get2TweetsCountsAllResponse}
\alias{Get2TweetsCountsAllResponse}
\title{Get2TweetsCountsAllResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsCountsAllResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{SearchCount} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{meta}}{\link{Get2TweetsCountsAllResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsCountsAllResponse-new}{\code{Get2TweetsCountsAllResponse$new()}}
\item \href{#method-Get2TweetsCountsAllResponse-toJSON}{\code{Get2TweetsCountsAllResponse$toJSON()}}
\item \href{#method-Get2TweetsCountsAllResponse-fromJSON}{\code{Get2TweetsCountsAllResponse$fromJSON()}}
\item \href{#method-Get2TweetsCountsAllResponse-toJSONString}{\code{Get2TweetsCountsAllResponse$toJSONString()}}
\item \href{#method-Get2TweetsCountsAllResponse-fromJSONString}{\code{Get2TweetsCountsAllResponse$fromJSONString()}}
\item \href{#method-Get2TweetsCountsAllResponse-validateJSON}{\code{Get2TweetsCountsAllResponse$validateJSON()}}
\item \href{#method-Get2TweetsCountsAllResponse-toString}{\code{Get2TweetsCountsAllResponse$toString()}}
\item \href{#method-Get2TweetsCountsAllResponse-isValid}{\code{Get2TweetsCountsAllResponse$isValid()}}
\item \href{#method-Get2TweetsCountsAllResponse-getInvalidFields}{\code{Get2TweetsCountsAllResponse$getInvalidFields()}}
\item \href{#method-Get2TweetsCountsAllResponse-clone}{\code{Get2TweetsCountsAllResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsCountsAllResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponse$new(data = NULL, errors = NULL, meta = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsCountsAllResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsCountsAllResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsCountsAllResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsCountsAllResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsCountsAllResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsCountsAllResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsCountsAllResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsCountsAllResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
