% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_counts_all_response_meta.R
\docType{class}
\name{Get2TweetsCountsAllResponseMeta}
\alias{Get2TweetsCountsAllResponseMeta}
\title{Get2TweetsCountsAllResponseMeta}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsCountsAllResponseMeta Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{newest_id}}{character [optional]}

\item{\code{next_token}}{character [optional]}

\item{\code{oldest_id}}{character [optional]}

\item{\code{total_tweet_count}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsCountsAllResponseMeta-new}{\code{Get2TweetsCountsAllResponseMeta$new()}}
\item \href{#method-Get2TweetsCountsAllResponseMeta-toJSON}{\code{Get2TweetsCountsAllResponseMeta$toJSON()}}
\item \href{#method-Get2TweetsCountsAllResponseMeta-fromJSON}{\code{Get2TweetsCountsAllResponseMeta$fromJSON()}}
\item \href{#method-Get2TweetsCountsAllResponseMeta-toJSONString}{\code{Get2TweetsCountsAllResponseMeta$toJSONString()}}
\item \href{#method-Get2TweetsCountsAllResponseMeta-fromJSONString}{\code{Get2TweetsCountsAllResponseMeta$fromJSONString()}}
\item \href{#method-Get2TweetsCountsAllResponseMeta-validateJSON}{\code{Get2TweetsCountsAllResponseMeta$validateJSON()}}
\item \href{#method-Get2TweetsCountsAllResponseMeta-toString}{\code{Get2TweetsCountsAllResponseMeta$toString()}}
\item \href{#method-Get2TweetsCountsAllResponseMeta-isValid}{\code{Get2TweetsCountsAllResponseMeta$isValid()}}
\item \href{#method-Get2TweetsCountsAllResponseMeta-getInvalidFields}{\code{Get2TweetsCountsAllResponseMeta$getInvalidFields()}}
\item \href{#method-Get2TweetsCountsAllResponseMeta-clone}{\code{Get2TweetsCountsAllResponseMeta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponseMeta-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponseMeta-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsCountsAllResponseMeta class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponseMeta$new(
  newest_id = NULL,
  next_token = NULL,
  oldest_id = NULL,
  total_tweet_count = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newest_id}}{The newest id in this response.}

\item{\code{next_token}}{The next token.}

\item{\code{oldest_id}}{The oldest id in this response.}

\item{\code{total_tweet_count}}{The sum of results returned in this response.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponseMeta-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponseMeta-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponseMeta$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsCountsAllResponseMeta in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponseMeta-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponseMeta-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsCountsAllResponseMeta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponseMeta$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsCountsAllResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponseMeta-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponseMeta-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponseMeta$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsCountsAllResponseMeta in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponseMeta-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponseMeta-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsCountsAllResponseMeta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponseMeta$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsCountsAllResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponseMeta-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponseMeta-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsCountsAllResponseMeta and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponseMeta$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponseMeta-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponseMeta-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponseMeta$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsCountsAllResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponseMeta-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponseMeta-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponseMeta$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponseMeta-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponseMeta-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponseMeta$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsAllResponseMeta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsAllResponseMeta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsAllResponseMeta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
