% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_counts_recent_response.R
\docType{class}
\name{Get2TweetsCountsRecentResponse}
\alias{Get2TweetsCountsRecentResponse}
\title{Get2TweetsCountsRecentResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsCountsRecentResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{SearchCount} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{meta}}{\link{Get2TweetsCountsAllResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsCountsRecentResponse-new}{\code{Get2TweetsCountsRecentResponse$new()}}
\item \href{#method-Get2TweetsCountsRecentResponse-toJSON}{\code{Get2TweetsCountsRecentResponse$toJSON()}}
\item \href{#method-Get2TweetsCountsRecentResponse-fromJSON}{\code{Get2TweetsCountsRecentResponse$fromJSON()}}
\item \href{#method-Get2TweetsCountsRecentResponse-toJSONString}{\code{Get2TweetsCountsRecentResponse$toJSONString()}}
\item \href{#method-Get2TweetsCountsRecentResponse-fromJSONString}{\code{Get2TweetsCountsRecentResponse$fromJSONString()}}
\item \href{#method-Get2TweetsCountsRecentResponse-validateJSON}{\code{Get2TweetsCountsRecentResponse$validateJSON()}}
\item \href{#method-Get2TweetsCountsRecentResponse-toString}{\code{Get2TweetsCountsRecentResponse$toString()}}
\item \href{#method-Get2TweetsCountsRecentResponse-isValid}{\code{Get2TweetsCountsRecentResponse$isValid()}}
\item \href{#method-Get2TweetsCountsRecentResponse-getInvalidFields}{\code{Get2TweetsCountsRecentResponse$getInvalidFields()}}
\item \href{#method-Get2TweetsCountsRecentResponse-clone}{\code{Get2TweetsCountsRecentResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsRecentResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsRecentResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsCountsRecentResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsRecentResponse$new(
  data = NULL,
  errors = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsRecentResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsRecentResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsRecentResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsCountsRecentResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsRecentResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsRecentResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsCountsRecentResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsRecentResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsCountsRecentResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsRecentResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsRecentResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsRecentResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsCountsRecentResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsRecentResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsRecentResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsCountsRecentResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsRecentResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsCountsRecentResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsRecentResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsRecentResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsCountsRecentResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsRecentResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsRecentResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsRecentResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsRecentResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsCountsRecentResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsRecentResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsRecentResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsRecentResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsRecentResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsRecentResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsRecentResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsCountsRecentResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsCountsRecentResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsCountsRecentResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
