% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_firehose_stream_response.R
\docType{class}
\name{Get2TweetsFirehoseStreamResponse}
\alias{Get2TweetsFirehoseStreamResponse}
\title{Get2TweetsFirehoseStreamResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsFirehoseStreamResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{Tweet} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsFirehoseStreamResponse-new}{\code{Get2TweetsFirehoseStreamResponse$new()}}
\item \href{#method-Get2TweetsFirehoseStreamResponse-toJSON}{\code{Get2TweetsFirehoseStreamResponse$toJSON()}}
\item \href{#method-Get2TweetsFirehoseStreamResponse-fromJSON}{\code{Get2TweetsFirehoseStreamResponse$fromJSON()}}
\item \href{#method-Get2TweetsFirehoseStreamResponse-toJSONString}{\code{Get2TweetsFirehoseStreamResponse$toJSONString()}}
\item \href{#method-Get2TweetsFirehoseStreamResponse-fromJSONString}{\code{Get2TweetsFirehoseStreamResponse$fromJSONString()}}
\item \href{#method-Get2TweetsFirehoseStreamResponse-validateJSON}{\code{Get2TweetsFirehoseStreamResponse$validateJSON()}}
\item \href{#method-Get2TweetsFirehoseStreamResponse-toString}{\code{Get2TweetsFirehoseStreamResponse$toString()}}
\item \href{#method-Get2TweetsFirehoseStreamResponse-isValid}{\code{Get2TweetsFirehoseStreamResponse$isValid()}}
\item \href{#method-Get2TweetsFirehoseStreamResponse-getInvalidFields}{\code{Get2TweetsFirehoseStreamResponse$getInvalidFields()}}
\item \href{#method-Get2TweetsFirehoseStreamResponse-clone}{\code{Get2TweetsFirehoseStreamResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsFirehoseStreamResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsFirehoseStreamResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsFirehoseStreamResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsFirehoseStreamResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsFirehoseStreamResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsFirehoseStreamResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsFirehoseStreamResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsFirehoseStreamResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsFirehoseStreamResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsFirehoseStreamResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsFirehoseStreamResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsFirehoseStreamResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsFirehoseStreamResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsFirehoseStreamResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsFirehoseStreamResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsFirehoseStreamResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsFirehoseStreamResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsFirehoseStreamResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsFirehoseStreamResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsFirehoseStreamResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsFirehoseStreamResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsFirehoseStreamResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsFirehoseStreamResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsFirehoseStreamResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsFirehoseStreamResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsFirehoseStreamResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsFirehoseStreamResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsFirehoseStreamResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsFirehoseStreamResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsFirehoseStreamResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsFirehoseStreamResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsFirehoseStreamResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsFirehoseStreamResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsFirehoseStreamResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsFirehoseStreamResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsFirehoseStreamResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsFirehoseStreamResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsFirehoseStreamResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsFirehoseStreamResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
