% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_id_liking_users_response.R
\docType{class}
\name{Get2TweetsIdLikingUsersResponse}
\alias{Get2TweetsIdLikingUsersResponse}
\title{Get2TweetsIdLikingUsersResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsIdLikingUsersResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{User} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ListsIdFollowersResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsIdLikingUsersResponse-new}{\code{Get2TweetsIdLikingUsersResponse$new()}}
\item \href{#method-Get2TweetsIdLikingUsersResponse-toJSON}{\code{Get2TweetsIdLikingUsersResponse$toJSON()}}
\item \href{#method-Get2TweetsIdLikingUsersResponse-fromJSON}{\code{Get2TweetsIdLikingUsersResponse$fromJSON()}}
\item \href{#method-Get2TweetsIdLikingUsersResponse-toJSONString}{\code{Get2TweetsIdLikingUsersResponse$toJSONString()}}
\item \href{#method-Get2TweetsIdLikingUsersResponse-fromJSONString}{\code{Get2TweetsIdLikingUsersResponse$fromJSONString()}}
\item \href{#method-Get2TweetsIdLikingUsersResponse-validateJSON}{\code{Get2TweetsIdLikingUsersResponse$validateJSON()}}
\item \href{#method-Get2TweetsIdLikingUsersResponse-toString}{\code{Get2TweetsIdLikingUsersResponse$toString()}}
\item \href{#method-Get2TweetsIdLikingUsersResponse-isValid}{\code{Get2TweetsIdLikingUsersResponse$isValid()}}
\item \href{#method-Get2TweetsIdLikingUsersResponse-getInvalidFields}{\code{Get2TweetsIdLikingUsersResponse$getInvalidFields()}}
\item \href{#method-Get2TweetsIdLikingUsersResponse-clone}{\code{Get2TweetsIdLikingUsersResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdLikingUsersResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdLikingUsersResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsIdLikingUsersResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdLikingUsersResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdLikingUsersResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdLikingUsersResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdLikingUsersResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsIdLikingUsersResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdLikingUsersResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdLikingUsersResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsIdLikingUsersResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdLikingUsersResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsIdLikingUsersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdLikingUsersResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdLikingUsersResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdLikingUsersResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsIdLikingUsersResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdLikingUsersResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdLikingUsersResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsIdLikingUsersResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdLikingUsersResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsIdLikingUsersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdLikingUsersResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdLikingUsersResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsIdLikingUsersResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdLikingUsersResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdLikingUsersResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdLikingUsersResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdLikingUsersResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsIdLikingUsersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdLikingUsersResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdLikingUsersResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdLikingUsersResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdLikingUsersResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdLikingUsersResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdLikingUsersResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdLikingUsersResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdLikingUsersResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdLikingUsersResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
