% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_id_quote_tweets_response.R
\docType{class}
\name{Get2TweetsIdQuoteTweetsResponse}
\alias{Get2TweetsIdQuoteTweetsResponse}
\title{Get2TweetsIdQuoteTweetsResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsIdQuoteTweetsResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{Tweet} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2TweetsIdQuoteTweetsResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsIdQuoteTweetsResponse-new}{\code{Get2TweetsIdQuoteTweetsResponse$new()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponse-toJSON}{\code{Get2TweetsIdQuoteTweetsResponse$toJSON()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponse-fromJSON}{\code{Get2TweetsIdQuoteTweetsResponse$fromJSON()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponse-toJSONString}{\code{Get2TweetsIdQuoteTweetsResponse$toJSONString()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponse-fromJSONString}{\code{Get2TweetsIdQuoteTweetsResponse$fromJSONString()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponse-validateJSON}{\code{Get2TweetsIdQuoteTweetsResponse$validateJSON()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponse-toString}{\code{Get2TweetsIdQuoteTweetsResponse$toString()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponse-isValid}{\code{Get2TweetsIdQuoteTweetsResponse$isValid()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponse-getInvalidFields}{\code{Get2TweetsIdQuoteTweetsResponse$getInvalidFields()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponse-clone}{\code{Get2TweetsIdQuoteTweetsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsIdQuoteTweetsResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsIdQuoteTweetsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsIdQuoteTweetsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsIdQuoteTweetsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsIdQuoteTweetsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsIdQuoteTweetsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsIdQuoteTweetsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsIdQuoteTweetsResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsIdQuoteTweetsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
