% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_id_quote_tweets_response_meta.R
\docType{class}
\name{Get2TweetsIdQuoteTweetsResponseMeta}
\alias{Get2TweetsIdQuoteTweetsResponseMeta}
\title{Get2TweetsIdQuoteTweetsResponseMeta}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsIdQuoteTweetsResponseMeta Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{next_token}}{character [optional]}

\item{\code{result_count}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsIdQuoteTweetsResponseMeta-new}{\code{Get2TweetsIdQuoteTweetsResponseMeta$new()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponseMeta-toJSON}{\code{Get2TweetsIdQuoteTweetsResponseMeta$toJSON()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponseMeta-fromJSON}{\code{Get2TweetsIdQuoteTweetsResponseMeta$fromJSON()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponseMeta-toJSONString}{\code{Get2TweetsIdQuoteTweetsResponseMeta$toJSONString()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponseMeta-fromJSONString}{\code{Get2TweetsIdQuoteTweetsResponseMeta$fromJSONString()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponseMeta-validateJSON}{\code{Get2TweetsIdQuoteTweetsResponseMeta$validateJSON()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponseMeta-toString}{\code{Get2TweetsIdQuoteTweetsResponseMeta$toString()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponseMeta-isValid}{\code{Get2TweetsIdQuoteTweetsResponseMeta$isValid()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponseMeta-getInvalidFields}{\code{Get2TweetsIdQuoteTweetsResponseMeta$getInvalidFields()}}
\item \href{#method-Get2TweetsIdQuoteTweetsResponseMeta-clone}{\code{Get2TweetsIdQuoteTweetsResponseMeta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponseMeta-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponseMeta-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsIdQuoteTweetsResponseMeta class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponseMeta$new(
  next_token = NULL,
  result_count = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{next_token}}{The next token.}

\item{\code{result_count}}{The number of results returned in this response.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponseMeta-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponseMeta-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponseMeta$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsIdQuoteTweetsResponseMeta in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponseMeta-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponseMeta-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsIdQuoteTweetsResponseMeta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponseMeta$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsIdQuoteTweetsResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponseMeta-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponseMeta-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponseMeta$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsIdQuoteTweetsResponseMeta in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponseMeta-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponseMeta-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsIdQuoteTweetsResponseMeta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponseMeta$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsIdQuoteTweetsResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponseMeta-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponseMeta-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsIdQuoteTweetsResponseMeta and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponseMeta$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponseMeta-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponseMeta-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponseMeta$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsIdQuoteTweetsResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponseMeta-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponseMeta-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponseMeta$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponseMeta-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponseMeta-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponseMeta$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdQuoteTweetsResponseMeta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdQuoteTweetsResponseMeta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdQuoteTweetsResponseMeta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
