% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_id_response.R
\docType{class}
\name{Get2TweetsIdResponse}
\alias{Get2TweetsIdResponse}
\title{Get2TweetsIdResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsIdResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{Tweet} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsIdResponse-new}{\code{Get2TweetsIdResponse$new()}}
\item \href{#method-Get2TweetsIdResponse-toJSON}{\code{Get2TweetsIdResponse$toJSON()}}
\item \href{#method-Get2TweetsIdResponse-fromJSON}{\code{Get2TweetsIdResponse$fromJSON()}}
\item \href{#method-Get2TweetsIdResponse-toJSONString}{\code{Get2TweetsIdResponse$toJSONString()}}
\item \href{#method-Get2TweetsIdResponse-fromJSONString}{\code{Get2TweetsIdResponse$fromJSONString()}}
\item \href{#method-Get2TweetsIdResponse-validateJSON}{\code{Get2TweetsIdResponse$validateJSON()}}
\item \href{#method-Get2TweetsIdResponse-toString}{\code{Get2TweetsIdResponse$toString()}}
\item \href{#method-Get2TweetsIdResponse-isValid}{\code{Get2TweetsIdResponse$isValid()}}
\item \href{#method-Get2TweetsIdResponse-getInvalidFields}{\code{Get2TweetsIdResponse$getInvalidFields()}}
\item \href{#method-Get2TweetsIdResponse-clone}{\code{Get2TweetsIdResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsIdResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdResponse$new(data = NULL, errors = NULL, includes = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsIdResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsIdResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsIdResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsIdResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsIdResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsIdResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsIdResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsIdResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
