% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_id_retweeted_by_response.R
\docType{class}
\name{Get2TweetsIdRetweetedByResponse}
\alias{Get2TweetsIdRetweetedByResponse}
\title{Get2TweetsIdRetweetedByResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsIdRetweetedByResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{User} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ListsIdFollowersResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsIdRetweetedByResponse-new}{\code{Get2TweetsIdRetweetedByResponse$new()}}
\item \href{#method-Get2TweetsIdRetweetedByResponse-toJSON}{\code{Get2TweetsIdRetweetedByResponse$toJSON()}}
\item \href{#method-Get2TweetsIdRetweetedByResponse-fromJSON}{\code{Get2TweetsIdRetweetedByResponse$fromJSON()}}
\item \href{#method-Get2TweetsIdRetweetedByResponse-toJSONString}{\code{Get2TweetsIdRetweetedByResponse$toJSONString()}}
\item \href{#method-Get2TweetsIdRetweetedByResponse-fromJSONString}{\code{Get2TweetsIdRetweetedByResponse$fromJSONString()}}
\item \href{#method-Get2TweetsIdRetweetedByResponse-validateJSON}{\code{Get2TweetsIdRetweetedByResponse$validateJSON()}}
\item \href{#method-Get2TweetsIdRetweetedByResponse-toString}{\code{Get2TweetsIdRetweetedByResponse$toString()}}
\item \href{#method-Get2TweetsIdRetweetedByResponse-isValid}{\code{Get2TweetsIdRetweetedByResponse$isValid()}}
\item \href{#method-Get2TweetsIdRetweetedByResponse-getInvalidFields}{\code{Get2TweetsIdRetweetedByResponse$getInvalidFields()}}
\item \href{#method-Get2TweetsIdRetweetedByResponse-clone}{\code{Get2TweetsIdRetweetedByResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdRetweetedByResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdRetweetedByResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsIdRetweetedByResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdRetweetedByResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdRetweetedByResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdRetweetedByResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdRetweetedByResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsIdRetweetedByResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdRetweetedByResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdRetweetedByResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsIdRetweetedByResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdRetweetedByResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsIdRetweetedByResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdRetweetedByResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdRetweetedByResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdRetweetedByResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsIdRetweetedByResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdRetweetedByResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdRetweetedByResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsIdRetweetedByResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdRetweetedByResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsIdRetweetedByResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdRetweetedByResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdRetweetedByResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsIdRetweetedByResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdRetweetedByResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdRetweetedByResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdRetweetedByResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdRetweetedByResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsIdRetweetedByResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdRetweetedByResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdRetweetedByResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdRetweetedByResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdRetweetedByResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdRetweetedByResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdRetweetedByResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsIdRetweetedByResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsIdRetweetedByResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsIdRetweetedByResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
