% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_sample10_stream_response.R
\docType{class}
\name{Get2TweetsSample10StreamResponse}
\alias{Get2TweetsSample10StreamResponse}
\title{Get2TweetsSample10StreamResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsSample10StreamResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{Tweet} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsSample10StreamResponse-new}{\code{Get2TweetsSample10StreamResponse$new()}}
\item \href{#method-Get2TweetsSample10StreamResponse-toJSON}{\code{Get2TweetsSample10StreamResponse$toJSON()}}
\item \href{#method-Get2TweetsSample10StreamResponse-fromJSON}{\code{Get2TweetsSample10StreamResponse$fromJSON()}}
\item \href{#method-Get2TweetsSample10StreamResponse-toJSONString}{\code{Get2TweetsSample10StreamResponse$toJSONString()}}
\item \href{#method-Get2TweetsSample10StreamResponse-fromJSONString}{\code{Get2TweetsSample10StreamResponse$fromJSONString()}}
\item \href{#method-Get2TweetsSample10StreamResponse-validateJSON}{\code{Get2TweetsSample10StreamResponse$validateJSON()}}
\item \href{#method-Get2TweetsSample10StreamResponse-toString}{\code{Get2TweetsSample10StreamResponse$toString()}}
\item \href{#method-Get2TweetsSample10StreamResponse-isValid}{\code{Get2TweetsSample10StreamResponse$isValid()}}
\item \href{#method-Get2TweetsSample10StreamResponse-getInvalidFields}{\code{Get2TweetsSample10StreamResponse$getInvalidFields()}}
\item \href{#method-Get2TweetsSample10StreamResponse-clone}{\code{Get2TweetsSample10StreamResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSample10StreamResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSample10StreamResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsSample10StreamResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSample10StreamResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSample10StreamResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSample10StreamResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSample10StreamResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsSample10StreamResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSample10StreamResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSample10StreamResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsSample10StreamResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSample10StreamResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsSample10StreamResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSample10StreamResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSample10StreamResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSample10StreamResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsSample10StreamResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSample10StreamResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSample10StreamResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsSample10StreamResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSample10StreamResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsSample10StreamResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSample10StreamResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSample10StreamResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsSample10StreamResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSample10StreamResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSample10StreamResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSample10StreamResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSample10StreamResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsSample10StreamResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSample10StreamResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSample10StreamResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSample10StreamResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSample10StreamResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSample10StreamResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSample10StreamResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSample10StreamResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSample10StreamResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSample10StreamResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
