% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_sample_stream_response.R
\docType{class}
\name{Get2TweetsSampleStreamResponse}
\alias{Get2TweetsSampleStreamResponse}
\title{Get2TweetsSampleStreamResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsSampleStreamResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{Tweet} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsSampleStreamResponse-new}{\code{Get2TweetsSampleStreamResponse$new()}}
\item \href{#method-Get2TweetsSampleStreamResponse-toJSON}{\code{Get2TweetsSampleStreamResponse$toJSON()}}
\item \href{#method-Get2TweetsSampleStreamResponse-fromJSON}{\code{Get2TweetsSampleStreamResponse$fromJSON()}}
\item \href{#method-Get2TweetsSampleStreamResponse-toJSONString}{\code{Get2TweetsSampleStreamResponse$toJSONString()}}
\item \href{#method-Get2TweetsSampleStreamResponse-fromJSONString}{\code{Get2TweetsSampleStreamResponse$fromJSONString()}}
\item \href{#method-Get2TweetsSampleStreamResponse-validateJSON}{\code{Get2TweetsSampleStreamResponse$validateJSON()}}
\item \href{#method-Get2TweetsSampleStreamResponse-toString}{\code{Get2TweetsSampleStreamResponse$toString()}}
\item \href{#method-Get2TweetsSampleStreamResponse-isValid}{\code{Get2TweetsSampleStreamResponse$isValid()}}
\item \href{#method-Get2TweetsSampleStreamResponse-getInvalidFields}{\code{Get2TweetsSampleStreamResponse$getInvalidFields()}}
\item \href{#method-Get2TweetsSampleStreamResponse-clone}{\code{Get2TweetsSampleStreamResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSampleStreamResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSampleStreamResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsSampleStreamResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSampleStreamResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSampleStreamResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSampleStreamResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSampleStreamResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsSampleStreamResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSampleStreamResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSampleStreamResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsSampleStreamResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSampleStreamResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsSampleStreamResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSampleStreamResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSampleStreamResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSampleStreamResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsSampleStreamResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSampleStreamResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSampleStreamResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsSampleStreamResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSampleStreamResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsSampleStreamResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSampleStreamResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSampleStreamResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsSampleStreamResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSampleStreamResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSampleStreamResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSampleStreamResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSampleStreamResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsSampleStreamResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSampleStreamResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSampleStreamResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSampleStreamResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSampleStreamResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSampleStreamResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSampleStreamResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSampleStreamResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSampleStreamResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSampleStreamResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
