% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_search_all_response.R
\docType{class}
\name{Get2TweetsSearchAllResponse}
\alias{Get2TweetsSearchAllResponse}
\title{Get2TweetsSearchAllResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsSearchAllResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{Tweet} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2TweetsSearchAllResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsSearchAllResponse-new}{\code{Get2TweetsSearchAllResponse$new()}}
\item \href{#method-Get2TweetsSearchAllResponse-toJSON}{\code{Get2TweetsSearchAllResponse$toJSON()}}
\item \href{#method-Get2TweetsSearchAllResponse-fromJSON}{\code{Get2TweetsSearchAllResponse$fromJSON()}}
\item \href{#method-Get2TweetsSearchAllResponse-toJSONString}{\code{Get2TweetsSearchAllResponse$toJSONString()}}
\item \href{#method-Get2TweetsSearchAllResponse-fromJSONString}{\code{Get2TweetsSearchAllResponse$fromJSONString()}}
\item \href{#method-Get2TweetsSearchAllResponse-validateJSON}{\code{Get2TweetsSearchAllResponse$validateJSON()}}
\item \href{#method-Get2TweetsSearchAllResponse-toString}{\code{Get2TweetsSearchAllResponse$toString()}}
\item \href{#method-Get2TweetsSearchAllResponse-isValid}{\code{Get2TweetsSearchAllResponse$isValid()}}
\item \href{#method-Get2TweetsSearchAllResponse-getInvalidFields}{\code{Get2TweetsSearchAllResponse$getInvalidFields()}}
\item \href{#method-Get2TweetsSearchAllResponse-clone}{\code{Get2TweetsSearchAllResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsSearchAllResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsSearchAllResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsSearchAllResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsSearchAllResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsSearchAllResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsSearchAllResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsSearchAllResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsSearchAllResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsSearchAllResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
