% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_search_all_response_meta.R
\docType{class}
\name{Get2TweetsSearchAllResponseMeta}
\alias{Get2TweetsSearchAllResponseMeta}
\title{Get2TweetsSearchAllResponseMeta}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsSearchAllResponseMeta Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{newest_id}}{character [optional]}

\item{\code{next_token}}{character [optional]}

\item{\code{oldest_id}}{character [optional]}

\item{\code{result_count}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsSearchAllResponseMeta-new}{\code{Get2TweetsSearchAllResponseMeta$new()}}
\item \href{#method-Get2TweetsSearchAllResponseMeta-toJSON}{\code{Get2TweetsSearchAllResponseMeta$toJSON()}}
\item \href{#method-Get2TweetsSearchAllResponseMeta-fromJSON}{\code{Get2TweetsSearchAllResponseMeta$fromJSON()}}
\item \href{#method-Get2TweetsSearchAllResponseMeta-toJSONString}{\code{Get2TweetsSearchAllResponseMeta$toJSONString()}}
\item \href{#method-Get2TweetsSearchAllResponseMeta-fromJSONString}{\code{Get2TweetsSearchAllResponseMeta$fromJSONString()}}
\item \href{#method-Get2TweetsSearchAllResponseMeta-validateJSON}{\code{Get2TweetsSearchAllResponseMeta$validateJSON()}}
\item \href{#method-Get2TweetsSearchAllResponseMeta-toString}{\code{Get2TweetsSearchAllResponseMeta$toString()}}
\item \href{#method-Get2TweetsSearchAllResponseMeta-isValid}{\code{Get2TweetsSearchAllResponseMeta$isValid()}}
\item \href{#method-Get2TweetsSearchAllResponseMeta-getInvalidFields}{\code{Get2TweetsSearchAllResponseMeta$getInvalidFields()}}
\item \href{#method-Get2TweetsSearchAllResponseMeta-clone}{\code{Get2TweetsSearchAllResponseMeta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponseMeta-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponseMeta-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsSearchAllResponseMeta class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponseMeta$new(
  newest_id = NULL,
  next_token = NULL,
  oldest_id = NULL,
  result_count = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newest_id}}{The newest id in this response.}

\item{\code{next_token}}{The next token.}

\item{\code{oldest_id}}{The oldest id in this response.}

\item{\code{result_count}}{The number of results returned in this response.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponseMeta-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponseMeta-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponseMeta$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsSearchAllResponseMeta in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponseMeta-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponseMeta-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsSearchAllResponseMeta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponseMeta$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsSearchAllResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponseMeta-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponseMeta-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponseMeta$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsSearchAllResponseMeta in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponseMeta-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponseMeta-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsSearchAllResponseMeta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponseMeta$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsSearchAllResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponseMeta-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponseMeta-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsSearchAllResponseMeta and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponseMeta$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponseMeta-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponseMeta-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponseMeta$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsSearchAllResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponseMeta-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponseMeta-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponseMeta$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponseMeta-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponseMeta-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponseMeta$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchAllResponseMeta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchAllResponseMeta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchAllResponseMeta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
