% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_search_recent_response.R
\docType{class}
\name{Get2TweetsSearchRecentResponse}
\alias{Get2TweetsSearchRecentResponse}
\title{Get2TweetsSearchRecentResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsSearchRecentResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{Tweet} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2TweetsSearchAllResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsSearchRecentResponse-new}{\code{Get2TweetsSearchRecentResponse$new()}}
\item \href{#method-Get2TweetsSearchRecentResponse-toJSON}{\code{Get2TweetsSearchRecentResponse$toJSON()}}
\item \href{#method-Get2TweetsSearchRecentResponse-fromJSON}{\code{Get2TweetsSearchRecentResponse$fromJSON()}}
\item \href{#method-Get2TweetsSearchRecentResponse-toJSONString}{\code{Get2TweetsSearchRecentResponse$toJSONString()}}
\item \href{#method-Get2TweetsSearchRecentResponse-fromJSONString}{\code{Get2TweetsSearchRecentResponse$fromJSONString()}}
\item \href{#method-Get2TweetsSearchRecentResponse-validateJSON}{\code{Get2TweetsSearchRecentResponse$validateJSON()}}
\item \href{#method-Get2TweetsSearchRecentResponse-toString}{\code{Get2TweetsSearchRecentResponse$toString()}}
\item \href{#method-Get2TweetsSearchRecentResponse-isValid}{\code{Get2TweetsSearchRecentResponse$isValid()}}
\item \href{#method-Get2TweetsSearchRecentResponse-getInvalidFields}{\code{Get2TweetsSearchRecentResponse$getInvalidFields()}}
\item \href{#method-Get2TweetsSearchRecentResponse-clone}{\code{Get2TweetsSearchRecentResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchRecentResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchRecentResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsSearchRecentResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchRecentResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchRecentResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchRecentResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchRecentResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsSearchRecentResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchRecentResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchRecentResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsSearchRecentResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchRecentResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsSearchRecentResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchRecentResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchRecentResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchRecentResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsSearchRecentResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchRecentResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchRecentResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsSearchRecentResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchRecentResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsSearchRecentResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchRecentResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchRecentResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsSearchRecentResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchRecentResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchRecentResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchRecentResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchRecentResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsSearchRecentResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchRecentResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchRecentResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchRecentResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchRecentResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchRecentResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchRecentResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchRecentResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchRecentResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchRecentResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
