% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_tweets_search_stream_response.R
\docType{class}
\name{Get2TweetsSearchStreamResponse}
\alias{Get2TweetsSearchStreamResponse}
\title{Get2TweetsSearchStreamResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2TweetsSearchStreamResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{Tweet} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2TweetsSearchStreamResponse-new}{\code{Get2TweetsSearchStreamResponse$new()}}
\item \href{#method-Get2TweetsSearchStreamResponse-toJSON}{\code{Get2TweetsSearchStreamResponse$toJSON()}}
\item \href{#method-Get2TweetsSearchStreamResponse-fromJSON}{\code{Get2TweetsSearchStreamResponse$fromJSON()}}
\item \href{#method-Get2TweetsSearchStreamResponse-toJSONString}{\code{Get2TweetsSearchStreamResponse$toJSONString()}}
\item \href{#method-Get2TweetsSearchStreamResponse-fromJSONString}{\code{Get2TweetsSearchStreamResponse$fromJSONString()}}
\item \href{#method-Get2TweetsSearchStreamResponse-validateJSON}{\code{Get2TweetsSearchStreamResponse$validateJSON()}}
\item \href{#method-Get2TweetsSearchStreamResponse-toString}{\code{Get2TweetsSearchStreamResponse$toString()}}
\item \href{#method-Get2TweetsSearchStreamResponse-isValid}{\code{Get2TweetsSearchStreamResponse$isValid()}}
\item \href{#method-Get2TweetsSearchStreamResponse-getInvalidFields}{\code{Get2TweetsSearchStreamResponse$getInvalidFields()}}
\item \href{#method-Get2TweetsSearchStreamResponse-clone}{\code{Get2TweetsSearchStreamResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchStreamResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchStreamResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2TweetsSearchStreamResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchStreamResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchStreamResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchStreamResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchStreamResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsSearchStreamResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchStreamResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchStreamResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2TweetsSearchStreamResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchStreamResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsSearchStreamResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchStreamResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchStreamResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchStreamResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2TweetsSearchStreamResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchStreamResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchStreamResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2TweetsSearchStreamResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchStreamResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2TweetsSearchStreamResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchStreamResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchStreamResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2TweetsSearchStreamResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchStreamResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchStreamResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchStreamResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchStreamResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2TweetsSearchStreamResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchStreamResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchStreamResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchStreamResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchStreamResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchStreamResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchStreamResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2TweetsSearchStreamResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2TweetsSearchStreamResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2TweetsSearchStreamResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
