% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_by_response.R
\docType{class}
\name{Get2UsersByResponse}
\alias{Get2UsersByResponse}
\title{Get2UsersByResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersByResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{User} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersByResponse-new}{\code{Get2UsersByResponse$new()}}
\item \href{#method-Get2UsersByResponse-toJSON}{\code{Get2UsersByResponse$toJSON()}}
\item \href{#method-Get2UsersByResponse-fromJSON}{\code{Get2UsersByResponse$fromJSON()}}
\item \href{#method-Get2UsersByResponse-toJSONString}{\code{Get2UsersByResponse$toJSONString()}}
\item \href{#method-Get2UsersByResponse-fromJSONString}{\code{Get2UsersByResponse$fromJSONString()}}
\item \href{#method-Get2UsersByResponse-validateJSON}{\code{Get2UsersByResponse$validateJSON()}}
\item \href{#method-Get2UsersByResponse-toString}{\code{Get2UsersByResponse$toString()}}
\item \href{#method-Get2UsersByResponse-isValid}{\code{Get2UsersByResponse$isValid()}}
\item \href{#method-Get2UsersByResponse-getInvalidFields}{\code{Get2UsersByResponse$getInvalidFields()}}
\item \href{#method-Get2UsersByResponse-clone}{\code{Get2UsersByResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersByResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByResponse$new(data = NULL, errors = NULL, includes = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersByResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersByResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersByResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersByResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersByResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersByResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersByResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersByResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
