% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_by_username_username_response.R
\docType{class}
\name{Get2UsersByUsernameUsernameResponse}
\alias{Get2UsersByUsernameUsernameResponse}
\title{Get2UsersByUsernameUsernameResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersByUsernameUsernameResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{User} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersByUsernameUsernameResponse-new}{\code{Get2UsersByUsernameUsernameResponse$new()}}
\item \href{#method-Get2UsersByUsernameUsernameResponse-toJSON}{\code{Get2UsersByUsernameUsernameResponse$toJSON()}}
\item \href{#method-Get2UsersByUsernameUsernameResponse-fromJSON}{\code{Get2UsersByUsernameUsernameResponse$fromJSON()}}
\item \href{#method-Get2UsersByUsernameUsernameResponse-toJSONString}{\code{Get2UsersByUsernameUsernameResponse$toJSONString()}}
\item \href{#method-Get2UsersByUsernameUsernameResponse-fromJSONString}{\code{Get2UsersByUsernameUsernameResponse$fromJSONString()}}
\item \href{#method-Get2UsersByUsernameUsernameResponse-validateJSON}{\code{Get2UsersByUsernameUsernameResponse$validateJSON()}}
\item \href{#method-Get2UsersByUsernameUsernameResponse-toString}{\code{Get2UsersByUsernameUsernameResponse$toString()}}
\item \href{#method-Get2UsersByUsernameUsernameResponse-isValid}{\code{Get2UsersByUsernameUsernameResponse$isValid()}}
\item \href{#method-Get2UsersByUsernameUsernameResponse-getInvalidFields}{\code{Get2UsersByUsernameUsernameResponse$getInvalidFields()}}
\item \href{#method-Get2UsersByUsernameUsernameResponse-clone}{\code{Get2UsersByUsernameUsernameResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByUsernameUsernameResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByUsernameUsernameResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersByUsernameUsernameResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByUsernameUsernameResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByUsernameUsernameResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByUsernameUsernameResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByUsernameUsernameResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersByUsernameUsernameResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByUsernameUsernameResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByUsernameUsernameResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersByUsernameUsernameResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByUsernameUsernameResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersByUsernameUsernameResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByUsernameUsernameResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByUsernameUsernameResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByUsernameUsernameResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersByUsernameUsernameResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByUsernameUsernameResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByUsernameUsernameResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersByUsernameUsernameResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByUsernameUsernameResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersByUsernameUsernameResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByUsernameUsernameResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByUsernameUsernameResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersByUsernameUsernameResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByUsernameUsernameResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByUsernameUsernameResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByUsernameUsernameResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByUsernameUsernameResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersByUsernameUsernameResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByUsernameUsernameResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByUsernameUsernameResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByUsernameUsernameResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByUsernameUsernameResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByUsernameUsernameResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByUsernameUsernameResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersByUsernameUsernameResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersByUsernameUsernameResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersByUsernameUsernameResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
