% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_id_blocking_response.R
\docType{class}
\name{Get2UsersIdBlockingResponse}
\alias{Get2UsersIdBlockingResponse}
\title{Get2UsersIdBlockingResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersIdBlockingResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{User} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ListsIdFollowersResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersIdBlockingResponse-new}{\code{Get2UsersIdBlockingResponse$new()}}
\item \href{#method-Get2UsersIdBlockingResponse-toJSON}{\code{Get2UsersIdBlockingResponse$toJSON()}}
\item \href{#method-Get2UsersIdBlockingResponse-fromJSON}{\code{Get2UsersIdBlockingResponse$fromJSON()}}
\item \href{#method-Get2UsersIdBlockingResponse-toJSONString}{\code{Get2UsersIdBlockingResponse$toJSONString()}}
\item \href{#method-Get2UsersIdBlockingResponse-fromJSONString}{\code{Get2UsersIdBlockingResponse$fromJSONString()}}
\item \href{#method-Get2UsersIdBlockingResponse-validateJSON}{\code{Get2UsersIdBlockingResponse$validateJSON()}}
\item \href{#method-Get2UsersIdBlockingResponse-toString}{\code{Get2UsersIdBlockingResponse$toString()}}
\item \href{#method-Get2UsersIdBlockingResponse-isValid}{\code{Get2UsersIdBlockingResponse$isValid()}}
\item \href{#method-Get2UsersIdBlockingResponse-getInvalidFields}{\code{Get2UsersIdBlockingResponse$getInvalidFields()}}
\item \href{#method-Get2UsersIdBlockingResponse-clone}{\code{Get2UsersIdBlockingResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBlockingResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBlockingResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersIdBlockingResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBlockingResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBlockingResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBlockingResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBlockingResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdBlockingResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBlockingResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBlockingResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersIdBlockingResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBlockingResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdBlockingResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBlockingResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBlockingResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBlockingResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdBlockingResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBlockingResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBlockingResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersIdBlockingResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBlockingResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdBlockingResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBlockingResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBlockingResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersIdBlockingResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBlockingResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBlockingResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBlockingResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBlockingResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersIdBlockingResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBlockingResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBlockingResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBlockingResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBlockingResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBlockingResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBlockingResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBlockingResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBlockingResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBlockingResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
