% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_id_bookmarks_response.R
\docType{class}
\name{Get2UsersIdBookmarksResponse}
\alias{Get2UsersIdBookmarksResponse}
\title{Get2UsersIdBookmarksResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersIdBookmarksResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{Tweet} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ListsIdFollowersResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersIdBookmarksResponse-new}{\code{Get2UsersIdBookmarksResponse$new()}}
\item \href{#method-Get2UsersIdBookmarksResponse-toJSON}{\code{Get2UsersIdBookmarksResponse$toJSON()}}
\item \href{#method-Get2UsersIdBookmarksResponse-fromJSON}{\code{Get2UsersIdBookmarksResponse$fromJSON()}}
\item \href{#method-Get2UsersIdBookmarksResponse-toJSONString}{\code{Get2UsersIdBookmarksResponse$toJSONString()}}
\item \href{#method-Get2UsersIdBookmarksResponse-fromJSONString}{\code{Get2UsersIdBookmarksResponse$fromJSONString()}}
\item \href{#method-Get2UsersIdBookmarksResponse-validateJSON}{\code{Get2UsersIdBookmarksResponse$validateJSON()}}
\item \href{#method-Get2UsersIdBookmarksResponse-toString}{\code{Get2UsersIdBookmarksResponse$toString()}}
\item \href{#method-Get2UsersIdBookmarksResponse-isValid}{\code{Get2UsersIdBookmarksResponse$isValid()}}
\item \href{#method-Get2UsersIdBookmarksResponse-getInvalidFields}{\code{Get2UsersIdBookmarksResponse$getInvalidFields()}}
\item \href{#method-Get2UsersIdBookmarksResponse-clone}{\code{Get2UsersIdBookmarksResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBookmarksResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBookmarksResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersIdBookmarksResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBookmarksResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBookmarksResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBookmarksResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBookmarksResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdBookmarksResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBookmarksResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBookmarksResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersIdBookmarksResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBookmarksResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdBookmarksResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBookmarksResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBookmarksResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBookmarksResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdBookmarksResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBookmarksResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBookmarksResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersIdBookmarksResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBookmarksResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdBookmarksResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBookmarksResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBookmarksResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersIdBookmarksResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBookmarksResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBookmarksResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBookmarksResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBookmarksResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersIdBookmarksResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBookmarksResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBookmarksResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBookmarksResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBookmarksResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBookmarksResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBookmarksResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdBookmarksResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdBookmarksResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdBookmarksResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
