% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_id_followed_lists_response.R
\docType{class}
\name{Get2UsersIdFollowedListsResponse}
\alias{Get2UsersIdFollowedListsResponse}
\title{Get2UsersIdFollowedListsResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersIdFollowedListsResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{List} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ListsIdFollowersResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersIdFollowedListsResponse-new}{\code{Get2UsersIdFollowedListsResponse$new()}}
\item \href{#method-Get2UsersIdFollowedListsResponse-toJSON}{\code{Get2UsersIdFollowedListsResponse$toJSON()}}
\item \href{#method-Get2UsersIdFollowedListsResponse-fromJSON}{\code{Get2UsersIdFollowedListsResponse$fromJSON()}}
\item \href{#method-Get2UsersIdFollowedListsResponse-toJSONString}{\code{Get2UsersIdFollowedListsResponse$toJSONString()}}
\item \href{#method-Get2UsersIdFollowedListsResponse-fromJSONString}{\code{Get2UsersIdFollowedListsResponse$fromJSONString()}}
\item \href{#method-Get2UsersIdFollowedListsResponse-validateJSON}{\code{Get2UsersIdFollowedListsResponse$validateJSON()}}
\item \href{#method-Get2UsersIdFollowedListsResponse-toString}{\code{Get2UsersIdFollowedListsResponse$toString()}}
\item \href{#method-Get2UsersIdFollowedListsResponse-isValid}{\code{Get2UsersIdFollowedListsResponse$isValid()}}
\item \href{#method-Get2UsersIdFollowedListsResponse-getInvalidFields}{\code{Get2UsersIdFollowedListsResponse$getInvalidFields()}}
\item \href{#method-Get2UsersIdFollowedListsResponse-clone}{\code{Get2UsersIdFollowedListsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowedListsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowedListsResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersIdFollowedListsResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowedListsResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowedListsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowedListsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowedListsResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdFollowedListsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowedListsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowedListsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersIdFollowedListsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowedListsResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdFollowedListsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowedListsResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowedListsResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowedListsResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdFollowedListsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowedListsResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowedListsResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersIdFollowedListsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowedListsResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdFollowedListsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowedListsResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowedListsResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersIdFollowedListsResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowedListsResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowedListsResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowedListsResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowedListsResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersIdFollowedListsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowedListsResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowedListsResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowedListsResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowedListsResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowedListsResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowedListsResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowedListsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowedListsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowedListsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
