% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_id_followers_response.R
\docType{class}
\name{Get2UsersIdFollowersResponse}
\alias{Get2UsersIdFollowersResponse}
\title{Get2UsersIdFollowersResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersIdFollowersResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{User} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ListsIdFollowersResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersIdFollowersResponse-new}{\code{Get2UsersIdFollowersResponse$new()}}
\item \href{#method-Get2UsersIdFollowersResponse-toJSON}{\code{Get2UsersIdFollowersResponse$toJSON()}}
\item \href{#method-Get2UsersIdFollowersResponse-fromJSON}{\code{Get2UsersIdFollowersResponse$fromJSON()}}
\item \href{#method-Get2UsersIdFollowersResponse-toJSONString}{\code{Get2UsersIdFollowersResponse$toJSONString()}}
\item \href{#method-Get2UsersIdFollowersResponse-fromJSONString}{\code{Get2UsersIdFollowersResponse$fromJSONString()}}
\item \href{#method-Get2UsersIdFollowersResponse-validateJSON}{\code{Get2UsersIdFollowersResponse$validateJSON()}}
\item \href{#method-Get2UsersIdFollowersResponse-toString}{\code{Get2UsersIdFollowersResponse$toString()}}
\item \href{#method-Get2UsersIdFollowersResponse-isValid}{\code{Get2UsersIdFollowersResponse$isValid()}}
\item \href{#method-Get2UsersIdFollowersResponse-getInvalidFields}{\code{Get2UsersIdFollowersResponse$getInvalidFields()}}
\item \href{#method-Get2UsersIdFollowersResponse-clone}{\code{Get2UsersIdFollowersResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowersResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowersResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersIdFollowersResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowersResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowersResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowersResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowersResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdFollowersResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowersResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowersResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersIdFollowersResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowersResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdFollowersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowersResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowersResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowersResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdFollowersResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowersResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowersResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersIdFollowersResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowersResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdFollowersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowersResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowersResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersIdFollowersResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowersResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowersResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowersResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowersResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersIdFollowersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowersResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowersResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowersResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowersResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowersResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowersResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowersResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowersResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowersResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
