% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_id_following_response.R
\docType{class}
\name{Get2UsersIdFollowingResponse}
\alias{Get2UsersIdFollowingResponse}
\title{Get2UsersIdFollowingResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersIdFollowingResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{User} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ListsIdFollowersResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersIdFollowingResponse-new}{\code{Get2UsersIdFollowingResponse$new()}}
\item \href{#method-Get2UsersIdFollowingResponse-toJSON}{\code{Get2UsersIdFollowingResponse$toJSON()}}
\item \href{#method-Get2UsersIdFollowingResponse-fromJSON}{\code{Get2UsersIdFollowingResponse$fromJSON()}}
\item \href{#method-Get2UsersIdFollowingResponse-toJSONString}{\code{Get2UsersIdFollowingResponse$toJSONString()}}
\item \href{#method-Get2UsersIdFollowingResponse-fromJSONString}{\code{Get2UsersIdFollowingResponse$fromJSONString()}}
\item \href{#method-Get2UsersIdFollowingResponse-validateJSON}{\code{Get2UsersIdFollowingResponse$validateJSON()}}
\item \href{#method-Get2UsersIdFollowingResponse-toString}{\code{Get2UsersIdFollowingResponse$toString()}}
\item \href{#method-Get2UsersIdFollowingResponse-isValid}{\code{Get2UsersIdFollowingResponse$isValid()}}
\item \href{#method-Get2UsersIdFollowingResponse-getInvalidFields}{\code{Get2UsersIdFollowingResponse$getInvalidFields()}}
\item \href{#method-Get2UsersIdFollowingResponse-clone}{\code{Get2UsersIdFollowingResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowingResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowingResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersIdFollowingResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowingResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowingResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowingResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowingResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdFollowingResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowingResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowingResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersIdFollowingResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowingResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdFollowingResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowingResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowingResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowingResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdFollowingResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowingResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowingResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersIdFollowingResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowingResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdFollowingResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowingResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowingResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersIdFollowingResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowingResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowingResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowingResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowingResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersIdFollowingResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowingResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowingResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowingResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowingResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowingResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowingResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdFollowingResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdFollowingResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdFollowingResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
