% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_id_liked_tweets_response.R
\docType{class}
\name{Get2UsersIdLikedTweetsResponse}
\alias{Get2UsersIdLikedTweetsResponse}
\title{Get2UsersIdLikedTweetsResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersIdLikedTweetsResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{Tweet} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ListsIdFollowersResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersIdLikedTweetsResponse-new}{\code{Get2UsersIdLikedTweetsResponse$new()}}
\item \href{#method-Get2UsersIdLikedTweetsResponse-toJSON}{\code{Get2UsersIdLikedTweetsResponse$toJSON()}}
\item \href{#method-Get2UsersIdLikedTweetsResponse-fromJSON}{\code{Get2UsersIdLikedTweetsResponse$fromJSON()}}
\item \href{#method-Get2UsersIdLikedTweetsResponse-toJSONString}{\code{Get2UsersIdLikedTweetsResponse$toJSONString()}}
\item \href{#method-Get2UsersIdLikedTweetsResponse-fromJSONString}{\code{Get2UsersIdLikedTweetsResponse$fromJSONString()}}
\item \href{#method-Get2UsersIdLikedTweetsResponse-validateJSON}{\code{Get2UsersIdLikedTweetsResponse$validateJSON()}}
\item \href{#method-Get2UsersIdLikedTweetsResponse-toString}{\code{Get2UsersIdLikedTweetsResponse$toString()}}
\item \href{#method-Get2UsersIdLikedTweetsResponse-isValid}{\code{Get2UsersIdLikedTweetsResponse$isValid()}}
\item \href{#method-Get2UsersIdLikedTweetsResponse-getInvalidFields}{\code{Get2UsersIdLikedTweetsResponse$getInvalidFields()}}
\item \href{#method-Get2UsersIdLikedTweetsResponse-clone}{\code{Get2UsersIdLikedTweetsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdLikedTweetsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdLikedTweetsResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersIdLikedTweetsResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdLikedTweetsResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdLikedTweetsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdLikedTweetsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdLikedTweetsResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdLikedTweetsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdLikedTweetsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdLikedTweetsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersIdLikedTweetsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdLikedTweetsResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdLikedTweetsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdLikedTweetsResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdLikedTweetsResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdLikedTweetsResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdLikedTweetsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdLikedTweetsResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdLikedTweetsResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersIdLikedTweetsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdLikedTweetsResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdLikedTweetsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdLikedTweetsResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdLikedTweetsResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersIdLikedTweetsResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdLikedTweetsResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdLikedTweetsResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdLikedTweetsResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdLikedTweetsResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersIdLikedTweetsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdLikedTweetsResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdLikedTweetsResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdLikedTweetsResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdLikedTweetsResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdLikedTweetsResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdLikedTweetsResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdLikedTweetsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdLikedTweetsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdLikedTweetsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
