% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_id_mentions_response.R
\docType{class}
\name{Get2UsersIdMentionsResponse}
\alias{Get2UsersIdMentionsResponse}
\title{Get2UsersIdMentionsResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersIdMentionsResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{Tweet} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2UsersIdMentionsResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersIdMentionsResponse-new}{\code{Get2UsersIdMentionsResponse$new()}}
\item \href{#method-Get2UsersIdMentionsResponse-toJSON}{\code{Get2UsersIdMentionsResponse$toJSON()}}
\item \href{#method-Get2UsersIdMentionsResponse-fromJSON}{\code{Get2UsersIdMentionsResponse$fromJSON()}}
\item \href{#method-Get2UsersIdMentionsResponse-toJSONString}{\code{Get2UsersIdMentionsResponse$toJSONString()}}
\item \href{#method-Get2UsersIdMentionsResponse-fromJSONString}{\code{Get2UsersIdMentionsResponse$fromJSONString()}}
\item \href{#method-Get2UsersIdMentionsResponse-validateJSON}{\code{Get2UsersIdMentionsResponse$validateJSON()}}
\item \href{#method-Get2UsersIdMentionsResponse-toString}{\code{Get2UsersIdMentionsResponse$toString()}}
\item \href{#method-Get2UsersIdMentionsResponse-isValid}{\code{Get2UsersIdMentionsResponse$isValid()}}
\item \href{#method-Get2UsersIdMentionsResponse-getInvalidFields}{\code{Get2UsersIdMentionsResponse$getInvalidFields()}}
\item \href{#method-Get2UsersIdMentionsResponse-clone}{\code{Get2UsersIdMentionsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersIdMentionsResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdMentionsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersIdMentionsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdMentionsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdMentionsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersIdMentionsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdMentionsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersIdMentionsResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersIdMentionsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
