% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_id_mentions_response_meta.R
\docType{class}
\name{Get2UsersIdMentionsResponseMeta}
\alias{Get2UsersIdMentionsResponseMeta}
\title{Get2UsersIdMentionsResponseMeta}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersIdMentionsResponseMeta Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{newest_id}}{character [optional]}

\item{\code{next_token}}{character [optional]}

\item{\code{oldest_id}}{character [optional]}

\item{\code{previous_token}}{character [optional]}

\item{\code{result_count}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersIdMentionsResponseMeta-new}{\code{Get2UsersIdMentionsResponseMeta$new()}}
\item \href{#method-Get2UsersIdMentionsResponseMeta-toJSON}{\code{Get2UsersIdMentionsResponseMeta$toJSON()}}
\item \href{#method-Get2UsersIdMentionsResponseMeta-fromJSON}{\code{Get2UsersIdMentionsResponseMeta$fromJSON()}}
\item \href{#method-Get2UsersIdMentionsResponseMeta-toJSONString}{\code{Get2UsersIdMentionsResponseMeta$toJSONString()}}
\item \href{#method-Get2UsersIdMentionsResponseMeta-fromJSONString}{\code{Get2UsersIdMentionsResponseMeta$fromJSONString()}}
\item \href{#method-Get2UsersIdMentionsResponseMeta-validateJSON}{\code{Get2UsersIdMentionsResponseMeta$validateJSON()}}
\item \href{#method-Get2UsersIdMentionsResponseMeta-toString}{\code{Get2UsersIdMentionsResponseMeta$toString()}}
\item \href{#method-Get2UsersIdMentionsResponseMeta-isValid}{\code{Get2UsersIdMentionsResponseMeta$isValid()}}
\item \href{#method-Get2UsersIdMentionsResponseMeta-getInvalidFields}{\code{Get2UsersIdMentionsResponseMeta$getInvalidFields()}}
\item \href{#method-Get2UsersIdMentionsResponseMeta-clone}{\code{Get2UsersIdMentionsResponseMeta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponseMeta-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponseMeta-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersIdMentionsResponseMeta class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponseMeta$new(
  newest_id = NULL,
  next_token = NULL,
  oldest_id = NULL,
  previous_token = NULL,
  result_count = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newest_id}}{The newest id in this response.}

\item{\code{next_token}}{The next token.}

\item{\code{oldest_id}}{The oldest id in this response.}

\item{\code{previous_token}}{The previous token.}

\item{\code{result_count}}{The number of results returned in this response.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponseMeta-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponseMeta-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponseMeta$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdMentionsResponseMeta in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponseMeta-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponseMeta-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersIdMentionsResponseMeta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponseMeta$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdMentionsResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponseMeta-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponseMeta-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponseMeta$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdMentionsResponseMeta in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponseMeta-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponseMeta-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersIdMentionsResponseMeta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponseMeta$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdMentionsResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponseMeta-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponseMeta-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersIdMentionsResponseMeta and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponseMeta$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponseMeta-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponseMeta-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponseMeta$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersIdMentionsResponseMeta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponseMeta-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponseMeta-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponseMeta$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponseMeta-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponseMeta-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponseMeta$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMentionsResponseMeta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMentionsResponseMeta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMentionsResponseMeta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
