% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_id_muting_response.R
\docType{class}
\name{Get2UsersIdMutingResponse}
\alias{Get2UsersIdMutingResponse}
\title{Get2UsersIdMutingResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersIdMutingResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{User} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ListsIdFollowersResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersIdMutingResponse-new}{\code{Get2UsersIdMutingResponse$new()}}
\item \href{#method-Get2UsersIdMutingResponse-toJSON}{\code{Get2UsersIdMutingResponse$toJSON()}}
\item \href{#method-Get2UsersIdMutingResponse-fromJSON}{\code{Get2UsersIdMutingResponse$fromJSON()}}
\item \href{#method-Get2UsersIdMutingResponse-toJSONString}{\code{Get2UsersIdMutingResponse$toJSONString()}}
\item \href{#method-Get2UsersIdMutingResponse-fromJSONString}{\code{Get2UsersIdMutingResponse$fromJSONString()}}
\item \href{#method-Get2UsersIdMutingResponse-validateJSON}{\code{Get2UsersIdMutingResponse$validateJSON()}}
\item \href{#method-Get2UsersIdMutingResponse-toString}{\code{Get2UsersIdMutingResponse$toString()}}
\item \href{#method-Get2UsersIdMutingResponse-isValid}{\code{Get2UsersIdMutingResponse$isValid()}}
\item \href{#method-Get2UsersIdMutingResponse-getInvalidFields}{\code{Get2UsersIdMutingResponse$getInvalidFields()}}
\item \href{#method-Get2UsersIdMutingResponse-clone}{\code{Get2UsersIdMutingResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMutingResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMutingResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersIdMutingResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMutingResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMutingResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMutingResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMutingResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdMutingResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMutingResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMutingResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersIdMutingResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMutingResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdMutingResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMutingResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMutingResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMutingResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdMutingResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMutingResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMutingResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersIdMutingResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMutingResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdMutingResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMutingResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMutingResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersIdMutingResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMutingResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMutingResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMutingResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMutingResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersIdMutingResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMutingResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMutingResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMutingResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMutingResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMutingResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMutingResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdMutingResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdMutingResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdMutingResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
