% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_id_owned_lists_response.R
\docType{class}
\name{Get2UsersIdOwnedListsResponse}
\alias{Get2UsersIdOwnedListsResponse}
\title{Get2UsersIdOwnedListsResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersIdOwnedListsResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{List} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ListsIdFollowersResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersIdOwnedListsResponse-new}{\code{Get2UsersIdOwnedListsResponse$new()}}
\item \href{#method-Get2UsersIdOwnedListsResponse-toJSON}{\code{Get2UsersIdOwnedListsResponse$toJSON()}}
\item \href{#method-Get2UsersIdOwnedListsResponse-fromJSON}{\code{Get2UsersIdOwnedListsResponse$fromJSON()}}
\item \href{#method-Get2UsersIdOwnedListsResponse-toJSONString}{\code{Get2UsersIdOwnedListsResponse$toJSONString()}}
\item \href{#method-Get2UsersIdOwnedListsResponse-fromJSONString}{\code{Get2UsersIdOwnedListsResponse$fromJSONString()}}
\item \href{#method-Get2UsersIdOwnedListsResponse-validateJSON}{\code{Get2UsersIdOwnedListsResponse$validateJSON()}}
\item \href{#method-Get2UsersIdOwnedListsResponse-toString}{\code{Get2UsersIdOwnedListsResponse$toString()}}
\item \href{#method-Get2UsersIdOwnedListsResponse-isValid}{\code{Get2UsersIdOwnedListsResponse$isValid()}}
\item \href{#method-Get2UsersIdOwnedListsResponse-getInvalidFields}{\code{Get2UsersIdOwnedListsResponse$getInvalidFields()}}
\item \href{#method-Get2UsersIdOwnedListsResponse-clone}{\code{Get2UsersIdOwnedListsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdOwnedListsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdOwnedListsResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersIdOwnedListsResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdOwnedListsResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdOwnedListsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdOwnedListsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdOwnedListsResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdOwnedListsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdOwnedListsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdOwnedListsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersIdOwnedListsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdOwnedListsResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdOwnedListsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdOwnedListsResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdOwnedListsResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdOwnedListsResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdOwnedListsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdOwnedListsResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdOwnedListsResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersIdOwnedListsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdOwnedListsResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdOwnedListsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdOwnedListsResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdOwnedListsResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersIdOwnedListsResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdOwnedListsResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdOwnedListsResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdOwnedListsResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdOwnedListsResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersIdOwnedListsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdOwnedListsResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdOwnedListsResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdOwnedListsResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdOwnedListsResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdOwnedListsResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdOwnedListsResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdOwnedListsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdOwnedListsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdOwnedListsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
