% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_id_pinned_lists_response.R
\docType{class}
\name{Get2UsersIdPinnedListsResponse}
\alias{Get2UsersIdPinnedListsResponse}
\title{Get2UsersIdPinnedListsResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersIdPinnedListsResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{List} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2ComplianceJobsResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersIdPinnedListsResponse-new}{\code{Get2UsersIdPinnedListsResponse$new()}}
\item \href{#method-Get2UsersIdPinnedListsResponse-toJSON}{\code{Get2UsersIdPinnedListsResponse$toJSON()}}
\item \href{#method-Get2UsersIdPinnedListsResponse-fromJSON}{\code{Get2UsersIdPinnedListsResponse$fromJSON()}}
\item \href{#method-Get2UsersIdPinnedListsResponse-toJSONString}{\code{Get2UsersIdPinnedListsResponse$toJSONString()}}
\item \href{#method-Get2UsersIdPinnedListsResponse-fromJSONString}{\code{Get2UsersIdPinnedListsResponse$fromJSONString()}}
\item \href{#method-Get2UsersIdPinnedListsResponse-validateJSON}{\code{Get2UsersIdPinnedListsResponse$validateJSON()}}
\item \href{#method-Get2UsersIdPinnedListsResponse-toString}{\code{Get2UsersIdPinnedListsResponse$toString()}}
\item \href{#method-Get2UsersIdPinnedListsResponse-isValid}{\code{Get2UsersIdPinnedListsResponse$isValid()}}
\item \href{#method-Get2UsersIdPinnedListsResponse-getInvalidFields}{\code{Get2UsersIdPinnedListsResponse$getInvalidFields()}}
\item \href{#method-Get2UsersIdPinnedListsResponse-clone}{\code{Get2UsersIdPinnedListsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdPinnedListsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdPinnedListsResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersIdPinnedListsResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdPinnedListsResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdPinnedListsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdPinnedListsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdPinnedListsResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdPinnedListsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdPinnedListsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdPinnedListsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersIdPinnedListsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdPinnedListsResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdPinnedListsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdPinnedListsResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdPinnedListsResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdPinnedListsResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdPinnedListsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdPinnedListsResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdPinnedListsResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersIdPinnedListsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdPinnedListsResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdPinnedListsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdPinnedListsResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdPinnedListsResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersIdPinnedListsResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdPinnedListsResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdPinnedListsResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdPinnedListsResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdPinnedListsResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersIdPinnedListsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdPinnedListsResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdPinnedListsResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdPinnedListsResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdPinnedListsResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdPinnedListsResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdPinnedListsResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdPinnedListsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdPinnedListsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdPinnedListsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
