% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get2_users_id_timelines_reverse_chronological_response.R
\docType{class}
\name{Get2UsersIdTimelinesReverseChronologicalResponse}
\alias{Get2UsersIdTimelinesReverseChronologicalResponse}
\title{Get2UsersIdTimelinesReverseChronologicalResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersIdTimelinesReverseChronologicalResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{Tweet} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2UsersIdMentionsResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersIdTimelinesReverseChronologicalResponse-new}{\code{Get2UsersIdTimelinesReverseChronologicalResponse$new()}}
\item \href{#method-Get2UsersIdTimelinesReverseChronologicalResponse-toJSON}{\code{Get2UsersIdTimelinesReverseChronologicalResponse$toJSON()}}
\item \href{#method-Get2UsersIdTimelinesReverseChronologicalResponse-fromJSON}{\code{Get2UsersIdTimelinesReverseChronologicalResponse$fromJSON()}}
\item \href{#method-Get2UsersIdTimelinesReverseChronologicalResponse-toJSONString}{\code{Get2UsersIdTimelinesReverseChronologicalResponse$toJSONString()}}
\item \href{#method-Get2UsersIdTimelinesReverseChronologicalResponse-fromJSONString}{\code{Get2UsersIdTimelinesReverseChronologicalResponse$fromJSONString()}}
\item \href{#method-Get2UsersIdTimelinesReverseChronologicalResponse-validateJSON}{\code{Get2UsersIdTimelinesReverseChronologicalResponse$validateJSON()}}
\item \href{#method-Get2UsersIdTimelinesReverseChronologicalResponse-toString}{\code{Get2UsersIdTimelinesReverseChronologicalResponse$toString()}}
\item \href{#method-Get2UsersIdTimelinesReverseChronologicalResponse-isValid}{\code{Get2UsersIdTimelinesReverseChronologicalResponse$isValid()}}
\item \href{#method-Get2UsersIdTimelinesReverseChronologicalResponse-getInvalidFields}{\code{Get2UsersIdTimelinesReverseChronologicalResponse$getInvalidFields()}}
\item \href{#method-Get2UsersIdTimelinesReverseChronologicalResponse-clone}{\code{Get2UsersIdTimelinesReverseChronologicalResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTimelinesReverseChronologicalResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTimelinesReverseChronologicalResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersIdTimelinesReverseChronologicalResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTimelinesReverseChronologicalResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTimelinesReverseChronologicalResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTimelinesReverseChronologicalResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTimelinesReverseChronologicalResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdTimelinesReverseChronologicalResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTimelinesReverseChronologicalResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTimelinesReverseChronologicalResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersIdTimelinesReverseChronologicalResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTimelinesReverseChronologicalResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdTimelinesReverseChronologicalResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTimelinesReverseChronologicalResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTimelinesReverseChronologicalResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTimelinesReverseChronologicalResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdTimelinesReverseChronologicalResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTimelinesReverseChronologicalResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTimelinesReverseChronologicalResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersIdTimelinesReverseChronologicalResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTimelinesReverseChronologicalResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdTimelinesReverseChronologicalResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTimelinesReverseChronologicalResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTimelinesReverseChronologicalResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersIdTimelinesReverseChronologicalResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTimelinesReverseChronologicalResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTimelinesReverseChronologicalResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTimelinesReverseChronologicalResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTimelinesReverseChronologicalResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersIdTimelinesReverseChronologicalResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTimelinesReverseChronologicalResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTimelinesReverseChronologicalResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTimelinesReverseChronologicalResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTimelinesReverseChronologicalResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTimelinesReverseChronologicalResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTimelinesReverseChronologicalResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTimelinesReverseChronologicalResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTimelinesReverseChronologicalResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTimelinesReverseChronologicalResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
