% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_id_tweets_response.R
\docType{class}
\name{Get2UsersIdTweetsResponse}
\alias{Get2UsersIdTweetsResponse}
\title{Get2UsersIdTweetsResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersIdTweetsResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{Tweet} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}

\item{\code{meta}}{\link{Get2UsersIdMentionsResponseMeta} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersIdTweetsResponse-new}{\code{Get2UsersIdTweetsResponse$new()}}
\item \href{#method-Get2UsersIdTweetsResponse-toJSON}{\code{Get2UsersIdTweetsResponse$toJSON()}}
\item \href{#method-Get2UsersIdTweetsResponse-fromJSON}{\code{Get2UsersIdTweetsResponse$fromJSON()}}
\item \href{#method-Get2UsersIdTweetsResponse-toJSONString}{\code{Get2UsersIdTweetsResponse$toJSONString()}}
\item \href{#method-Get2UsersIdTweetsResponse-fromJSONString}{\code{Get2UsersIdTweetsResponse$fromJSONString()}}
\item \href{#method-Get2UsersIdTweetsResponse-validateJSON}{\code{Get2UsersIdTweetsResponse$validateJSON()}}
\item \href{#method-Get2UsersIdTweetsResponse-toString}{\code{Get2UsersIdTweetsResponse$toString()}}
\item \href{#method-Get2UsersIdTweetsResponse-isValid}{\code{Get2UsersIdTweetsResponse$isValid()}}
\item \href{#method-Get2UsersIdTweetsResponse-getInvalidFields}{\code{Get2UsersIdTweetsResponse$getInvalidFields()}}
\item \href{#method-Get2UsersIdTweetsResponse-clone}{\code{Get2UsersIdTweetsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTweetsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTweetsResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersIdTweetsResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTweetsResponse$new(
  data = NULL,
  errors = NULL,
  includes = NULL,
  meta = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{meta}}{meta}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTweetsResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTweetsResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTweetsResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdTweetsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTweetsResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTweetsResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersIdTweetsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTweetsResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdTweetsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTweetsResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTweetsResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTweetsResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersIdTweetsResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTweetsResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTweetsResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersIdTweetsResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTweetsResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersIdTweetsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTweetsResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTweetsResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersIdTweetsResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTweetsResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTweetsResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTweetsResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTweetsResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersIdTweetsResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTweetsResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTweetsResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTweetsResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTweetsResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTweetsResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTweetsResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersIdTweetsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersIdTweetsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersIdTweetsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
