% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_me_response.R
\docType{class}
\name{Get2UsersMeResponse}
\alias{Get2UsersMeResponse}
\title{Get2UsersMeResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersMeResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{User} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersMeResponse-new}{\code{Get2UsersMeResponse$new()}}
\item \href{#method-Get2UsersMeResponse-toJSON}{\code{Get2UsersMeResponse$toJSON()}}
\item \href{#method-Get2UsersMeResponse-fromJSON}{\code{Get2UsersMeResponse$fromJSON()}}
\item \href{#method-Get2UsersMeResponse-toJSONString}{\code{Get2UsersMeResponse$toJSONString()}}
\item \href{#method-Get2UsersMeResponse-fromJSONString}{\code{Get2UsersMeResponse$fromJSONString()}}
\item \href{#method-Get2UsersMeResponse-validateJSON}{\code{Get2UsersMeResponse$validateJSON()}}
\item \href{#method-Get2UsersMeResponse-toString}{\code{Get2UsersMeResponse$toString()}}
\item \href{#method-Get2UsersMeResponse-isValid}{\code{Get2UsersMeResponse$isValid()}}
\item \href{#method-Get2UsersMeResponse-getInvalidFields}{\code{Get2UsersMeResponse$getInvalidFields()}}
\item \href{#method-Get2UsersMeResponse-clone}{\code{Get2UsersMeResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersMeResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersMeResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersMeResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersMeResponse$new(data = NULL, errors = NULL, includes = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersMeResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersMeResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersMeResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersMeResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersMeResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersMeResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersMeResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersMeResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersMeResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersMeResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersMeResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersMeResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersMeResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersMeResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersMeResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersMeResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersMeResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersMeResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersMeResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersMeResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersMeResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersMeResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersMeResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersMeResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersMeResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersMeResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersMeResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersMeResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersMeResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersMeResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersMeResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersMeResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersMeResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersMeResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersMeResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
