% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2_users_response.R
\docType{class}
\name{Get2UsersResponse}
\alias{Get2UsersResponse}
\title{Get2UsersResponse}
\format{
An \code{R6Class} generator object
}
\description{
Get2UsersResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list( \link{User} ) [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Get2UsersResponse-new}{\code{Get2UsersResponse$new()}}
\item \href{#method-Get2UsersResponse-toJSON}{\code{Get2UsersResponse$toJSON()}}
\item \href{#method-Get2UsersResponse-fromJSON}{\code{Get2UsersResponse$fromJSON()}}
\item \href{#method-Get2UsersResponse-toJSONString}{\code{Get2UsersResponse$toJSONString()}}
\item \href{#method-Get2UsersResponse-fromJSONString}{\code{Get2UsersResponse$fromJSONString()}}
\item \href{#method-Get2UsersResponse-validateJSON}{\code{Get2UsersResponse$validateJSON()}}
\item \href{#method-Get2UsersResponse-toString}{\code{Get2UsersResponse$toString()}}
\item \href{#method-Get2UsersResponse-isValid}{\code{Get2UsersResponse$isValid()}}
\item \href{#method-Get2UsersResponse-getInvalidFields}{\code{Get2UsersResponse$getInvalidFields()}}
\item \href{#method-Get2UsersResponse-clone}{\code{Get2UsersResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Get2UsersResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersResponse$new(data = NULL, errors = NULL, includes = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Get2UsersResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get2UsersResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Get2UsersResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Get2UsersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Get2UsersResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Get2UsersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get2UsersResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Get2UsersResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Get2UsersResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
