% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtag_entity.R
\docType{class}
\name{HashtagEntity}
\alias{HashtagEntity}
\title{HashtagEntity}
\format{
An \code{R6Class} generator object
}
\description{
HashtagEntity Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{end}}{integer}

\item{\code{start}}{integer}

\item{\code{tag}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HashtagEntity-new}{\code{HashtagEntity$new()}}
\item \href{#method-HashtagEntity-toJSON}{\code{HashtagEntity$toJSON()}}
\item \href{#method-HashtagEntity-fromJSON}{\code{HashtagEntity$fromJSON()}}
\item \href{#method-HashtagEntity-toJSONString}{\code{HashtagEntity$toJSONString()}}
\item \href{#method-HashtagEntity-fromJSONString}{\code{HashtagEntity$fromJSONString()}}
\item \href{#method-HashtagEntity-validateJSON}{\code{HashtagEntity$validateJSON()}}
\item \href{#method-HashtagEntity-toString}{\code{HashtagEntity$toString()}}
\item \href{#method-HashtagEntity-isValid}{\code{HashtagEntity$isValid()}}
\item \href{#method-HashtagEntity-getInvalidFields}{\code{HashtagEntity$getInvalidFields()}}
\item \href{#method-HashtagEntity-clone}{\code{HashtagEntity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagEntity-new"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagEntity-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new HashtagEntity class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagEntity$new(end, start, tag, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{end}}{Index (zero-based) at which position this entity ends.  The index is exclusive.}

\item{\code{start}}{Index (zero-based) at which position this entity starts.  The index is inclusive.}

\item{\code{tag}}{The text of the Hashtag.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagEntity-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagEntity-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagEntity$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
HashtagEntity in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagEntity-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagEntity-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of HashtagEntity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagEntity$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of HashtagEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagEntity-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagEntity-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagEntity$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
HashtagEntity in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagEntity-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagEntity-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of HashtagEntity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagEntity$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of HashtagEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagEntity-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagEntity-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to HashtagEntity and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagEntity$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagEntity-toString"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagEntity-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagEntity$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of HashtagEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagEntity-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagEntity-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagEntity$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagEntity-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagEntity-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagEntity$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagEntity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagEntity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagEntity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
