% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtag_fields.R
\docType{class}
\name{HashtagFields}
\alias{HashtagFields}
\title{HashtagFields}
\format{
An \code{R6Class} generator object
}
\description{
HashtagFields Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tag}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HashtagFields-new}{\code{HashtagFields$new()}}
\item \href{#method-HashtagFields-toJSON}{\code{HashtagFields$toJSON()}}
\item \href{#method-HashtagFields-fromJSON}{\code{HashtagFields$fromJSON()}}
\item \href{#method-HashtagFields-toJSONString}{\code{HashtagFields$toJSONString()}}
\item \href{#method-HashtagFields-fromJSONString}{\code{HashtagFields$fromJSONString()}}
\item \href{#method-HashtagFields-validateJSON}{\code{HashtagFields$validateJSON()}}
\item \href{#method-HashtagFields-toString}{\code{HashtagFields$toString()}}
\item \href{#method-HashtagFields-isValid}{\code{HashtagFields$isValid()}}
\item \href{#method-HashtagFields-getInvalidFields}{\code{HashtagFields$getInvalidFields()}}
\item \href{#method-HashtagFields-clone}{\code{HashtagFields$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagFields-new"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagFields-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new HashtagFields class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagFields$new(tag, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag}}{The text of the Hashtag.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagFields-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagFields-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagFields$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
HashtagFields in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagFields-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagFields-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of HashtagFields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagFields$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of HashtagFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagFields-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagFields-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagFields$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
HashtagFields in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagFields-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagFields-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of HashtagFields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagFields$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of HashtagFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagFields-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagFields-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to HashtagFields and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagFields$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagFields-toString"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagFields-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagFields$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of HashtagFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagFields-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagFields-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagFields$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagFields-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagFields-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagFields$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HashtagFields-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HashtagFields-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashtagFields$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
