% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalid_request_problem.R
\docType{class}
\name{InvalidRequestProblem}
\alias{InvalidRequestProblem}
\title{InvalidRequestProblem}
\format{
An \code{R6Class} generator object
}
\description{
InvalidRequestProblem Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Super class}{
\code{\link[twitter:Problem]{twitter::Problem}} -> \code{InvalidRequestProblem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detail}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character}

\item{\code{type}}{character}

\item{\code{errors}}{list( \link{InvalidRequestProblemAllOfErrors} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InvalidRequestProblem-new}{\code{InvalidRequestProblem$new()}}
\item \href{#method-InvalidRequestProblem-toJSON}{\code{InvalidRequestProblem$toJSON()}}
\item \href{#method-InvalidRequestProblem-fromJSON}{\code{InvalidRequestProblem$fromJSON()}}
\item \href{#method-InvalidRequestProblem-toJSONString}{\code{InvalidRequestProblem$toJSONString()}}
\item \href{#method-InvalidRequestProblem-fromJSONString}{\code{InvalidRequestProblem$fromJSONString()}}
\item \href{#method-InvalidRequestProblem-validateJSON}{\code{InvalidRequestProblem$validateJSON()}}
\item \href{#method-InvalidRequestProblem-toString}{\code{InvalidRequestProblem$toString()}}
\item \href{#method-InvalidRequestProblem-isValid}{\code{InvalidRequestProblem$isValid()}}
\item \href{#method-InvalidRequestProblem-getInvalidFields}{\code{InvalidRequestProblem$getInvalidFields()}}
\item \href{#method-InvalidRequestProblem-clone}{\code{InvalidRequestProblem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InvalidRequestProblem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblem$new(
  title,
  type,
  detail = NULL,
  status = NULL,
  errors = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{type}}{type}

\item{\code{detail}}{detail}

\item{\code{status}}{status}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InvalidRequestProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of InvalidRequestProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InvalidRequestProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InvalidRequestProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of InvalidRequestProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InvalidRequestProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to InvalidRequestProblem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of InvalidRequestProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
