% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalid_request_problem_all_of.R
\docType{class}
\name{InvalidRequestProblemAllOf}
\alias{InvalidRequestProblemAllOf}
\title{InvalidRequestProblemAllOf}
\format{
An \code{R6Class} generator object
}
\description{
InvalidRequestProblemAllOf Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{errors}}{list( \link{InvalidRequestProblemAllOfErrors} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InvalidRequestProblemAllOf-new}{\code{InvalidRequestProblemAllOf$new()}}
\item \href{#method-InvalidRequestProblemAllOf-toJSON}{\code{InvalidRequestProblemAllOf$toJSON()}}
\item \href{#method-InvalidRequestProblemAllOf-fromJSON}{\code{InvalidRequestProblemAllOf$fromJSON()}}
\item \href{#method-InvalidRequestProblemAllOf-toJSONString}{\code{InvalidRequestProblemAllOf$toJSONString()}}
\item \href{#method-InvalidRequestProblemAllOf-fromJSONString}{\code{InvalidRequestProblemAllOf$fromJSONString()}}
\item \href{#method-InvalidRequestProblemAllOf-validateJSON}{\code{InvalidRequestProblemAllOf$validateJSON()}}
\item \href{#method-InvalidRequestProblemAllOf-toString}{\code{InvalidRequestProblemAllOf$toString()}}
\item \href{#method-InvalidRequestProblemAllOf-isValid}{\code{InvalidRequestProblemAllOf$isValid()}}
\item \href{#method-InvalidRequestProblemAllOf-getInvalidFields}{\code{InvalidRequestProblemAllOf$getInvalidFields()}}
\item \href{#method-InvalidRequestProblemAllOf-clone}{\code{InvalidRequestProblemAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblemAllOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblemAllOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InvalidRequestProblemAllOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblemAllOf$new(errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblemAllOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblemAllOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblemAllOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InvalidRequestProblemAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblemAllOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblemAllOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of InvalidRequestProblemAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblemAllOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InvalidRequestProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblemAllOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblemAllOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblemAllOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InvalidRequestProblemAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblemAllOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblemAllOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of InvalidRequestProblemAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblemAllOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InvalidRequestProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblemAllOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblemAllOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to InvalidRequestProblemAllOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblemAllOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblemAllOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblemAllOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblemAllOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of InvalidRequestProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblemAllOf-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblemAllOf-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblemAllOf$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblemAllOf-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblemAllOf-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblemAllOf$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRequestProblemAllOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRequestProblemAllOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRequestProblemAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
