% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalid_rule_problem.R
\docType{class}
\name{InvalidRuleProblem}
\alias{InvalidRuleProblem}
\title{InvalidRuleProblem}
\format{
An \code{R6Class} generator object
}
\description{
InvalidRuleProblem Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Super class}{
\code{\link[twitter:Problem]{twitter::Problem}} -> \code{InvalidRuleProblem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detail}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character}

\item{\code{type}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InvalidRuleProblem-new}{\code{InvalidRuleProblem$new()}}
\item \href{#method-InvalidRuleProblem-toJSON}{\code{InvalidRuleProblem$toJSON()}}
\item \href{#method-InvalidRuleProblem-fromJSON}{\code{InvalidRuleProblem$fromJSON()}}
\item \href{#method-InvalidRuleProblem-toJSONString}{\code{InvalidRuleProblem$toJSONString()}}
\item \href{#method-InvalidRuleProblem-fromJSONString}{\code{InvalidRuleProblem$fromJSONString()}}
\item \href{#method-InvalidRuleProblem-validateJSON}{\code{InvalidRuleProblem$validateJSON()}}
\item \href{#method-InvalidRuleProblem-toString}{\code{InvalidRuleProblem$toString()}}
\item \href{#method-InvalidRuleProblem-isValid}{\code{InvalidRuleProblem$isValid()}}
\item \href{#method-InvalidRuleProblem-getInvalidFields}{\code{InvalidRuleProblem$getInvalidFields()}}
\item \href{#method-InvalidRuleProblem-clone}{\code{InvalidRuleProblem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRuleProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRuleProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InvalidRuleProblem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRuleProblem$new(title, type, detail = NULL, status = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{type}}{type}

\item{\code{detail}}{detail}

\item{\code{status}}{status}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRuleProblem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRuleProblem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRuleProblem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InvalidRuleProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRuleProblem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRuleProblem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of InvalidRuleProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRuleProblem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InvalidRuleProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRuleProblem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRuleProblem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRuleProblem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InvalidRuleProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRuleProblem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRuleProblem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of InvalidRuleProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRuleProblem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InvalidRuleProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRuleProblem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRuleProblem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to InvalidRuleProblem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRuleProblem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRuleProblem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRuleProblem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRuleProblem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of InvalidRuleProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRuleProblem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRuleProblem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRuleProblem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRuleProblem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRuleProblem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRuleProblem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvalidRuleProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InvalidRuleProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvalidRuleProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
