% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_create_request.R
\docType{class}
\name{ListCreateRequest}
\alias{ListCreateRequest}
\title{ListCreateRequest}
\format{
An \code{R6Class} generator object
}
\description{
ListCreateRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character [optional]}

\item{\code{name}}{character}

\item{\code{item_private}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListCreateRequest-new}{\code{ListCreateRequest$new()}}
\item \href{#method-ListCreateRequest-toJSON}{\code{ListCreateRequest$toJSON()}}
\item \href{#method-ListCreateRequest-fromJSON}{\code{ListCreateRequest$fromJSON()}}
\item \href{#method-ListCreateRequest-toJSONString}{\code{ListCreateRequest$toJSONString()}}
\item \href{#method-ListCreateRequest-fromJSONString}{\code{ListCreateRequest$fromJSONString()}}
\item \href{#method-ListCreateRequest-validateJSON}{\code{ListCreateRequest$validateJSON()}}
\item \href{#method-ListCreateRequest-toString}{\code{ListCreateRequest$toString()}}
\item \href{#method-ListCreateRequest-isValid}{\code{ListCreateRequest$isValid()}}
\item \href{#method-ListCreateRequest-getInvalidFields}{\code{ListCreateRequest$getInvalidFields()}}
\item \href{#method-ListCreateRequest-clone}{\code{ListCreateRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListCreateRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateRequest$new(name, description = NULL, item_private = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{description}}{description}

\item{\code{item_private}}{item_private. Default to FALSE.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListCreateRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListCreateRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListCreateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListCreateRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListCreateRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListCreateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListCreateRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListCreateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
