% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_create_response.R
\docType{class}
\name{ListCreateResponse}
\alias{ListCreateResponse}
\title{ListCreateResponse}
\format{
An \code{R6Class} generator object
}
\description{
ListCreateResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{ListCreateResponseData} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListCreateResponse-new}{\code{ListCreateResponse$new()}}
\item \href{#method-ListCreateResponse-toJSON}{\code{ListCreateResponse$toJSON()}}
\item \href{#method-ListCreateResponse-fromJSON}{\code{ListCreateResponse$fromJSON()}}
\item \href{#method-ListCreateResponse-toJSONString}{\code{ListCreateResponse$toJSONString()}}
\item \href{#method-ListCreateResponse-fromJSONString}{\code{ListCreateResponse$fromJSONString()}}
\item \href{#method-ListCreateResponse-validateJSON}{\code{ListCreateResponse$validateJSON()}}
\item \href{#method-ListCreateResponse-toString}{\code{ListCreateResponse$toString()}}
\item \href{#method-ListCreateResponse-isValid}{\code{ListCreateResponse$isValid()}}
\item \href{#method-ListCreateResponse-getInvalidFields}{\code{ListCreateResponse$getInvalidFields()}}
\item \href{#method-ListCreateResponse-clone}{\code{ListCreateResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListCreateResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListCreateResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListCreateResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListCreateResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListCreateResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListCreateResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
