% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_create_response_data.R
\docType{class}
\name{ListCreateResponseData}
\alias{ListCreateResponseData}
\title{ListCreateResponseData}
\format{
An \code{R6Class} generator object
}
\description{
ListCreateResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character}

\item{\code{name}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListCreateResponseData-new}{\code{ListCreateResponseData$new()}}
\item \href{#method-ListCreateResponseData-toJSON}{\code{ListCreateResponseData$toJSON()}}
\item \href{#method-ListCreateResponseData-fromJSON}{\code{ListCreateResponseData$fromJSON()}}
\item \href{#method-ListCreateResponseData-toJSONString}{\code{ListCreateResponseData$toJSONString()}}
\item \href{#method-ListCreateResponseData-fromJSONString}{\code{ListCreateResponseData$fromJSONString()}}
\item \href{#method-ListCreateResponseData-validateJSON}{\code{ListCreateResponseData$validateJSON()}}
\item \href{#method-ListCreateResponseData-toString}{\code{ListCreateResponseData$toString()}}
\item \href{#method-ListCreateResponseData-isValid}{\code{ListCreateResponseData$isValid()}}
\item \href{#method-ListCreateResponseData-getInvalidFields}{\code{ListCreateResponseData$getInvalidFields()}}
\item \href{#method-ListCreateResponseData-clone}{\code{ListCreateResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListCreateResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponseData$new(id, name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The unique identifier of this List.}

\item{\code{name}}{The name of this List.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListCreateResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListCreateResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListCreateResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListCreateResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListCreateResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListCreateResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListCreateResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListCreateResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
