% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_delete_response_data.R
\docType{class}
\name{ListDeleteResponseData}
\alias{ListDeleteResponseData}
\title{ListDeleteResponseData}
\format{
An \code{R6Class} generator object
}
\description{
ListDeleteResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deleted}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListDeleteResponseData-new}{\code{ListDeleteResponseData$new()}}
\item \href{#method-ListDeleteResponseData-toJSON}{\code{ListDeleteResponseData$toJSON()}}
\item \href{#method-ListDeleteResponseData-fromJSON}{\code{ListDeleteResponseData$fromJSON()}}
\item \href{#method-ListDeleteResponseData-toJSONString}{\code{ListDeleteResponseData$toJSONString()}}
\item \href{#method-ListDeleteResponseData-fromJSONString}{\code{ListDeleteResponseData$fromJSONString()}}
\item \href{#method-ListDeleteResponseData-validateJSON}{\code{ListDeleteResponseData$validateJSON()}}
\item \href{#method-ListDeleteResponseData-toString}{\code{ListDeleteResponseData$toString()}}
\item \href{#method-ListDeleteResponseData-isValid}{\code{ListDeleteResponseData$isValid()}}
\item \href{#method-ListDeleteResponseData-getInvalidFields}{\code{ListDeleteResponseData$getInvalidFields()}}
\item \href{#method-ListDeleteResponseData-clone}{\code{ListDeleteResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListDeleteResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListDeleteResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListDeleteResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListDeleteResponseData$new(deleted = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deleted}}{deleted}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListDeleteResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListDeleteResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListDeleteResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListDeleteResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListDeleteResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListDeleteResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListDeleteResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListDeleteResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListDeleteResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListDeleteResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListDeleteResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListDeleteResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListDeleteResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListDeleteResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListDeleteResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListDeleteResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListDeleteResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListDeleteResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListDeleteResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListDeleteResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListDeleteResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListDeleteResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListDeleteResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListDeleteResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListDeleteResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListDeleteResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListDeleteResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListDeleteResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListDeleteResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListDeleteResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListDeleteResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListDeleteResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListDeleteResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListDeleteResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListDeleteResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
