% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_followed_request.R
\docType{class}
\name{ListFollowedRequest}
\alias{ListFollowedRequest}
\title{ListFollowedRequest}
\format{
An \code{R6Class} generator object
}
\description{
ListFollowedRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{list_id}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListFollowedRequest-new}{\code{ListFollowedRequest$new()}}
\item \href{#method-ListFollowedRequest-toJSON}{\code{ListFollowedRequest$toJSON()}}
\item \href{#method-ListFollowedRequest-fromJSON}{\code{ListFollowedRequest$fromJSON()}}
\item \href{#method-ListFollowedRequest-toJSONString}{\code{ListFollowedRequest$toJSONString()}}
\item \href{#method-ListFollowedRequest-fromJSONString}{\code{ListFollowedRequest$fromJSONString()}}
\item \href{#method-ListFollowedRequest-validateJSON}{\code{ListFollowedRequest$validateJSON()}}
\item \href{#method-ListFollowedRequest-toString}{\code{ListFollowedRequest$toString()}}
\item \href{#method-ListFollowedRequest-isValid}{\code{ListFollowedRequest$isValid()}}
\item \href{#method-ListFollowedRequest-getInvalidFields}{\code{ListFollowedRequest$getInvalidFields()}}
\item \href{#method-ListFollowedRequest-clone}{\code{ListFollowedRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListFollowedRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedRequest$new(list_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{list_id}}{The unique identifier of this List.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListFollowedRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListFollowedRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListFollowedRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListFollowedRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListFollowedRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListFollowedRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListFollowedRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListFollowedRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
