% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_followed_response_data.R
\docType{class}
\name{ListFollowedResponseData}
\alias{ListFollowedResponseData}
\title{ListFollowedResponseData}
\format{
An \code{R6Class} generator object
}
\description{
ListFollowedResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{following}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListFollowedResponseData-new}{\code{ListFollowedResponseData$new()}}
\item \href{#method-ListFollowedResponseData-toJSON}{\code{ListFollowedResponseData$toJSON()}}
\item \href{#method-ListFollowedResponseData-fromJSON}{\code{ListFollowedResponseData$fromJSON()}}
\item \href{#method-ListFollowedResponseData-toJSONString}{\code{ListFollowedResponseData$toJSONString()}}
\item \href{#method-ListFollowedResponseData-fromJSONString}{\code{ListFollowedResponseData$fromJSONString()}}
\item \href{#method-ListFollowedResponseData-validateJSON}{\code{ListFollowedResponseData$validateJSON()}}
\item \href{#method-ListFollowedResponseData-toString}{\code{ListFollowedResponseData$toString()}}
\item \href{#method-ListFollowedResponseData-isValid}{\code{ListFollowedResponseData$isValid()}}
\item \href{#method-ListFollowedResponseData-getInvalidFields}{\code{ListFollowedResponseData$getInvalidFields()}}
\item \href{#method-ListFollowedResponseData-clone}{\code{ListFollowedResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListFollowedResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedResponseData$new(following = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{following}}{following}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListFollowedResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListFollowedResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListFollowedResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListFollowedResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListFollowedResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListFollowedResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListFollowedResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListFollowedResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFollowedResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListFollowedResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListFollowedResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
