% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_mutate_response.R
\docType{class}
\name{ListMutateResponse}
\alias{ListMutateResponse}
\title{ListMutateResponse}
\format{
An \code{R6Class} generator object
}
\description{
ListMutateResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{ListMutateResponseData} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListMutateResponse-new}{\code{ListMutateResponse$new()}}
\item \href{#method-ListMutateResponse-toJSON}{\code{ListMutateResponse$toJSON()}}
\item \href{#method-ListMutateResponse-fromJSON}{\code{ListMutateResponse$fromJSON()}}
\item \href{#method-ListMutateResponse-toJSONString}{\code{ListMutateResponse$toJSONString()}}
\item \href{#method-ListMutateResponse-fromJSONString}{\code{ListMutateResponse$fromJSONString()}}
\item \href{#method-ListMutateResponse-validateJSON}{\code{ListMutateResponse$validateJSON()}}
\item \href{#method-ListMutateResponse-toString}{\code{ListMutateResponse$toString()}}
\item \href{#method-ListMutateResponse-isValid}{\code{ListMutateResponse$isValid()}}
\item \href{#method-ListMutateResponse-getInvalidFields}{\code{ListMutateResponse$getInvalidFields()}}
\item \href{#method-ListMutateResponse-clone}{\code{ListMutateResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListMutateResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListMutateResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListMutateResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListMutateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListMutateResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListMutateResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListMutateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListMutateResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListMutateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
