% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_mutate_response_data.R
\docType{class}
\name{ListMutateResponseData}
\alias{ListMutateResponseData}
\title{ListMutateResponseData}
\format{
An \code{R6Class} generator object
}
\description{
ListMutateResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{is_member}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListMutateResponseData-new}{\code{ListMutateResponseData$new()}}
\item \href{#method-ListMutateResponseData-toJSON}{\code{ListMutateResponseData$toJSON()}}
\item \href{#method-ListMutateResponseData-fromJSON}{\code{ListMutateResponseData$fromJSON()}}
\item \href{#method-ListMutateResponseData-toJSONString}{\code{ListMutateResponseData$toJSONString()}}
\item \href{#method-ListMutateResponseData-fromJSONString}{\code{ListMutateResponseData$fromJSONString()}}
\item \href{#method-ListMutateResponseData-validateJSON}{\code{ListMutateResponseData$validateJSON()}}
\item \href{#method-ListMutateResponseData-toString}{\code{ListMutateResponseData$toString()}}
\item \href{#method-ListMutateResponseData-isValid}{\code{ListMutateResponseData$isValid()}}
\item \href{#method-ListMutateResponseData-getInvalidFields}{\code{ListMutateResponseData$getInvalidFields()}}
\item \href{#method-ListMutateResponseData-clone}{\code{ListMutateResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListMutateResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponseData$new(is_member = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{is_member}}{is_member}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListMutateResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListMutateResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListMutateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListMutateResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListMutateResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListMutateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListMutateResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListMutateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMutateResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListMutateResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListMutateResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
