% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pinned_request.R
\docType{class}
\name{ListPinnedRequest}
\alias{ListPinnedRequest}
\title{ListPinnedRequest}
\format{
An \code{R6Class} generator object
}
\description{
ListPinnedRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{list_id}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListPinnedRequest-new}{\code{ListPinnedRequest$new()}}
\item \href{#method-ListPinnedRequest-toJSON}{\code{ListPinnedRequest$toJSON()}}
\item \href{#method-ListPinnedRequest-fromJSON}{\code{ListPinnedRequest$fromJSON()}}
\item \href{#method-ListPinnedRequest-toJSONString}{\code{ListPinnedRequest$toJSONString()}}
\item \href{#method-ListPinnedRequest-fromJSONString}{\code{ListPinnedRequest$fromJSONString()}}
\item \href{#method-ListPinnedRequest-validateJSON}{\code{ListPinnedRequest$validateJSON()}}
\item \href{#method-ListPinnedRequest-toString}{\code{ListPinnedRequest$toString()}}
\item \href{#method-ListPinnedRequest-isValid}{\code{ListPinnedRequest$isValid()}}
\item \href{#method-ListPinnedRequest-getInvalidFields}{\code{ListPinnedRequest$getInvalidFields()}}
\item \href{#method-ListPinnedRequest-clone}{\code{ListPinnedRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListPinnedRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedRequest$new(list_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{list_id}}{The unique identifier of this List.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListPinnedRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListPinnedRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListPinnedRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListPinnedRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListPinnedRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListPinnedRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListPinnedRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListPinnedRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
