% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pinned_response.R
\docType{class}
\name{ListPinnedResponse}
\alias{ListPinnedResponse}
\title{ListPinnedResponse}
\format{
An \code{R6Class} generator object
}
\description{
ListPinnedResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{ListPinnedResponseData} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListPinnedResponse-new}{\code{ListPinnedResponse$new()}}
\item \href{#method-ListPinnedResponse-toJSON}{\code{ListPinnedResponse$toJSON()}}
\item \href{#method-ListPinnedResponse-fromJSON}{\code{ListPinnedResponse$fromJSON()}}
\item \href{#method-ListPinnedResponse-toJSONString}{\code{ListPinnedResponse$toJSONString()}}
\item \href{#method-ListPinnedResponse-fromJSONString}{\code{ListPinnedResponse$fromJSONString()}}
\item \href{#method-ListPinnedResponse-validateJSON}{\code{ListPinnedResponse$validateJSON()}}
\item \href{#method-ListPinnedResponse-toString}{\code{ListPinnedResponse$toString()}}
\item \href{#method-ListPinnedResponse-isValid}{\code{ListPinnedResponse$isValid()}}
\item \href{#method-ListPinnedResponse-getInvalidFields}{\code{ListPinnedResponse$getInvalidFields()}}
\item \href{#method-ListPinnedResponse-clone}{\code{ListPinnedResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListPinnedResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListPinnedResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListPinnedResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListPinnedResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListPinnedResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListPinnedResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListPinnedResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListPinnedResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListPinnedResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
