% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pinned_response_data.R
\docType{class}
\name{ListPinnedResponseData}
\alias{ListPinnedResponseData}
\title{ListPinnedResponseData}
\format{
An \code{R6Class} generator object
}
\description{
ListPinnedResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pinned}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListPinnedResponseData-new}{\code{ListPinnedResponseData$new()}}
\item \href{#method-ListPinnedResponseData-toJSON}{\code{ListPinnedResponseData$toJSON()}}
\item \href{#method-ListPinnedResponseData-fromJSON}{\code{ListPinnedResponseData$fromJSON()}}
\item \href{#method-ListPinnedResponseData-toJSONString}{\code{ListPinnedResponseData$toJSONString()}}
\item \href{#method-ListPinnedResponseData-fromJSONString}{\code{ListPinnedResponseData$fromJSONString()}}
\item \href{#method-ListPinnedResponseData-validateJSON}{\code{ListPinnedResponseData$validateJSON()}}
\item \href{#method-ListPinnedResponseData-toString}{\code{ListPinnedResponseData$toString()}}
\item \href{#method-ListPinnedResponseData-isValid}{\code{ListPinnedResponseData$isValid()}}
\item \href{#method-ListPinnedResponseData-getInvalidFields}{\code{ListPinnedResponseData$getInvalidFields()}}
\item \href{#method-ListPinnedResponseData-clone}{\code{ListPinnedResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListPinnedResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponseData$new(pinned = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pinned}}{pinned}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListPinnedResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListPinnedResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListPinnedResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListPinnedResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListPinnedResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListPinnedResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListPinnedResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListPinnedResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListPinnedResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListPinnedResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListPinnedResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
