% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_update_request.R
\docType{class}
\name{ListUpdateRequest}
\alias{ListUpdateRequest}
\title{ListUpdateRequest}
\format{
An \code{R6Class} generator object
}
\description{
ListUpdateRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{item_private}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListUpdateRequest-new}{\code{ListUpdateRequest$new()}}
\item \href{#method-ListUpdateRequest-toJSON}{\code{ListUpdateRequest$toJSON()}}
\item \href{#method-ListUpdateRequest-fromJSON}{\code{ListUpdateRequest$fromJSON()}}
\item \href{#method-ListUpdateRequest-toJSONString}{\code{ListUpdateRequest$toJSONString()}}
\item \href{#method-ListUpdateRequest-fromJSONString}{\code{ListUpdateRequest$fromJSONString()}}
\item \href{#method-ListUpdateRequest-validateJSON}{\code{ListUpdateRequest$validateJSON()}}
\item \href{#method-ListUpdateRequest-toString}{\code{ListUpdateRequest$toString()}}
\item \href{#method-ListUpdateRequest-isValid}{\code{ListUpdateRequest$isValid()}}
\item \href{#method-ListUpdateRequest-getInvalidFields}{\code{ListUpdateRequest$getInvalidFields()}}
\item \href{#method-ListUpdateRequest-clone}{\code{ListUpdateRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListUpdateRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateRequest$new(
  description = NULL,
  name = NULL,
  item_private = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{description}

\item{\code{name}}{name}

\item{\code{item_private}}{item_private}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListUpdateRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListUpdateRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListUpdateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListUpdateRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListUpdateRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListUpdateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListUpdateRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListUpdateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
