% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_update_response.R
\docType{class}
\name{ListUpdateResponse}
\alias{ListUpdateResponse}
\title{ListUpdateResponse}
\format{
An \code{R6Class} generator object
}
\description{
ListUpdateResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{ListUpdateResponseData} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListUpdateResponse-new}{\code{ListUpdateResponse$new()}}
\item \href{#method-ListUpdateResponse-toJSON}{\code{ListUpdateResponse$toJSON()}}
\item \href{#method-ListUpdateResponse-fromJSON}{\code{ListUpdateResponse$fromJSON()}}
\item \href{#method-ListUpdateResponse-toJSONString}{\code{ListUpdateResponse$toJSONString()}}
\item \href{#method-ListUpdateResponse-fromJSONString}{\code{ListUpdateResponse$fromJSONString()}}
\item \href{#method-ListUpdateResponse-validateJSON}{\code{ListUpdateResponse$validateJSON()}}
\item \href{#method-ListUpdateResponse-toString}{\code{ListUpdateResponse$toString()}}
\item \href{#method-ListUpdateResponse-isValid}{\code{ListUpdateResponse$isValid()}}
\item \href{#method-ListUpdateResponse-getInvalidFields}{\code{ListUpdateResponse$getInvalidFields()}}
\item \href{#method-ListUpdateResponse-clone}{\code{ListUpdateResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListUpdateResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListUpdateResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListUpdateResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListUpdateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListUpdateResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListUpdateResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListUpdateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListUpdateResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListUpdateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
