% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_update_response_data.R
\docType{class}
\name{ListUpdateResponseData}
\alias{ListUpdateResponseData}
\title{ListUpdateResponseData}
\format{
An \code{R6Class} generator object
}
\description{
ListUpdateResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{updated}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListUpdateResponseData-new}{\code{ListUpdateResponseData$new()}}
\item \href{#method-ListUpdateResponseData-toJSON}{\code{ListUpdateResponseData$toJSON()}}
\item \href{#method-ListUpdateResponseData-fromJSON}{\code{ListUpdateResponseData$fromJSON()}}
\item \href{#method-ListUpdateResponseData-toJSONString}{\code{ListUpdateResponseData$toJSONString()}}
\item \href{#method-ListUpdateResponseData-fromJSONString}{\code{ListUpdateResponseData$fromJSONString()}}
\item \href{#method-ListUpdateResponseData-validateJSON}{\code{ListUpdateResponseData$validateJSON()}}
\item \href{#method-ListUpdateResponseData-toString}{\code{ListUpdateResponseData$toString()}}
\item \href{#method-ListUpdateResponseData-isValid}{\code{ListUpdateResponseData$isValid()}}
\item \href{#method-ListUpdateResponseData-getInvalidFields}{\code{ListUpdateResponseData$getInvalidFields()}}
\item \href{#method-ListUpdateResponseData-clone}{\code{ListUpdateResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ListUpdateResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponseData$new(updated = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{updated}}{updated}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListUpdateResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ListUpdateResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListUpdateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ListUpdateResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ListUpdateResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ListUpdateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ListUpdateResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ListUpdateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUpdateResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListUpdateResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ListUpdateResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
