% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mention_fields.R
\docType{class}
\name{MentionFields}
\alias{MentionFields}
\title{MentionFields}
\format{
An \code{R6Class} generator object
}
\description{
MentionFields Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{username}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MentionFields-new}{\code{MentionFields$new()}}
\item \href{#method-MentionFields-toJSON}{\code{MentionFields$toJSON()}}
\item \href{#method-MentionFields-fromJSON}{\code{MentionFields$fromJSON()}}
\item \href{#method-MentionFields-toJSONString}{\code{MentionFields$toJSONString()}}
\item \href{#method-MentionFields-fromJSONString}{\code{MentionFields$fromJSONString()}}
\item \href{#method-MentionFields-validateJSON}{\code{MentionFields$validateJSON()}}
\item \href{#method-MentionFields-toString}{\code{MentionFields$toString()}}
\item \href{#method-MentionFields-isValid}{\code{MentionFields$isValid()}}
\item \href{#method-MentionFields-getInvalidFields}{\code{MentionFields$getInvalidFields()}}
\item \href{#method-MentionFields-clone}{\code{MentionFields$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MentionFields-new"></a>}}
\if{latex}{\out{\hypertarget{method-MentionFields-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new MentionFields class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MentionFields$new(username, id = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{The Twitter handle (screen name) of this user.}

\item{\code{id}}{Unique identifier of this User. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MentionFields-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MentionFields-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MentionFields$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MentionFields in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MentionFields-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MentionFields-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of MentionFields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MentionFields$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MentionFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MentionFields-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MentionFields-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MentionFields$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MentionFields in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MentionFields-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MentionFields-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of MentionFields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MentionFields$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MentionFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MentionFields-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MentionFields-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to MentionFields and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MentionFields$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MentionFields-toString"></a>}}
\if{latex}{\out{\hypertarget{method-MentionFields-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MentionFields$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of MentionFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MentionFields-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-MentionFields-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MentionFields$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MentionFields-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-MentionFields-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MentionFields$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MentionFields-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MentionFields-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MentionFields$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
