% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mute_user_mutation_response_data.R
\docType{class}
\name{MuteUserMutationResponseData}
\alias{MuteUserMutationResponseData}
\title{MuteUserMutationResponseData}
\format{
An \code{R6Class} generator object
}
\description{
MuteUserMutationResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{muting}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MuteUserMutationResponseData-new}{\code{MuteUserMutationResponseData$new()}}
\item \href{#method-MuteUserMutationResponseData-toJSON}{\code{MuteUserMutationResponseData$toJSON()}}
\item \href{#method-MuteUserMutationResponseData-fromJSON}{\code{MuteUserMutationResponseData$fromJSON()}}
\item \href{#method-MuteUserMutationResponseData-toJSONString}{\code{MuteUserMutationResponseData$toJSONString()}}
\item \href{#method-MuteUserMutationResponseData-fromJSONString}{\code{MuteUserMutationResponseData$fromJSONString()}}
\item \href{#method-MuteUserMutationResponseData-validateJSON}{\code{MuteUserMutationResponseData$validateJSON()}}
\item \href{#method-MuteUserMutationResponseData-toString}{\code{MuteUserMutationResponseData$toString()}}
\item \href{#method-MuteUserMutationResponseData-isValid}{\code{MuteUserMutationResponseData$isValid()}}
\item \href{#method-MuteUserMutationResponseData-getInvalidFields}{\code{MuteUserMutationResponseData$getInvalidFields()}}
\item \href{#method-MuteUserMutationResponseData-clone}{\code{MuteUserMutationResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserMutationResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserMutationResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new MuteUserMutationResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserMutationResponseData$new(muting = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{muting}}{muting}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserMutationResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserMutationResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserMutationResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MuteUserMutationResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserMutationResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserMutationResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of MuteUserMutationResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserMutationResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MuteUserMutationResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserMutationResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserMutationResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserMutationResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MuteUserMutationResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserMutationResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserMutationResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of MuteUserMutationResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserMutationResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MuteUserMutationResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserMutationResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserMutationResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to MuteUserMutationResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserMutationResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserMutationResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserMutationResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserMutationResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of MuteUserMutationResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserMutationResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserMutationResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserMutationResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserMutationResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserMutationResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserMutationResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserMutationResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserMutationResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserMutationResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
