% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mute_user_request.R
\docType{class}
\name{MuteUserRequest}
\alias{MuteUserRequest}
\title{MuteUserRequest}
\format{
An \code{R6Class} generator object
}
\description{
MuteUserRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{target_user_id}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MuteUserRequest-new}{\code{MuteUserRequest$new()}}
\item \href{#method-MuteUserRequest-toJSON}{\code{MuteUserRequest$toJSON()}}
\item \href{#method-MuteUserRequest-fromJSON}{\code{MuteUserRequest$fromJSON()}}
\item \href{#method-MuteUserRequest-toJSONString}{\code{MuteUserRequest$toJSONString()}}
\item \href{#method-MuteUserRequest-fromJSONString}{\code{MuteUserRequest$fromJSONString()}}
\item \href{#method-MuteUserRequest-validateJSON}{\code{MuteUserRequest$validateJSON()}}
\item \href{#method-MuteUserRequest-toString}{\code{MuteUserRequest$toString()}}
\item \href{#method-MuteUserRequest-isValid}{\code{MuteUserRequest$isValid()}}
\item \href{#method-MuteUserRequest-getInvalidFields}{\code{MuteUserRequest$getInvalidFields()}}
\item \href{#method-MuteUserRequest-clone}{\code{MuteUserRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new MuteUserRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserRequest$new(target_user_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_user_id}}{Unique identifier of this User. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MuteUserRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of MuteUserRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MuteUserRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MuteUserRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of MuteUserRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MuteUserRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to MuteUserRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of MuteUserRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MuteUserRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MuteUserRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MuteUserRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
