% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_disconnect_problem.R
\docType{class}
\name{OperationalDisconnectProblem}
\alias{OperationalDisconnectProblem}
\title{OperationalDisconnectProblem}
\format{
An \code{R6Class} generator object
}
\description{
OperationalDisconnectProblem Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Super class}{
\code{\link[twitter:Problem]{twitter::Problem}} -> \code{OperationalDisconnectProblem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detail}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character}

\item{\code{type}}{character}

\item{\code{disconnect_type}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OperationalDisconnectProblem-new}{\code{OperationalDisconnectProblem$new()}}
\item \href{#method-OperationalDisconnectProblem-toJSON}{\code{OperationalDisconnectProblem$toJSON()}}
\item \href{#method-OperationalDisconnectProblem-fromJSON}{\code{OperationalDisconnectProblem$fromJSON()}}
\item \href{#method-OperationalDisconnectProblem-toJSONString}{\code{OperationalDisconnectProblem$toJSONString()}}
\item \href{#method-OperationalDisconnectProblem-fromJSONString}{\code{OperationalDisconnectProblem$fromJSONString()}}
\item \href{#method-OperationalDisconnectProblem-validateJSON}{\code{OperationalDisconnectProblem$validateJSON()}}
\item \href{#method-OperationalDisconnectProblem-toString}{\code{OperationalDisconnectProblem$toString()}}
\item \href{#method-OperationalDisconnectProblem-isValid}{\code{OperationalDisconnectProblem$isValid()}}
\item \href{#method-OperationalDisconnectProblem-getInvalidFields}{\code{OperationalDisconnectProblem$getInvalidFields()}}
\item \href{#method-OperationalDisconnectProblem-clone}{\code{OperationalDisconnectProblem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OperationalDisconnectProblem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblem$new(
  title,
  type,
  detail = NULL,
  status = NULL,
  disconnect_type = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{type}}{type}

\item{\code{detail}}{detail}

\item{\code{status}}{status}

\item{\code{disconnect_type}}{disconnect_type}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OperationalDisconnectProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OperationalDisconnectProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OperationalDisconnectProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OperationalDisconnectProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OperationalDisconnectProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OperationalDisconnectProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OperationalDisconnectProblem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OperationalDisconnectProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
