% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_disconnect_problem_all_of.R
\docType{class}
\name{OperationalDisconnectProblemAllOf}
\alias{OperationalDisconnectProblemAllOf}
\title{OperationalDisconnectProblemAllOf}
\format{
An \code{R6Class} generator object
}
\description{
OperationalDisconnectProblemAllOf Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{disconnect_type}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OperationalDisconnectProblemAllOf-new}{\code{OperationalDisconnectProblemAllOf$new()}}
\item \href{#method-OperationalDisconnectProblemAllOf-toJSON}{\code{OperationalDisconnectProblemAllOf$toJSON()}}
\item \href{#method-OperationalDisconnectProblemAllOf-fromJSON}{\code{OperationalDisconnectProblemAllOf$fromJSON()}}
\item \href{#method-OperationalDisconnectProblemAllOf-toJSONString}{\code{OperationalDisconnectProblemAllOf$toJSONString()}}
\item \href{#method-OperationalDisconnectProblemAllOf-fromJSONString}{\code{OperationalDisconnectProblemAllOf$fromJSONString()}}
\item \href{#method-OperationalDisconnectProblemAllOf-validateJSON}{\code{OperationalDisconnectProblemAllOf$validateJSON()}}
\item \href{#method-OperationalDisconnectProblemAllOf-toString}{\code{OperationalDisconnectProblemAllOf$toString()}}
\item \href{#method-OperationalDisconnectProblemAllOf-isValid}{\code{OperationalDisconnectProblemAllOf$isValid()}}
\item \href{#method-OperationalDisconnectProblemAllOf-getInvalidFields}{\code{OperationalDisconnectProblemAllOf$getInvalidFields()}}
\item \href{#method-OperationalDisconnectProblemAllOf-clone}{\code{OperationalDisconnectProblemAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblemAllOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblemAllOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OperationalDisconnectProblemAllOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblemAllOf$new(disconnect_type = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{disconnect_type}}{disconnect_type}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblemAllOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblemAllOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblemAllOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OperationalDisconnectProblemAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblemAllOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblemAllOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OperationalDisconnectProblemAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblemAllOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OperationalDisconnectProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblemAllOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblemAllOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblemAllOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
OperationalDisconnectProblemAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblemAllOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblemAllOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OperationalDisconnectProblemAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblemAllOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of OperationalDisconnectProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblemAllOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblemAllOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to OperationalDisconnectProblemAllOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblemAllOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblemAllOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblemAllOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblemAllOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of OperationalDisconnectProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblemAllOf-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblemAllOf-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblemAllOf$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblemAllOf-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblemAllOf-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblemAllOf$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationalDisconnectProblemAllOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OperationalDisconnectProblemAllOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationalDisconnectProblemAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
