% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_type.R
\docType{class}
\name{PlaceType}
\alias{PlaceType}
\title{PlaceType}
\format{
An \code{R6Class} generator object
}
\description{
PlaceType Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PlaceType-new}{\code{PlaceType$new()}}
\item \href{#method-PlaceType-toJSON}{\code{PlaceType$toJSON()}}
\item \href{#method-PlaceType-fromJSON}{\code{PlaceType$fromJSON()}}
\item \href{#method-PlaceType-toJSONString}{\code{PlaceType$toJSONString()}}
\item \href{#method-PlaceType-fromJSONString}{\code{PlaceType$fromJSONString()}}
\item \href{#method-PlaceType-clone}{\code{PlaceType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlaceType-new"></a>}}
\if{latex}{\out{\hypertarget{method-PlaceType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PlaceType class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaceType$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlaceType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PlaceType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaceType$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PlaceType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlaceType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PlaceType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PlaceType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaceType$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PlaceType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlaceType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PlaceType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaceType$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PlaceType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlaceType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PlaceType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PlaceType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaceType$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PlaceType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlaceType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlaceType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaceType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
