% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_errors.R
\docType{class}
\name{ProblemErrors}
\alias{ProblemErrors}
\title{ProblemErrors}
\format{
An \code{R6Class} generator object
}
\description{
ProblemErrors Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{errors}}{list( \link{Problem} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProblemErrors-new}{\code{ProblemErrors$new()}}
\item \href{#method-ProblemErrors-toJSON}{\code{ProblemErrors$toJSON()}}
\item \href{#method-ProblemErrors-fromJSON}{\code{ProblemErrors$fromJSON()}}
\item \href{#method-ProblemErrors-toJSONString}{\code{ProblemErrors$toJSONString()}}
\item \href{#method-ProblemErrors-fromJSONString}{\code{ProblemErrors$fromJSONString()}}
\item \href{#method-ProblemErrors-validateJSON}{\code{ProblemErrors$validateJSON()}}
\item \href{#method-ProblemErrors-toString}{\code{ProblemErrors$toString()}}
\item \href{#method-ProblemErrors-isValid}{\code{ProblemErrors$isValid()}}
\item \href{#method-ProblemErrors-getInvalidFields}{\code{ProblemErrors$getInvalidFields()}}
\item \href{#method-ProblemErrors-clone}{\code{ProblemErrors$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemErrors-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemErrors-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ProblemErrors class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemErrors$new(errors, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemErrors-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemErrors-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemErrors$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ProblemErrors in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemErrors-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemErrors-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ProblemErrors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemErrors$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ProblemErrors
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemErrors-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemErrors-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemErrors$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ProblemErrors in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemErrors-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemErrors-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ProblemErrors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemErrors$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ProblemErrors
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemErrors-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemErrors-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ProblemErrors and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemErrors$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemErrors-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemErrors-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemErrors$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ProblemErrors
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemErrors-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemErrors-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemErrors$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemErrors-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemErrors-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemErrors$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemErrors-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemErrors-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemErrors$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
