% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_or_error.R
\docType{class}
\name{ProblemOrError}
\alias{ProblemOrError}
\title{ProblemOrError}
\format{
An \code{R6Class} generator object
}
\description{
ProblemOrError Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of object types defined in the oneOf schema.
Initialize a new ProblemOrError.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProblemOrError-new}{\code{ProblemOrError$new()}}
\item \href{#method-ProblemOrError-fromJSON}{\code{ProblemOrError$fromJSON()}}
\item \href{#method-ProblemOrError-toJSONString}{\code{ProblemOrError$toJSONString()}}
\item \href{#method-ProblemOrError-toJSON}{\code{ProblemOrError$toJSON()}}
\item \href{#method-ProblemOrError-validateJSON}{\code{ProblemOrError$validateJSON()}}
\item \href{#method-ProblemOrError-toString}{\code{ProblemOrError$toString()}}
\item \href{#method-ProblemOrError-clone}{\code{ProblemOrError$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemOrError-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemOrError-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ProblemOrError.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemOrError$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "Error", "Problem"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemOrError-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemOrError-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ProblemOrError.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemOrError$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of ProblemOrError.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemOrError-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemOrError-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize ProblemOrError to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemOrError$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the ProblemOrError.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemOrError-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemOrError-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize ProblemOrError to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemOrError$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the ProblemOrError.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemOrError-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemOrError-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to ProblemOrError and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemOrError$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemOrError-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemOrError-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemOrError$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProblemOrError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProblemOrError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProblemOrError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
