% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reply_settings.R
\docType{class}
\name{ReplySettings}
\alias{ReplySettings}
\title{ReplySettings}
\format{
An \code{R6Class} generator object
}
\description{
ReplySettings Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReplySettings-new}{\code{ReplySettings$new()}}
\item \href{#method-ReplySettings-toJSON}{\code{ReplySettings$toJSON()}}
\item \href{#method-ReplySettings-fromJSON}{\code{ReplySettings$fromJSON()}}
\item \href{#method-ReplySettings-toJSONString}{\code{ReplySettings$toJSONString()}}
\item \href{#method-ReplySettings-fromJSONString}{\code{ReplySettings$fromJSONString()}}
\item \href{#method-ReplySettings-clone}{\code{ReplySettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplySettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReplySettings-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ReplySettings class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplySettings$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplySettings-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ReplySettings-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplySettings$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ReplySettings in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplySettings-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ReplySettings-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ReplySettings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplySettings$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ReplySettings
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplySettings-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ReplySettings-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplySettings$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ReplySettings in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplySettings-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ReplySettings-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ReplySettings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplySettings$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ReplySettings
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplySettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReplySettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplySettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
