% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_users_response.R
\docType{class}
\name{ReportUsersResponse}
\alias{ReportUsersResponse}
\title{ReportUsersResponse}
\format{
An \code{R6Class} generator object
}
\description{
ReportUsersResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{ReportUsersResponseData} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReportUsersResponse-new}{\code{ReportUsersResponse$new()}}
\item \href{#method-ReportUsersResponse-toJSON}{\code{ReportUsersResponse$toJSON()}}
\item \href{#method-ReportUsersResponse-fromJSON}{\code{ReportUsersResponse$fromJSON()}}
\item \href{#method-ReportUsersResponse-toJSONString}{\code{ReportUsersResponse$toJSONString()}}
\item \href{#method-ReportUsersResponse-fromJSONString}{\code{ReportUsersResponse$fromJSONString()}}
\item \href{#method-ReportUsersResponse-validateJSON}{\code{ReportUsersResponse$validateJSON()}}
\item \href{#method-ReportUsersResponse-toString}{\code{ReportUsersResponse$toString()}}
\item \href{#method-ReportUsersResponse-isValid}{\code{ReportUsersResponse$isValid()}}
\item \href{#method-ReportUsersResponse-getInvalidFields}{\code{ReportUsersResponse$getInvalidFields()}}
\item \href{#method-ReportUsersResponse-clone}{\code{ReportUsersResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ReportUsersResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponse$new(data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ReportUsersResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ReportUsersResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ReportUsersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ReportUsersResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ReportUsersResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ReportUsersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ReportUsersResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ReportUsersResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
