% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_users_response_data.R
\docType{class}
\name{ReportUsersResponseData}
\alias{ReportUsersResponseData}
\title{ReportUsersResponseData}
\format{
An \code{R6Class} generator object
}
\description{
ReportUsersResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReportUsersResponseData-new}{\code{ReportUsersResponseData$new()}}
\item \href{#method-ReportUsersResponseData-toJSON}{\code{ReportUsersResponseData$toJSON()}}
\item \href{#method-ReportUsersResponseData-fromJSON}{\code{ReportUsersResponseData$fromJSON()}}
\item \href{#method-ReportUsersResponseData-toJSONString}{\code{ReportUsersResponseData$toJSONString()}}
\item \href{#method-ReportUsersResponseData-fromJSONString}{\code{ReportUsersResponseData$fromJSONString()}}
\item \href{#method-ReportUsersResponseData-validateJSON}{\code{ReportUsersResponseData$validateJSON()}}
\item \href{#method-ReportUsersResponseData-toString}{\code{ReportUsersResponseData$toString()}}
\item \href{#method-ReportUsersResponseData-isValid}{\code{ReportUsersResponseData$isValid()}}
\item \href{#method-ReportUsersResponseData-getInvalidFields}{\code{ReportUsersResponseData$getInvalidFields()}}
\item \href{#method-ReportUsersResponseData-clone}{\code{ReportUsersResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ReportUsersResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponseData$new(id = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ReportUsersResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ReportUsersResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ReportUsersResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ReportUsersResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ReportUsersResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ReportUsersResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ReportUsersResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ReportUsersResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReportUsersResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReportUsersResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportUsersResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
