% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_not_found_problem.R
\docType{class}
\name{ResourceNotFoundProblem}
\alias{ResourceNotFoundProblem}
\title{ResourceNotFoundProblem}
\format{
An \code{R6Class} generator object
}
\description{
ResourceNotFoundProblem Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Super class}{
\code{\link[twitter:Problem]{twitter::Problem}} -> \code{ResourceNotFoundProblem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detail}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character}

\item{\code{type}}{character}

\item{\code{parameter}}{character}

\item{\code{resource_id}}{character}

\item{\code{resource_type}}{character}

\item{\code{value}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResourceNotFoundProblem-new}{\code{ResourceNotFoundProblem$new()}}
\item \href{#method-ResourceNotFoundProblem-toJSON}{\code{ResourceNotFoundProblem$toJSON()}}
\item \href{#method-ResourceNotFoundProblem-fromJSON}{\code{ResourceNotFoundProblem$fromJSON()}}
\item \href{#method-ResourceNotFoundProblem-toJSONString}{\code{ResourceNotFoundProblem$toJSONString()}}
\item \href{#method-ResourceNotFoundProblem-fromJSONString}{\code{ResourceNotFoundProblem$fromJSONString()}}
\item \href{#method-ResourceNotFoundProblem-validateJSON}{\code{ResourceNotFoundProblem$validateJSON()}}
\item \href{#method-ResourceNotFoundProblem-toString}{\code{ResourceNotFoundProblem$toString()}}
\item \href{#method-ResourceNotFoundProblem-isValid}{\code{ResourceNotFoundProblem$isValid()}}
\item \href{#method-ResourceNotFoundProblem-getInvalidFields}{\code{ResourceNotFoundProblem$getInvalidFields()}}
\item \href{#method-ResourceNotFoundProblem-clone}{\code{ResourceNotFoundProblem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceNotFoundProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceNotFoundProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ResourceNotFoundProblem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceNotFoundProblem$new(
  title,
  type,
  parameter,
  resource_id,
  resource_type,
  value,
  detail = NULL,
  status = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{type}}{type}

\item{\code{parameter}}{parameter}

\item{\code{resource_id}}{resource_id}

\item{\code{resource_type}}{resource_type}

\item{\code{value}}{Value will match the schema of the field.}

\item{\code{detail}}{detail}

\item{\code{status}}{status}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceNotFoundProblem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceNotFoundProblem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceNotFoundProblem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResourceNotFoundProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceNotFoundProblem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceNotFoundProblem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ResourceNotFoundProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceNotFoundProblem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResourceNotFoundProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceNotFoundProblem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceNotFoundProblem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceNotFoundProblem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResourceNotFoundProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceNotFoundProblem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceNotFoundProblem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ResourceNotFoundProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceNotFoundProblem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResourceNotFoundProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceNotFoundProblem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceNotFoundProblem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ResourceNotFoundProblem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceNotFoundProblem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceNotFoundProblem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceNotFoundProblem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceNotFoundProblem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ResourceNotFoundProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceNotFoundProblem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceNotFoundProblem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceNotFoundProblem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceNotFoundProblem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceNotFoundProblem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceNotFoundProblem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceNotFoundProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceNotFoundProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceNotFoundProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
