% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_unauthorized_problem.R
\docType{class}
\name{ResourceUnauthorizedProblem}
\alias{ResourceUnauthorizedProblem}
\title{ResourceUnauthorizedProblem}
\format{
An \code{R6Class} generator object
}
\description{
ResourceUnauthorizedProblem Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Super class}{
\code{\link[twitter:Problem]{twitter::Problem}} -> \code{ResourceUnauthorizedProblem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detail}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character}

\item{\code{type}}{character}

\item{\code{parameter}}{character}

\item{\code{resource_id}}{character}

\item{\code{resource_type}}{character}

\item{\code{section}}{character}

\item{\code{value}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResourceUnauthorizedProblem-new}{\code{ResourceUnauthorizedProblem$new()}}
\item \href{#method-ResourceUnauthorizedProblem-toJSON}{\code{ResourceUnauthorizedProblem$toJSON()}}
\item \href{#method-ResourceUnauthorizedProblem-fromJSON}{\code{ResourceUnauthorizedProblem$fromJSON()}}
\item \href{#method-ResourceUnauthorizedProblem-toJSONString}{\code{ResourceUnauthorizedProblem$toJSONString()}}
\item \href{#method-ResourceUnauthorizedProblem-fromJSONString}{\code{ResourceUnauthorizedProblem$fromJSONString()}}
\item \href{#method-ResourceUnauthorizedProblem-validateJSON}{\code{ResourceUnauthorizedProblem$validateJSON()}}
\item \href{#method-ResourceUnauthorizedProblem-toString}{\code{ResourceUnauthorizedProblem$toString()}}
\item \href{#method-ResourceUnauthorizedProblem-isValid}{\code{ResourceUnauthorizedProblem$isValid()}}
\item \href{#method-ResourceUnauthorizedProblem-getInvalidFields}{\code{ResourceUnauthorizedProblem$getInvalidFields()}}
\item \href{#method-ResourceUnauthorizedProblem-clone}{\code{ResourceUnauthorizedProblem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ResourceUnauthorizedProblem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblem$new(
  title,
  type,
  parameter,
  resource_id,
  resource_type,
  section,
  value,
  detail = NULL,
  status = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{type}}{type}

\item{\code{parameter}}{parameter}

\item{\code{resource_id}}{resource_id}

\item{\code{resource_type}}{resource_type}

\item{\code{section}}{section}

\item{\code{value}}{value}

\item{\code{detail}}{detail}

\item{\code{status}}{status}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResourceUnauthorizedProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ResourceUnauthorizedProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResourceUnauthorizedProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResourceUnauthorizedProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ResourceUnauthorizedProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResourceUnauthorizedProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ResourceUnauthorizedProblem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ResourceUnauthorizedProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
