% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_unauthorized_problem_all_of.R
\docType{class}
\name{ResourceUnauthorizedProblemAllOf}
\alias{ResourceUnauthorizedProblemAllOf}
\title{ResourceUnauthorizedProblemAllOf}
\format{
An \code{R6Class} generator object
}
\description{
ResourceUnauthorizedProblemAllOf Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parameter}}{character}

\item{\code{resource_id}}{character}

\item{\code{resource_type}}{character}

\item{\code{section}}{character}

\item{\code{value}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResourceUnauthorizedProblemAllOf-new}{\code{ResourceUnauthorizedProblemAllOf$new()}}
\item \href{#method-ResourceUnauthorizedProblemAllOf-toJSON}{\code{ResourceUnauthorizedProblemAllOf$toJSON()}}
\item \href{#method-ResourceUnauthorizedProblemAllOf-fromJSON}{\code{ResourceUnauthorizedProblemAllOf$fromJSON()}}
\item \href{#method-ResourceUnauthorizedProblemAllOf-toJSONString}{\code{ResourceUnauthorizedProblemAllOf$toJSONString()}}
\item \href{#method-ResourceUnauthorizedProblemAllOf-fromJSONString}{\code{ResourceUnauthorizedProblemAllOf$fromJSONString()}}
\item \href{#method-ResourceUnauthorizedProblemAllOf-validateJSON}{\code{ResourceUnauthorizedProblemAllOf$validateJSON()}}
\item \href{#method-ResourceUnauthorizedProblemAllOf-toString}{\code{ResourceUnauthorizedProblemAllOf$toString()}}
\item \href{#method-ResourceUnauthorizedProblemAllOf-isValid}{\code{ResourceUnauthorizedProblemAllOf$isValid()}}
\item \href{#method-ResourceUnauthorizedProblemAllOf-getInvalidFields}{\code{ResourceUnauthorizedProblemAllOf$getInvalidFields()}}
\item \href{#method-ResourceUnauthorizedProblemAllOf-clone}{\code{ResourceUnauthorizedProblemAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblemAllOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblemAllOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ResourceUnauthorizedProblemAllOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblemAllOf$new(
  parameter,
  resource_id,
  resource_type,
  section,
  value,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameter}}{parameter}

\item{\code{resource_id}}{resource_id}

\item{\code{resource_type}}{resource_type}

\item{\code{section}}{section}

\item{\code{value}}{value}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblemAllOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblemAllOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblemAllOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResourceUnauthorizedProblemAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblemAllOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblemAllOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ResourceUnauthorizedProblemAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblemAllOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResourceUnauthorizedProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblemAllOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblemAllOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblemAllOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResourceUnauthorizedProblemAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblemAllOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblemAllOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ResourceUnauthorizedProblemAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblemAllOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResourceUnauthorizedProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblemAllOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblemAllOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ResourceUnauthorizedProblemAllOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblemAllOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblemAllOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblemAllOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblemAllOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ResourceUnauthorizedProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblemAllOf-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblemAllOf-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblemAllOf$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblemAllOf-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblemAllOf-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblemAllOf$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnauthorizedProblemAllOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnauthorizedProblemAllOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnauthorizedProblemAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
