% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_unavailable_problem_all_of.R
\docType{class}
\name{ResourceUnavailableProblemAllOf}
\alias{ResourceUnavailableProblemAllOf}
\title{ResourceUnavailableProblemAllOf}
\format{
An \code{R6Class} generator object
}
\description{
ResourceUnavailableProblemAllOf Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parameter}}{character}

\item{\code{resource_id}}{character}

\item{\code{resource_type}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResourceUnavailableProblemAllOf-new}{\code{ResourceUnavailableProblemAllOf$new()}}
\item \href{#method-ResourceUnavailableProblemAllOf-toJSON}{\code{ResourceUnavailableProblemAllOf$toJSON()}}
\item \href{#method-ResourceUnavailableProblemAllOf-fromJSON}{\code{ResourceUnavailableProblemAllOf$fromJSON()}}
\item \href{#method-ResourceUnavailableProblemAllOf-toJSONString}{\code{ResourceUnavailableProblemAllOf$toJSONString()}}
\item \href{#method-ResourceUnavailableProblemAllOf-fromJSONString}{\code{ResourceUnavailableProblemAllOf$fromJSONString()}}
\item \href{#method-ResourceUnavailableProblemAllOf-validateJSON}{\code{ResourceUnavailableProblemAllOf$validateJSON()}}
\item \href{#method-ResourceUnavailableProblemAllOf-toString}{\code{ResourceUnavailableProblemAllOf$toString()}}
\item \href{#method-ResourceUnavailableProblemAllOf-isValid}{\code{ResourceUnavailableProblemAllOf$isValid()}}
\item \href{#method-ResourceUnavailableProblemAllOf-getInvalidFields}{\code{ResourceUnavailableProblemAllOf$getInvalidFields()}}
\item \href{#method-ResourceUnavailableProblemAllOf-clone}{\code{ResourceUnavailableProblemAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnavailableProblemAllOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnavailableProblemAllOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ResourceUnavailableProblemAllOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnavailableProblemAllOf$new(parameter, resource_id, resource_type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameter}}{parameter}

\item{\code{resource_id}}{resource_id}

\item{\code{resource_type}}{resource_type}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnavailableProblemAllOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnavailableProblemAllOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnavailableProblemAllOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResourceUnavailableProblemAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnavailableProblemAllOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnavailableProblemAllOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ResourceUnavailableProblemAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnavailableProblemAllOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResourceUnavailableProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnavailableProblemAllOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnavailableProblemAllOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnavailableProblemAllOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResourceUnavailableProblemAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnavailableProblemAllOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnavailableProblemAllOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ResourceUnavailableProblemAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnavailableProblemAllOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResourceUnavailableProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnavailableProblemAllOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnavailableProblemAllOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ResourceUnavailableProblemAllOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnavailableProblemAllOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnavailableProblemAllOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnavailableProblemAllOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnavailableProblemAllOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ResourceUnavailableProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnavailableProblemAllOf-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnavailableProblemAllOf-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnavailableProblemAllOf$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnavailableProblemAllOf-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnavailableProblemAllOf-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnavailableProblemAllOf$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResourceUnavailableProblemAllOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResourceUnavailableProblemAllOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceUnavailableProblemAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
