% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_cap_problem.R
\docType{class}
\name{RulesCapProblem}
\alias{RulesCapProblem}
\title{RulesCapProblem}
\format{
An \code{R6Class} generator object
}
\description{
RulesCapProblem Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Super class}{
\code{\link[twitter:Problem]{twitter::Problem}} -> \code{RulesCapProblem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detail}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character}

\item{\code{type}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RulesCapProblem-new}{\code{RulesCapProblem$new()}}
\item \href{#method-RulesCapProblem-toJSON}{\code{RulesCapProblem$toJSON()}}
\item \href{#method-RulesCapProblem-fromJSON}{\code{RulesCapProblem$fromJSON()}}
\item \href{#method-RulesCapProblem-toJSONString}{\code{RulesCapProblem$toJSONString()}}
\item \href{#method-RulesCapProblem-fromJSONString}{\code{RulesCapProblem$fromJSONString()}}
\item \href{#method-RulesCapProblem-validateJSON}{\code{RulesCapProblem$validateJSON()}}
\item \href{#method-RulesCapProblem-toString}{\code{RulesCapProblem$toString()}}
\item \href{#method-RulesCapProblem-isValid}{\code{RulesCapProblem$isValid()}}
\item \href{#method-RulesCapProblem-getInvalidFields}{\code{RulesCapProblem$getInvalidFields()}}
\item \href{#method-RulesCapProblem-clone}{\code{RulesCapProblem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesCapProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-RulesCapProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RulesCapProblem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesCapProblem$new(title, type, detail = NULL, status = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{type}}{type}

\item{\code{detail}}{detail}

\item{\code{status}}{status}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesCapProblem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RulesCapProblem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesCapProblem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RulesCapProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesCapProblem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RulesCapProblem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of RulesCapProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesCapProblem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RulesCapProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesCapProblem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RulesCapProblem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesCapProblem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RulesCapProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesCapProblem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RulesCapProblem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of RulesCapProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesCapProblem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RulesCapProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesCapProblem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RulesCapProblem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to RulesCapProblem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesCapProblem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesCapProblem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-RulesCapProblem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesCapProblem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of RulesCapProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesCapProblem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-RulesCapProblem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesCapProblem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesCapProblem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-RulesCapProblem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesCapProblem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesCapProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RulesCapProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesCapProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
