% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_request_summary.R
\docType{class}
\name{RulesRequestSummary}
\alias{RulesRequestSummary}
\title{RulesRequestSummary}
\format{
An \code{R6Class} generator object
}
\description{
RulesRequestSummary Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of object types defined in the oneOf schema.
Initialize a new RulesRequestSummary.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RulesRequestSummary-new}{\code{RulesRequestSummary$new()}}
\item \href{#method-RulesRequestSummary-fromJSON}{\code{RulesRequestSummary$fromJSON()}}
\item \href{#method-RulesRequestSummary-toJSONString}{\code{RulesRequestSummary$toJSONString()}}
\item \href{#method-RulesRequestSummary-toJSON}{\code{RulesRequestSummary$toJSON()}}
\item \href{#method-RulesRequestSummary-validateJSON}{\code{RulesRequestSummary$validateJSON()}}
\item \href{#method-RulesRequestSummary-toString}{\code{RulesRequestSummary$toString()}}
\item \href{#method-RulesRequestSummary-clone}{\code{RulesRequestSummary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummary-new"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummary-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RulesRequestSummary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummary$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "RulesRequestSummaryOneOf", "RulesRequestSummaryOneOf1"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummary-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummary-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of RulesRequestSummary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummary$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of RulesRequestSummary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummary-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummary-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize RulesRequestSummary to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummary$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the RulesRequestSummary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummary-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummary-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize RulesRequestSummary to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummary$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the RulesRequestSummary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummary-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummary-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to RulesRequestSummary and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummary$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummary-toString"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummary-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummary$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummary-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummary-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
