% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_request_summary_one_of1.R
\docType{class}
\name{RulesRequestSummaryOneOf1}
\alias{RulesRequestSummaryOneOf1}
\title{RulesRequestSummaryOneOf1}
\format{
An \code{R6Class} generator object
}
\description{
RulesRequestSummaryOneOf1 Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deleted}}{integer}

\item{\code{not_deleted}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RulesRequestSummaryOneOf1-new}{\code{RulesRequestSummaryOneOf1$new()}}
\item \href{#method-RulesRequestSummaryOneOf1-toJSON}{\code{RulesRequestSummaryOneOf1$toJSON()}}
\item \href{#method-RulesRequestSummaryOneOf1-fromJSON}{\code{RulesRequestSummaryOneOf1$fromJSON()}}
\item \href{#method-RulesRequestSummaryOneOf1-toJSONString}{\code{RulesRequestSummaryOneOf1$toJSONString()}}
\item \href{#method-RulesRequestSummaryOneOf1-fromJSONString}{\code{RulesRequestSummaryOneOf1$fromJSONString()}}
\item \href{#method-RulesRequestSummaryOneOf1-validateJSON}{\code{RulesRequestSummaryOneOf1$validateJSON()}}
\item \href{#method-RulesRequestSummaryOneOf1-toString}{\code{RulesRequestSummaryOneOf1$toString()}}
\item \href{#method-RulesRequestSummaryOneOf1-isValid}{\code{RulesRequestSummaryOneOf1$isValid()}}
\item \href{#method-RulesRequestSummaryOneOf1-getInvalidFields}{\code{RulesRequestSummaryOneOf1$getInvalidFields()}}
\item \href{#method-RulesRequestSummaryOneOf1-clone}{\code{RulesRequestSummaryOneOf1$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummaryOneOf1-new"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummaryOneOf1-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RulesRequestSummaryOneOf1 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummaryOneOf1$new(deleted, not_deleted, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deleted}}{Number of user-specified stream filtering rules that were deleted.}

\item{\code{not_deleted}}{Number of user-specified stream filtering rules that were not deleted.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummaryOneOf1-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummaryOneOf1-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummaryOneOf1$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RulesRequestSummaryOneOf1 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummaryOneOf1-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummaryOneOf1-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of RulesRequestSummaryOneOf1
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummaryOneOf1$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RulesRequestSummaryOneOf1
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummaryOneOf1-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummaryOneOf1-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummaryOneOf1$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RulesRequestSummaryOneOf1 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummaryOneOf1-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummaryOneOf1-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of RulesRequestSummaryOneOf1
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummaryOneOf1$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RulesRequestSummaryOneOf1
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummaryOneOf1-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummaryOneOf1-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to RulesRequestSummaryOneOf1 and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummaryOneOf1$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummaryOneOf1-toString"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummaryOneOf1-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummaryOneOf1$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of RulesRequestSummaryOneOf1
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummaryOneOf1-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummaryOneOf1-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummaryOneOf1$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummaryOneOf1-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummaryOneOf1-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummaryOneOf1$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesRequestSummaryOneOf1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RulesRequestSummaryOneOf1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesRequestSummaryOneOf1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
