% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_response_metadata.R
\docType{class}
\name{RulesResponseMetadata}
\alias{RulesResponseMetadata}
\title{RulesResponseMetadata}
\format{
An \code{R6Class} generator object
}
\description{
RulesResponseMetadata Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{next_token}}{character [optional]}

\item{\code{result_count}}{integer [optional]}

\item{\code{sent}}{character}

\item{\code{summary}}{\link{RulesRequestSummary} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RulesResponseMetadata-new}{\code{RulesResponseMetadata$new()}}
\item \href{#method-RulesResponseMetadata-toJSON}{\code{RulesResponseMetadata$toJSON()}}
\item \href{#method-RulesResponseMetadata-fromJSON}{\code{RulesResponseMetadata$fromJSON()}}
\item \href{#method-RulesResponseMetadata-toJSONString}{\code{RulesResponseMetadata$toJSONString()}}
\item \href{#method-RulesResponseMetadata-fromJSONString}{\code{RulesResponseMetadata$fromJSONString()}}
\item \href{#method-RulesResponseMetadata-validateJSON}{\code{RulesResponseMetadata$validateJSON()}}
\item \href{#method-RulesResponseMetadata-toString}{\code{RulesResponseMetadata$toString()}}
\item \href{#method-RulesResponseMetadata-isValid}{\code{RulesResponseMetadata$isValid()}}
\item \href{#method-RulesResponseMetadata-getInvalidFields}{\code{RulesResponseMetadata$getInvalidFields()}}
\item \href{#method-RulesResponseMetadata-clone}{\code{RulesResponseMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesResponseMetadata-new"></a>}}
\if{latex}{\out{\hypertarget{method-RulesResponseMetadata-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new RulesResponseMetadata class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesResponseMetadata$new(
  sent,
  next_token = NULL,
  result_count = NULL,
  summary = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sent}}{sent}

\item{\code{next_token}}{The next token.}

\item{\code{result_count}}{Number of Rules in result set.}

\item{\code{summary}}{summary}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesResponseMetadata-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RulesResponseMetadata-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesResponseMetadata$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RulesResponseMetadata in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesResponseMetadata-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RulesResponseMetadata-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of RulesResponseMetadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesResponseMetadata$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RulesResponseMetadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesResponseMetadata-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RulesResponseMetadata-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesResponseMetadata$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
RulesResponseMetadata in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesResponseMetadata-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-RulesResponseMetadata-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of RulesResponseMetadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesResponseMetadata$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of RulesResponseMetadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesResponseMetadata-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-RulesResponseMetadata-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to RulesResponseMetadata and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesResponseMetadata$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesResponseMetadata-toString"></a>}}
\if{latex}{\out{\hypertarget{method-RulesResponseMetadata-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesResponseMetadata$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of RulesResponseMetadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesResponseMetadata-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-RulesResponseMetadata-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesResponseMetadata$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesResponseMetadata-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-RulesResponseMetadata-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesResponseMetadata$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RulesResponseMetadata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RulesResponseMetadata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RulesResponseMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
