% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_count.R
\docType{class}
\name{SearchCount}
\alias{SearchCount}
\title{SearchCount}
\format{
An \code{R6Class} generator object
}
\description{
SearchCount Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{end}}{character}

\item{\code{start}}{character}

\item{\code{tweet_count}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SearchCount-new}{\code{SearchCount$new()}}
\item \href{#method-SearchCount-toJSON}{\code{SearchCount$toJSON()}}
\item \href{#method-SearchCount-fromJSON}{\code{SearchCount$fromJSON()}}
\item \href{#method-SearchCount-toJSONString}{\code{SearchCount$toJSONString()}}
\item \href{#method-SearchCount-fromJSONString}{\code{SearchCount$fromJSONString()}}
\item \href{#method-SearchCount-validateJSON}{\code{SearchCount$validateJSON()}}
\item \href{#method-SearchCount-toString}{\code{SearchCount$toString()}}
\item \href{#method-SearchCount-isValid}{\code{SearchCount$isValid()}}
\item \href{#method-SearchCount-getInvalidFields}{\code{SearchCount$getInvalidFields()}}
\item \href{#method-SearchCount-clone}{\code{SearchCount$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchCount-new"></a>}}
\if{latex}{\out{\hypertarget{method-SearchCount-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SearchCount class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchCount$new(end, start, tweet_count, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{end}}{The end time of the bucket.}

\item{\code{start}}{The start time of the bucket.}

\item{\code{tweet_count}}{The count for the bucket.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchCount-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SearchCount-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchCount$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SearchCount in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchCount-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SearchCount-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of SearchCount
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchCount$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SearchCount
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchCount-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SearchCount-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchCount$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SearchCount in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchCount-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SearchCount-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of SearchCount
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchCount$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SearchCount
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchCount-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SearchCount-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to SearchCount and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchCount$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchCount-toString"></a>}}
\if{latex}{\out{\hypertarget{method-SearchCount-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchCount$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of SearchCount
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchCount-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-SearchCount-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchCount$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchCount-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-SearchCount-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchCount$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchCount-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SearchCount-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SearchCount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
