% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_topics_inner.R
\docType{class}
\name{SpaceTopicsInner}
\alias{SpaceTopicsInner}
\title{SpaceTopicsInner}
\format{
An \code{R6Class} generator object
}
\description{
SpaceTopicsInner Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character [optional]}

\item{\code{id}}{character}

\item{\code{name}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SpaceTopicsInner-new}{\code{SpaceTopicsInner$new()}}
\item \href{#method-SpaceTopicsInner-toJSON}{\code{SpaceTopicsInner$toJSON()}}
\item \href{#method-SpaceTopicsInner-fromJSON}{\code{SpaceTopicsInner$fromJSON()}}
\item \href{#method-SpaceTopicsInner-toJSONString}{\code{SpaceTopicsInner$toJSONString()}}
\item \href{#method-SpaceTopicsInner-fromJSONString}{\code{SpaceTopicsInner$fromJSONString()}}
\item \href{#method-SpaceTopicsInner-validateJSON}{\code{SpaceTopicsInner$validateJSON()}}
\item \href{#method-SpaceTopicsInner-toString}{\code{SpaceTopicsInner$toString()}}
\item \href{#method-SpaceTopicsInner-isValid}{\code{SpaceTopicsInner$isValid()}}
\item \href{#method-SpaceTopicsInner-getInvalidFields}{\code{SpaceTopicsInner$getInvalidFields()}}
\item \href{#method-SpaceTopicsInner-clone}{\code{SpaceTopicsInner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpaceTopicsInner-new"></a>}}
\if{latex}{\out{\hypertarget{method-SpaceTopicsInner-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SpaceTopicsInner class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceTopicsInner$new(id, name, description = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{An ID suitable for use in the REST API.}

\item{\code{name}}{The name of the given topic.}

\item{\code{description}}{The description of the given topic.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpaceTopicsInner-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SpaceTopicsInner-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceTopicsInner$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SpaceTopicsInner in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpaceTopicsInner-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SpaceTopicsInner-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of SpaceTopicsInner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceTopicsInner$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SpaceTopicsInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpaceTopicsInner-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SpaceTopicsInner-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceTopicsInner$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
SpaceTopicsInner in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpaceTopicsInner-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SpaceTopicsInner-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of SpaceTopicsInner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceTopicsInner$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of SpaceTopicsInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpaceTopicsInner-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SpaceTopicsInner-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to SpaceTopicsInner and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceTopicsInner$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpaceTopicsInner-toString"></a>}}
\if{latex}{\out{\hypertarget{method-SpaceTopicsInner-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceTopicsInner$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of SpaceTopicsInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpaceTopicsInner-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-SpaceTopicsInner-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceTopicsInner$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpaceTopicsInner-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-SpaceTopicsInner-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceTopicsInner$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpaceTopicsInner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpaceTopicsInner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpaceTopicsInner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
