% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming_tweet_response.R
\docType{class}
\name{StreamingTweetResponse}
\alias{StreamingTweetResponse}
\title{StreamingTweetResponse}
\format{
An \code{R6Class} generator object
}
\description{
StreamingTweetResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{Tweet} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}

\item{\code{includes}}{\link{Expansions} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StreamingTweetResponse-new}{\code{StreamingTweetResponse$new()}}
\item \href{#method-StreamingTweetResponse-toJSON}{\code{StreamingTweetResponse$toJSON()}}
\item \href{#method-StreamingTweetResponse-fromJSON}{\code{StreamingTweetResponse$fromJSON()}}
\item \href{#method-StreamingTweetResponse-toJSONString}{\code{StreamingTweetResponse$toJSONString()}}
\item \href{#method-StreamingTweetResponse-fromJSONString}{\code{StreamingTweetResponse$fromJSONString()}}
\item \href{#method-StreamingTweetResponse-validateJSON}{\code{StreamingTweetResponse$validateJSON()}}
\item \href{#method-StreamingTweetResponse-toString}{\code{StreamingTweetResponse$toString()}}
\item \href{#method-StreamingTweetResponse-isValid}{\code{StreamingTweetResponse$isValid()}}
\item \href{#method-StreamingTweetResponse-getInvalidFields}{\code{StreamingTweetResponse$getInvalidFields()}}
\item \href{#method-StreamingTweetResponse-clone}{\code{StreamingTweetResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamingTweetResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-StreamingTweetResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new StreamingTweetResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamingTweetResponse$new(data = NULL, errors = NULL, includes = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{includes}}{includes}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamingTweetResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StreamingTweetResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamingTweetResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StreamingTweetResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamingTweetResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StreamingTweetResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of StreamingTweetResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamingTweetResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StreamingTweetResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamingTweetResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StreamingTweetResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamingTweetResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
StreamingTweetResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamingTweetResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StreamingTweetResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of StreamingTweetResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamingTweetResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of StreamingTweetResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamingTweetResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StreamingTweetResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to StreamingTweetResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamingTweetResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamingTweetResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-StreamingTweetResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamingTweetResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of StreamingTweetResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamingTweetResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-StreamingTweetResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamingTweetResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamingTweetResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-StreamingTweetResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamingTweetResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamingTweetResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StreamingTweetResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamingTweetResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
