% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic.R
\docType{class}
\name{Topic}
\alias{Topic}
\title{Topic}
\format{
An \code{R6Class} generator object
}
\description{
Topic Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character [optional]}

\item{\code{id}}{character}

\item{\code{name}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Topic-new}{\code{Topic$new()}}
\item \href{#method-Topic-toJSON}{\code{Topic$toJSON()}}
\item \href{#method-Topic-fromJSON}{\code{Topic$fromJSON()}}
\item \href{#method-Topic-toJSONString}{\code{Topic$toJSONString()}}
\item \href{#method-Topic-fromJSONString}{\code{Topic$fromJSONString()}}
\item \href{#method-Topic-validateJSON}{\code{Topic$validateJSON()}}
\item \href{#method-Topic-toString}{\code{Topic$toString()}}
\item \href{#method-Topic-isValid}{\code{Topic$isValid()}}
\item \href{#method-Topic-getInvalidFields}{\code{Topic$getInvalidFields()}}
\item \href{#method-Topic-clone}{\code{Topic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Topic-new"></a>}}
\if{latex}{\out{\hypertarget{method-Topic-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Topic class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topic$new(id, name, description = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Unique identifier of this Topic.}

\item{\code{name}}{The name of the given topic.}

\item{\code{description}}{The description of the given topic.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Topic-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Topic-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topic$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Topic in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Topic-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Topic-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Topic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topic$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Topic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Topic-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Topic-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topic$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Topic in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Topic-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Topic-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Topic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topic$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Topic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Topic-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Topic-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Topic and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topic$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Topic-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Topic-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topic$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Topic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Topic-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Topic-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topic$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Topic-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Topic-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topic$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Topic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Topic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Topic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
