% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_attachments.R
\docType{class}
\name{TweetAttachments}
\alias{TweetAttachments}
\title{TweetAttachments}
\format{
An \code{R6Class} generator object
}
\description{
TweetAttachments Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{media_keys}}{list( character ) [optional]}

\item{\code{poll_ids}}{list( character ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetAttachments-new}{\code{TweetAttachments$new()}}
\item \href{#method-TweetAttachments-toJSON}{\code{TweetAttachments$toJSON()}}
\item \href{#method-TweetAttachments-fromJSON}{\code{TweetAttachments$fromJSON()}}
\item \href{#method-TweetAttachments-toJSONString}{\code{TweetAttachments$toJSONString()}}
\item \href{#method-TweetAttachments-fromJSONString}{\code{TweetAttachments$fromJSONString()}}
\item \href{#method-TweetAttachments-validateJSON}{\code{TweetAttachments$validateJSON()}}
\item \href{#method-TweetAttachments-toString}{\code{TweetAttachments$toString()}}
\item \href{#method-TweetAttachments-isValid}{\code{TweetAttachments$isValid()}}
\item \href{#method-TweetAttachments-getInvalidFields}{\code{TweetAttachments$getInvalidFields()}}
\item \href{#method-TweetAttachments-clone}{\code{TweetAttachments$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetAttachments-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetAttachments-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetAttachments class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAttachments$new(media_keys = NULL, poll_ids = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{media_keys}}{A list of Media Keys for each one of the media attachments (if media are attached).}

\item{\code{poll_ids}}{A list of poll IDs (if polls are attached).}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetAttachments-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetAttachments-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAttachments$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetAttachments in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetAttachments-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetAttachments-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetAttachments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAttachments$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetAttachments
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetAttachments-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetAttachments-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAttachments$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetAttachments in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetAttachments-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetAttachments-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetAttachments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAttachments$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetAttachments
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetAttachments-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetAttachments-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetAttachments and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAttachments$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetAttachments-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetAttachments-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAttachments$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetAttachments
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetAttachments-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetAttachments-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAttachments$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetAttachments-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetAttachments-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAttachments$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetAttachments-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetAttachments-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAttachments$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
