% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_compliance_data.R
\docType{class}
\name{TweetComplianceData}
\alias{TweetComplianceData}
\title{TweetComplianceData}
\format{
An \code{R6Class} generator object
}
\description{
TweetComplianceData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of object types defined in the oneOf schema.
Initialize a new TweetComplianceData.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetComplianceData-new}{\code{TweetComplianceData$new()}}
\item \href{#method-TweetComplianceData-fromJSON}{\code{TweetComplianceData$fromJSON()}}
\item \href{#method-TweetComplianceData-toJSONString}{\code{TweetComplianceData$toJSONString()}}
\item \href{#method-TweetComplianceData-toJSON}{\code{TweetComplianceData$toJSON()}}
\item \href{#method-TweetComplianceData-validateJSON}{\code{TweetComplianceData$validateJSON()}}
\item \href{#method-TweetComplianceData-toString}{\code{TweetComplianceData$toString()}}
\item \href{#method-TweetComplianceData-clone}{\code{TweetComplianceData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceData-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetComplianceData.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceData$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "TweetDeleteComplianceSchema", "TweetDropComplianceSchema", "TweetUndropComplianceSchema", "TweetWithheldComplianceSchema"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetComplianceData.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceData$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of TweetComplianceData.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize TweetComplianceData to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the TweetComplianceData.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize TweetComplianceData to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the TweetComplianceData.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to TweetComplianceData and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceData-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
