% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_compliance_schema.R
\docType{class}
\name{TweetComplianceSchema}
\alias{TweetComplianceSchema}
\title{TweetComplianceSchema}
\format{
An \code{R6Class} generator object
}
\description{
TweetComplianceSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{event_at}}{character}

\item{\code{quote_tweet_id}}{character [optional]}

\item{\code{tweet}}{\link{TweetComplianceSchemaTweet}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetComplianceSchema-new}{\code{TweetComplianceSchema$new()}}
\item \href{#method-TweetComplianceSchema-toJSON}{\code{TweetComplianceSchema$toJSON()}}
\item \href{#method-TweetComplianceSchema-fromJSON}{\code{TweetComplianceSchema$fromJSON()}}
\item \href{#method-TweetComplianceSchema-toJSONString}{\code{TweetComplianceSchema$toJSONString()}}
\item \href{#method-TweetComplianceSchema-fromJSONString}{\code{TweetComplianceSchema$fromJSONString()}}
\item \href{#method-TweetComplianceSchema-validateJSON}{\code{TweetComplianceSchema$validateJSON()}}
\item \href{#method-TweetComplianceSchema-toString}{\code{TweetComplianceSchema$toString()}}
\item \href{#method-TweetComplianceSchema-isValid}{\code{TweetComplianceSchema$isValid()}}
\item \href{#method-TweetComplianceSchema-getInvalidFields}{\code{TweetComplianceSchema$getInvalidFields()}}
\item \href{#method-TweetComplianceSchema-clone}{\code{TweetComplianceSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetComplianceSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceSchema$new(event_at, tweet, quote_tweet_id = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_at}}{Event time.}

\item{\code{tweet}}{tweet}

\item{\code{quote_tweet_id}}{Unique identifier of this Tweet. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetComplianceSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
