% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_compliance_stream_response.R
\docType{class}
\name{TweetComplianceStreamResponse}
\alias{TweetComplianceStreamResponse}
\title{TweetComplianceStreamResponse}
\format{
An \code{R6Class} generator object
}
\description{
TweetComplianceStreamResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of object types defined in the oneOf schema.
Initialize a new TweetComplianceStreamResponse.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetComplianceStreamResponse-new}{\code{TweetComplianceStreamResponse$new()}}
\item \href{#method-TweetComplianceStreamResponse-fromJSON}{\code{TweetComplianceStreamResponse$fromJSON()}}
\item \href{#method-TweetComplianceStreamResponse-toJSONString}{\code{TweetComplianceStreamResponse$toJSONString()}}
\item \href{#method-TweetComplianceStreamResponse-toJSON}{\code{TweetComplianceStreamResponse$toJSON()}}
\item \href{#method-TweetComplianceStreamResponse-validateJSON}{\code{TweetComplianceStreamResponse$validateJSON()}}
\item \href{#method-TweetComplianceStreamResponse-toString}{\code{TweetComplianceStreamResponse$toString()}}
\item \href{#method-TweetComplianceStreamResponse-clone}{\code{TweetComplianceStreamResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetComplianceStreamResponse.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponse$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "TweetComplianceStreamResponseOneOf", "TweetComplianceStreamResponseOneOf1"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetComplianceStreamResponse.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponse$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of TweetComplianceStreamResponse.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize TweetComplianceStreamResponse to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the TweetComplianceStreamResponse.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize TweetComplianceStreamResponse to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the TweetComplianceStreamResponse.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to TweetComplianceStreamResponse and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponse-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
