% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_compliance_stream_response_one_of.R
\docType{class}
\name{TweetComplianceStreamResponseOneOf}
\alias{TweetComplianceStreamResponseOneOf}
\title{TweetComplianceStreamResponseOneOf}
\format{
An \code{R6Class} generator object
}
\description{
TweetComplianceStreamResponseOneOf Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{TweetComplianceData}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetComplianceStreamResponseOneOf-new}{\code{TweetComplianceStreamResponseOneOf$new()}}
\item \href{#method-TweetComplianceStreamResponseOneOf-toJSON}{\code{TweetComplianceStreamResponseOneOf$toJSON()}}
\item \href{#method-TweetComplianceStreamResponseOneOf-fromJSON}{\code{TweetComplianceStreamResponseOneOf$fromJSON()}}
\item \href{#method-TweetComplianceStreamResponseOneOf-toJSONString}{\code{TweetComplianceStreamResponseOneOf$toJSONString()}}
\item \href{#method-TweetComplianceStreamResponseOneOf-fromJSONString}{\code{TweetComplianceStreamResponseOneOf$fromJSONString()}}
\item \href{#method-TweetComplianceStreamResponseOneOf-validateJSON}{\code{TweetComplianceStreamResponseOneOf$validateJSON()}}
\item \href{#method-TweetComplianceStreamResponseOneOf-toString}{\code{TweetComplianceStreamResponseOneOf$toString()}}
\item \href{#method-TweetComplianceStreamResponseOneOf-isValid}{\code{TweetComplianceStreamResponseOneOf$isValid()}}
\item \href{#method-TweetComplianceStreamResponseOneOf-getInvalidFields}{\code{TweetComplianceStreamResponseOneOf$getInvalidFields()}}
\item \href{#method-TweetComplianceStreamResponseOneOf-clone}{\code{TweetComplianceStreamResponseOneOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetComplianceStreamResponseOneOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf$new(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetComplianceStreamResponseOneOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetComplianceStreamResponseOneOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetComplianceStreamResponseOneOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetComplianceStreamResponseOneOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetComplianceStreamResponseOneOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetComplianceStreamResponseOneOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetComplianceStreamResponseOneOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetComplianceStreamResponseOneOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
