% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_compliance_stream_response_one_of1.R
\docType{class}
\name{TweetComplianceStreamResponseOneOf1}
\alias{TweetComplianceStreamResponseOneOf1}
\title{TweetComplianceStreamResponseOneOf1}
\format{
An \code{R6Class} generator object
}
\description{
TweetComplianceStreamResponseOneOf1 Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{errors}}{list( \link{Problem} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetComplianceStreamResponseOneOf1-new}{\code{TweetComplianceStreamResponseOneOf1$new()}}
\item \href{#method-TweetComplianceStreamResponseOneOf1-toJSON}{\code{TweetComplianceStreamResponseOneOf1$toJSON()}}
\item \href{#method-TweetComplianceStreamResponseOneOf1-fromJSON}{\code{TweetComplianceStreamResponseOneOf1$fromJSON()}}
\item \href{#method-TweetComplianceStreamResponseOneOf1-toJSONString}{\code{TweetComplianceStreamResponseOneOf1$toJSONString()}}
\item \href{#method-TweetComplianceStreamResponseOneOf1-fromJSONString}{\code{TweetComplianceStreamResponseOneOf1$fromJSONString()}}
\item \href{#method-TweetComplianceStreamResponseOneOf1-validateJSON}{\code{TweetComplianceStreamResponseOneOf1$validateJSON()}}
\item \href{#method-TweetComplianceStreamResponseOneOf1-toString}{\code{TweetComplianceStreamResponseOneOf1$toString()}}
\item \href{#method-TweetComplianceStreamResponseOneOf1-isValid}{\code{TweetComplianceStreamResponseOneOf1$isValid()}}
\item \href{#method-TweetComplianceStreamResponseOneOf1-getInvalidFields}{\code{TweetComplianceStreamResponseOneOf1$getInvalidFields()}}
\item \href{#method-TweetComplianceStreamResponseOneOf1-clone}{\code{TweetComplianceStreamResponseOneOf1$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf1-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf1-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetComplianceStreamResponseOneOf1 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf1$new(errors, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf1-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf1-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf1$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetComplianceStreamResponseOneOf1 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf1-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf1-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetComplianceStreamResponseOneOf1
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf1$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetComplianceStreamResponseOneOf1
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf1-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf1-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf1$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetComplianceStreamResponseOneOf1 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf1-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf1-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetComplianceStreamResponseOneOf1
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf1$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetComplianceStreamResponseOneOf1
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf1-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf1-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetComplianceStreamResponseOneOf1 and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf1$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf1-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf1-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf1$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetComplianceStreamResponseOneOf1
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf1-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf1-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf1$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf1-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf1-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf1$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetComplianceStreamResponseOneOf1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetComplianceStreamResponseOneOf1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetComplianceStreamResponseOneOf1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
