% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_create_request.R
\docType{class}
\name{TweetCreateRequest}
\alias{TweetCreateRequest}
\title{TweetCreateRequest}
\format{
An \code{R6Class} generator object
}
\description{
TweetCreateRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{direct_message_deep_link}}{character [optional]}

\item{\code{for_super_followers_only}}{character [optional]}

\item{\code{geo}}{\link{TweetCreateRequestGeo} [optional]}

\item{\code{media}}{\link{TweetCreateRequestMedia} [optional]}

\item{\code{poll}}{\link{TweetCreateRequestPoll} [optional]}

\item{\code{quote_tweet_id}}{character [optional]}

\item{\code{reply}}{\link{TweetCreateRequestReply} [optional]}

\item{\code{reply_settings}}{character [optional]}

\item{\code{text}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetCreateRequest-new}{\code{TweetCreateRequest$new()}}
\item \href{#method-TweetCreateRequest-toJSON}{\code{TweetCreateRequest$toJSON()}}
\item \href{#method-TweetCreateRequest-fromJSON}{\code{TweetCreateRequest$fromJSON()}}
\item \href{#method-TweetCreateRequest-toJSONString}{\code{TweetCreateRequest$toJSONString()}}
\item \href{#method-TweetCreateRequest-fromJSONString}{\code{TweetCreateRequest$fromJSONString()}}
\item \href{#method-TweetCreateRequest-validateJSON}{\code{TweetCreateRequest$validateJSON()}}
\item \href{#method-TweetCreateRequest-toString}{\code{TweetCreateRequest$toString()}}
\item \href{#method-TweetCreateRequest-isValid}{\code{TweetCreateRequest$isValid()}}
\item \href{#method-TweetCreateRequest-getInvalidFields}{\code{TweetCreateRequest$getInvalidFields()}}
\item \href{#method-TweetCreateRequest-clone}{\code{TweetCreateRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetCreateRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequest$new(
  direct_message_deep_link = NULL,
  for_super_followers_only = FALSE,
  geo = NULL,
  media = NULL,
  poll = NULL,
  quote_tweet_id = NULL,
  reply = NULL,
  reply_settings = NULL,
  text = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{direct_message_deep_link}}{Link to take the conversation from the public timeline to a private Direct Message.}

\item{\code{for_super_followers_only}}{Exclusive Tweet for super followers.. Default to FALSE.}

\item{\code{geo}}{geo}

\item{\code{media}}{media}

\item{\code{poll}}{poll}

\item{\code{quote_tweet_id}}{Unique identifier of this Tweet. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.}

\item{\code{reply}}{reply}

\item{\code{reply_settings}}{Settings to indicate who can reply to the Tweet.}

\item{\code{text}}{The content of the Tweet.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetCreateRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetCreateRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetCreateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetCreateRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetCreateRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetCreateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetCreateRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetCreateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
