% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_create_request_geo.R
\docType{class}
\name{TweetCreateRequestGeo}
\alias{TweetCreateRequestGeo}
\title{TweetCreateRequestGeo}
\format{
An \code{R6Class} generator object
}
\description{
TweetCreateRequestGeo Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{place_id}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetCreateRequestGeo-new}{\code{TweetCreateRequestGeo$new()}}
\item \href{#method-TweetCreateRequestGeo-toJSON}{\code{TweetCreateRequestGeo$toJSON()}}
\item \href{#method-TweetCreateRequestGeo-fromJSON}{\code{TweetCreateRequestGeo$fromJSON()}}
\item \href{#method-TweetCreateRequestGeo-toJSONString}{\code{TweetCreateRequestGeo$toJSONString()}}
\item \href{#method-TweetCreateRequestGeo-fromJSONString}{\code{TweetCreateRequestGeo$fromJSONString()}}
\item \href{#method-TweetCreateRequestGeo-validateJSON}{\code{TweetCreateRequestGeo$validateJSON()}}
\item \href{#method-TweetCreateRequestGeo-toString}{\code{TweetCreateRequestGeo$toString()}}
\item \href{#method-TweetCreateRequestGeo-isValid}{\code{TweetCreateRequestGeo$isValid()}}
\item \href{#method-TweetCreateRequestGeo-getInvalidFields}{\code{TweetCreateRequestGeo$getInvalidFields()}}
\item \href{#method-TweetCreateRequestGeo-clone}{\code{TweetCreateRequestGeo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestGeo-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestGeo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetCreateRequestGeo class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestGeo$new(place_id = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{place_id}}{place_id}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestGeo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestGeo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestGeo$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetCreateRequestGeo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestGeo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestGeo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetCreateRequestGeo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestGeo$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetCreateRequestGeo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestGeo-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestGeo-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestGeo$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetCreateRequestGeo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestGeo-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestGeo-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetCreateRequestGeo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestGeo$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetCreateRequestGeo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestGeo-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestGeo-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetCreateRequestGeo and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestGeo$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestGeo-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestGeo-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestGeo$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetCreateRequestGeo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestGeo-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestGeo-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestGeo$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestGeo-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestGeo-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestGeo$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestGeo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestGeo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestGeo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
