% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_create_request_media.R
\docType{class}
\name{TweetCreateRequestMedia}
\alias{TweetCreateRequestMedia}
\title{TweetCreateRequestMedia}
\format{
An \code{R6Class} generator object
}
\description{
TweetCreateRequestMedia Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{media_ids}}{list( character )}

\item{\code{tagged_user_ids}}{list( character ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetCreateRequestMedia-new}{\code{TweetCreateRequestMedia$new()}}
\item \href{#method-TweetCreateRequestMedia-toJSON}{\code{TweetCreateRequestMedia$toJSON()}}
\item \href{#method-TweetCreateRequestMedia-fromJSON}{\code{TweetCreateRequestMedia$fromJSON()}}
\item \href{#method-TweetCreateRequestMedia-toJSONString}{\code{TweetCreateRequestMedia$toJSONString()}}
\item \href{#method-TweetCreateRequestMedia-fromJSONString}{\code{TweetCreateRequestMedia$fromJSONString()}}
\item \href{#method-TweetCreateRequestMedia-validateJSON}{\code{TweetCreateRequestMedia$validateJSON()}}
\item \href{#method-TweetCreateRequestMedia-toString}{\code{TweetCreateRequestMedia$toString()}}
\item \href{#method-TweetCreateRequestMedia-isValid}{\code{TweetCreateRequestMedia$isValid()}}
\item \href{#method-TweetCreateRequestMedia-getInvalidFields}{\code{TweetCreateRequestMedia$getInvalidFields()}}
\item \href{#method-TweetCreateRequestMedia-clone}{\code{TweetCreateRequestMedia$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestMedia-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestMedia-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetCreateRequestMedia class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestMedia$new(media_ids, tagged_user_ids = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{media_ids}}{A list of Media Ids to be attached to a created Tweet.}

\item{\code{tagged_user_ids}}{A list of User Ids to be tagged in the media for created Tweet.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestMedia-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestMedia-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestMedia$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetCreateRequestMedia in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestMedia-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestMedia-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetCreateRequestMedia
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestMedia$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetCreateRequestMedia
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestMedia-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestMedia-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestMedia$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetCreateRequestMedia in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestMedia-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestMedia-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetCreateRequestMedia
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestMedia$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetCreateRequestMedia
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestMedia-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestMedia-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetCreateRequestMedia and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestMedia$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestMedia-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestMedia-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestMedia$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetCreateRequestMedia
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestMedia-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestMedia-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestMedia$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestMedia-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestMedia-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestMedia$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestMedia-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestMedia-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestMedia$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
