% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_create_request_poll.R
\docType{class}
\name{TweetCreateRequestPoll}
\alias{TweetCreateRequestPoll}
\title{TweetCreateRequestPoll}
\format{
An \code{R6Class} generator object
}
\description{
TweetCreateRequestPoll Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{duration_minutes}}{integer}

\item{\code{options}}{list( character )}

\item{\code{reply_settings}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetCreateRequestPoll-new}{\code{TweetCreateRequestPoll$new()}}
\item \href{#method-TweetCreateRequestPoll-toJSON}{\code{TweetCreateRequestPoll$toJSON()}}
\item \href{#method-TweetCreateRequestPoll-fromJSON}{\code{TweetCreateRequestPoll$fromJSON()}}
\item \href{#method-TweetCreateRequestPoll-toJSONString}{\code{TweetCreateRequestPoll$toJSONString()}}
\item \href{#method-TweetCreateRequestPoll-fromJSONString}{\code{TweetCreateRequestPoll$fromJSONString()}}
\item \href{#method-TweetCreateRequestPoll-validateJSON}{\code{TweetCreateRequestPoll$validateJSON()}}
\item \href{#method-TweetCreateRequestPoll-toString}{\code{TweetCreateRequestPoll$toString()}}
\item \href{#method-TweetCreateRequestPoll-isValid}{\code{TweetCreateRequestPoll$isValid()}}
\item \href{#method-TweetCreateRequestPoll-getInvalidFields}{\code{TweetCreateRequestPoll$getInvalidFields()}}
\item \href{#method-TweetCreateRequestPoll-clone}{\code{TweetCreateRequestPoll$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestPoll-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestPoll-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetCreateRequestPoll class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestPoll$new(
  duration_minutes,
  options,
  reply_settings = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{duration_minutes}}{Duration of the poll in minutes.}

\item{\code{options}}{options}

\item{\code{reply_settings}}{Settings to indicate who can reply to the Tweet.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestPoll-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestPoll-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestPoll$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetCreateRequestPoll in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestPoll-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestPoll-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetCreateRequestPoll
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestPoll$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetCreateRequestPoll
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestPoll-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestPoll-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestPoll$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetCreateRequestPoll in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestPoll-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestPoll-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetCreateRequestPoll
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestPoll$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetCreateRequestPoll
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestPoll-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestPoll-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetCreateRequestPoll and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestPoll$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestPoll-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestPoll-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestPoll$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetCreateRequestPoll
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestPoll-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestPoll-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestPoll$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestPoll-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestPoll-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestPoll$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateRequestPoll-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateRequestPoll-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateRequestPoll$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
