% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_create_response.R
\docType{class}
\name{TweetCreateResponse}
\alias{TweetCreateResponse}
\title{TweetCreateResponse}
\format{
An \code{R6Class} generator object
}
\description{
TweetCreateResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{TweetCreateResponseData} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetCreateResponse-new}{\code{TweetCreateResponse$new()}}
\item \href{#method-TweetCreateResponse-toJSON}{\code{TweetCreateResponse$toJSON()}}
\item \href{#method-TweetCreateResponse-fromJSON}{\code{TweetCreateResponse$fromJSON()}}
\item \href{#method-TweetCreateResponse-toJSONString}{\code{TweetCreateResponse$toJSONString()}}
\item \href{#method-TweetCreateResponse-fromJSONString}{\code{TweetCreateResponse$fromJSONString()}}
\item \href{#method-TweetCreateResponse-validateJSON}{\code{TweetCreateResponse$validateJSON()}}
\item \href{#method-TweetCreateResponse-toString}{\code{TweetCreateResponse$toString()}}
\item \href{#method-TweetCreateResponse-isValid}{\code{TweetCreateResponse$isValid()}}
\item \href{#method-TweetCreateResponse-getInvalidFields}{\code{TweetCreateResponse$getInvalidFields()}}
\item \href{#method-TweetCreateResponse-clone}{\code{TweetCreateResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetCreateResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetCreateResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetCreateResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetCreateResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetCreateResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetCreateResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
