% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_create_response_data.R
\docType{class}
\name{TweetCreateResponseData}
\alias{TweetCreateResponseData}
\title{TweetCreateResponseData}
\format{
An \code{R6Class} generator object
}
\description{
TweetCreateResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character}

\item{\code{text}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetCreateResponseData-new}{\code{TweetCreateResponseData$new()}}
\item \href{#method-TweetCreateResponseData-toJSON}{\code{TweetCreateResponseData$toJSON()}}
\item \href{#method-TweetCreateResponseData-fromJSON}{\code{TweetCreateResponseData$fromJSON()}}
\item \href{#method-TweetCreateResponseData-toJSONString}{\code{TweetCreateResponseData$toJSONString()}}
\item \href{#method-TweetCreateResponseData-fromJSONString}{\code{TweetCreateResponseData$fromJSONString()}}
\item \href{#method-TweetCreateResponseData-validateJSON}{\code{TweetCreateResponseData$validateJSON()}}
\item \href{#method-TweetCreateResponseData-toString}{\code{TweetCreateResponseData$toString()}}
\item \href{#method-TweetCreateResponseData-isValid}{\code{TweetCreateResponseData$isValid()}}
\item \href{#method-TweetCreateResponseData-getInvalidFields}{\code{TweetCreateResponseData$getInvalidFields()}}
\item \href{#method-TweetCreateResponseData-clone}{\code{TweetCreateResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetCreateResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponseData$new(id, text, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Unique identifier of this Tweet. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.}

\item{\code{text}}{The content of the Tweet.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetCreateResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetCreateResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetCreateResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetCreateResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetCreateResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetCreateResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetCreateResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetCreateResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
