% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_delete_compliance_schema.R
\docType{class}
\name{TweetDeleteComplianceSchema}
\alias{TweetDeleteComplianceSchema}
\title{TweetDeleteComplianceSchema}
\format{
An \code{R6Class} generator object
}
\description{
TweetDeleteComplianceSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{delete}}{\link{TweetComplianceSchema}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetDeleteComplianceSchema-new}{\code{TweetDeleteComplianceSchema$new()}}
\item \href{#method-TweetDeleteComplianceSchema-toJSON}{\code{TweetDeleteComplianceSchema$toJSON()}}
\item \href{#method-TweetDeleteComplianceSchema-fromJSON}{\code{TweetDeleteComplianceSchema$fromJSON()}}
\item \href{#method-TweetDeleteComplianceSchema-toJSONString}{\code{TweetDeleteComplianceSchema$toJSONString()}}
\item \href{#method-TweetDeleteComplianceSchema-fromJSONString}{\code{TweetDeleteComplianceSchema$fromJSONString()}}
\item \href{#method-TweetDeleteComplianceSchema-validateJSON}{\code{TweetDeleteComplianceSchema$validateJSON()}}
\item \href{#method-TweetDeleteComplianceSchema-toString}{\code{TweetDeleteComplianceSchema$toString()}}
\item \href{#method-TweetDeleteComplianceSchema-isValid}{\code{TweetDeleteComplianceSchema$isValid()}}
\item \href{#method-TweetDeleteComplianceSchema-getInvalidFields}{\code{TweetDeleteComplianceSchema$getInvalidFields()}}
\item \href{#method-TweetDeleteComplianceSchema-clone}{\code{TweetDeleteComplianceSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteComplianceSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteComplianceSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetDeleteComplianceSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteComplianceSchema$new(delete, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delete}}{delete}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteComplianceSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteComplianceSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteComplianceSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetDeleteComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteComplianceSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteComplianceSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetDeleteComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteComplianceSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetDeleteComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteComplianceSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteComplianceSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteComplianceSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetDeleteComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteComplianceSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteComplianceSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetDeleteComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteComplianceSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetDeleteComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteComplianceSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteComplianceSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetDeleteComplianceSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteComplianceSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteComplianceSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteComplianceSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteComplianceSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetDeleteComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteComplianceSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteComplianceSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteComplianceSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteComplianceSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteComplianceSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteComplianceSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteComplianceSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteComplianceSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteComplianceSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
