% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_delete_response.R
\docType{class}
\name{TweetDeleteResponse}
\alias{TweetDeleteResponse}
\title{TweetDeleteResponse}
\format{
An \code{R6Class} generator object
}
\description{
TweetDeleteResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{TweetDeleteResponseData} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetDeleteResponse-new}{\code{TweetDeleteResponse$new()}}
\item \href{#method-TweetDeleteResponse-toJSON}{\code{TweetDeleteResponse$toJSON()}}
\item \href{#method-TweetDeleteResponse-fromJSON}{\code{TweetDeleteResponse$fromJSON()}}
\item \href{#method-TweetDeleteResponse-toJSONString}{\code{TweetDeleteResponse$toJSONString()}}
\item \href{#method-TweetDeleteResponse-fromJSONString}{\code{TweetDeleteResponse$fromJSONString()}}
\item \href{#method-TweetDeleteResponse-validateJSON}{\code{TweetDeleteResponse$validateJSON()}}
\item \href{#method-TweetDeleteResponse-toString}{\code{TweetDeleteResponse$toString()}}
\item \href{#method-TweetDeleteResponse-isValid}{\code{TweetDeleteResponse$isValid()}}
\item \href{#method-TweetDeleteResponse-getInvalidFields}{\code{TweetDeleteResponse$getInvalidFields()}}
\item \href{#method-TweetDeleteResponse-clone}{\code{TweetDeleteResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetDeleteResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetDeleteResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetDeleteResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetDeleteResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetDeleteResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetDeleteResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetDeleteResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetDeleteResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetDeleteResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
