% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_delete_response_data.R
\docType{class}
\name{TweetDeleteResponseData}
\alias{TweetDeleteResponseData}
\title{TweetDeleteResponseData}
\format{
An \code{R6Class} generator object
}
\description{
TweetDeleteResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{deleted}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetDeleteResponseData-new}{\code{TweetDeleteResponseData$new()}}
\item \href{#method-TweetDeleteResponseData-toJSON}{\code{TweetDeleteResponseData$toJSON()}}
\item \href{#method-TweetDeleteResponseData-fromJSON}{\code{TweetDeleteResponseData$fromJSON()}}
\item \href{#method-TweetDeleteResponseData-toJSONString}{\code{TweetDeleteResponseData$toJSONString()}}
\item \href{#method-TweetDeleteResponseData-fromJSONString}{\code{TweetDeleteResponseData$fromJSONString()}}
\item \href{#method-TweetDeleteResponseData-validateJSON}{\code{TweetDeleteResponseData$validateJSON()}}
\item \href{#method-TweetDeleteResponseData-toString}{\code{TweetDeleteResponseData$toString()}}
\item \href{#method-TweetDeleteResponseData-isValid}{\code{TweetDeleteResponseData$isValid()}}
\item \href{#method-TweetDeleteResponseData-getInvalidFields}{\code{TweetDeleteResponseData$getInvalidFields()}}
\item \href{#method-TweetDeleteResponseData-clone}{\code{TweetDeleteResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetDeleteResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponseData$new(deleted, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deleted}}{deleted}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetDeleteResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetDeleteResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetDeleteResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetDeleteResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetDeleteResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetDeleteResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetDeleteResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetDeleteResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDeleteResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDeleteResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDeleteResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
