% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_drop_compliance_schema.R
\docType{class}
\name{TweetDropComplianceSchema}
\alias{TweetDropComplianceSchema}
\title{TweetDropComplianceSchema}
\format{
An \code{R6Class} generator object
}
\description{
TweetDropComplianceSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{drop}}{\link{TweetComplianceSchema}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetDropComplianceSchema-new}{\code{TweetDropComplianceSchema$new()}}
\item \href{#method-TweetDropComplianceSchema-toJSON}{\code{TweetDropComplianceSchema$toJSON()}}
\item \href{#method-TweetDropComplianceSchema-fromJSON}{\code{TweetDropComplianceSchema$fromJSON()}}
\item \href{#method-TweetDropComplianceSchema-toJSONString}{\code{TweetDropComplianceSchema$toJSONString()}}
\item \href{#method-TweetDropComplianceSchema-fromJSONString}{\code{TweetDropComplianceSchema$fromJSONString()}}
\item \href{#method-TweetDropComplianceSchema-validateJSON}{\code{TweetDropComplianceSchema$validateJSON()}}
\item \href{#method-TweetDropComplianceSchema-toString}{\code{TweetDropComplianceSchema$toString()}}
\item \href{#method-TweetDropComplianceSchema-isValid}{\code{TweetDropComplianceSchema$isValid()}}
\item \href{#method-TweetDropComplianceSchema-getInvalidFields}{\code{TweetDropComplianceSchema$getInvalidFields()}}
\item \href{#method-TweetDropComplianceSchema-clone}{\code{TweetDropComplianceSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDropComplianceSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDropComplianceSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetDropComplianceSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDropComplianceSchema$new(drop, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{drop}}{drop}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDropComplianceSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDropComplianceSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDropComplianceSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetDropComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDropComplianceSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDropComplianceSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetDropComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDropComplianceSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetDropComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDropComplianceSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDropComplianceSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDropComplianceSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetDropComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDropComplianceSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDropComplianceSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetDropComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDropComplianceSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetDropComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDropComplianceSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDropComplianceSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetDropComplianceSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDropComplianceSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDropComplianceSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDropComplianceSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDropComplianceSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetDropComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDropComplianceSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDropComplianceSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDropComplianceSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDropComplianceSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDropComplianceSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDropComplianceSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetDropComplianceSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetDropComplianceSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetDropComplianceSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
