% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_geo.R
\docType{class}
\name{TweetGeo}
\alias{TweetGeo}
\title{TweetGeo}
\format{
An \code{R6Class} generator object
}
\description{
TweetGeo Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{coordinates}}{\link{Point} [optional]}

\item{\code{place_id}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetGeo-new}{\code{TweetGeo$new()}}
\item \href{#method-TweetGeo-toJSON}{\code{TweetGeo$toJSON()}}
\item \href{#method-TweetGeo-fromJSON}{\code{TweetGeo$fromJSON()}}
\item \href{#method-TweetGeo-toJSONString}{\code{TweetGeo$toJSONString()}}
\item \href{#method-TweetGeo-fromJSONString}{\code{TweetGeo$fromJSONString()}}
\item \href{#method-TweetGeo-validateJSON}{\code{TweetGeo$validateJSON()}}
\item \href{#method-TweetGeo-toString}{\code{TweetGeo$toString()}}
\item \href{#method-TweetGeo-isValid}{\code{TweetGeo$isValid()}}
\item \href{#method-TweetGeo-getInvalidFields}{\code{TweetGeo$getInvalidFields()}}
\item \href{#method-TweetGeo-clone}{\code{TweetGeo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetGeo-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetGeo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetGeo class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetGeo$new(coordinates = NULL, place_id = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coordinates}}{coordinates}

\item{\code{place_id}}{The identifier for this place.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetGeo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetGeo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetGeo$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetGeo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetGeo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetGeo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetGeo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetGeo$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetGeo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetGeo-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetGeo-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetGeo$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetGeo in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetGeo-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetGeo-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetGeo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetGeo$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetGeo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetGeo-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetGeo-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetGeo and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetGeo$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetGeo-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetGeo-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetGeo$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetGeo
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetGeo-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetGeo-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetGeo$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetGeo-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetGeo-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetGeo$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetGeo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetGeo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetGeo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
