% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_hide_request.R
\docType{class}
\name{TweetHideRequest}
\alias{TweetHideRequest}
\title{TweetHideRequest}
\format{
An \code{R6Class} generator object
}
\description{
TweetHideRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{hidden}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetHideRequest-new}{\code{TweetHideRequest$new()}}
\item \href{#method-TweetHideRequest-toJSON}{\code{TweetHideRequest$toJSON()}}
\item \href{#method-TweetHideRequest-fromJSON}{\code{TweetHideRequest$fromJSON()}}
\item \href{#method-TweetHideRequest-toJSONString}{\code{TweetHideRequest$toJSONString()}}
\item \href{#method-TweetHideRequest-fromJSONString}{\code{TweetHideRequest$fromJSONString()}}
\item \href{#method-TweetHideRequest-validateJSON}{\code{TweetHideRequest$validateJSON()}}
\item \href{#method-TweetHideRequest-toString}{\code{TweetHideRequest$toString()}}
\item \href{#method-TweetHideRequest-isValid}{\code{TweetHideRequest$isValid()}}
\item \href{#method-TweetHideRequest-getInvalidFields}{\code{TweetHideRequest$getInvalidFields()}}
\item \href{#method-TweetHideRequest-clone}{\code{TweetHideRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetHideRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideRequest$new(hidden, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hidden}}{hidden}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetHideRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetHideRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetHideRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetHideRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetHideRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetHideRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetHideRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetHideRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
