% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_hide_response.R
\docType{class}
\name{TweetHideResponse}
\alias{TweetHideResponse}
\title{TweetHideResponse}
\format{
An \code{R6Class} generator object
}
\description{
TweetHideResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{TweetHideResponseData} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetHideResponse-new}{\code{TweetHideResponse$new()}}
\item \href{#method-TweetHideResponse-toJSON}{\code{TweetHideResponse$toJSON()}}
\item \href{#method-TweetHideResponse-fromJSON}{\code{TweetHideResponse$fromJSON()}}
\item \href{#method-TweetHideResponse-toJSONString}{\code{TweetHideResponse$toJSONString()}}
\item \href{#method-TweetHideResponse-fromJSONString}{\code{TweetHideResponse$fromJSONString()}}
\item \href{#method-TweetHideResponse-validateJSON}{\code{TweetHideResponse$validateJSON()}}
\item \href{#method-TweetHideResponse-toString}{\code{TweetHideResponse$toString()}}
\item \href{#method-TweetHideResponse-isValid}{\code{TweetHideResponse$isValid()}}
\item \href{#method-TweetHideResponse-getInvalidFields}{\code{TweetHideResponse$getInvalidFields()}}
\item \href{#method-TweetHideResponse-clone}{\code{TweetHideResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetHideResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponse$new(data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetHideResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetHideResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetHideResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetHideResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetHideResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetHideResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetHideResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetHideResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
