% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_hide_response_data.R
\docType{class}
\name{TweetHideResponseData}
\alias{TweetHideResponseData}
\title{TweetHideResponseData}
\format{
An \code{R6Class} generator object
}
\description{
TweetHideResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{hidden}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetHideResponseData-new}{\code{TweetHideResponseData$new()}}
\item \href{#method-TweetHideResponseData-toJSON}{\code{TweetHideResponseData$toJSON()}}
\item \href{#method-TweetHideResponseData-fromJSON}{\code{TweetHideResponseData$fromJSON()}}
\item \href{#method-TweetHideResponseData-toJSONString}{\code{TweetHideResponseData$toJSONString()}}
\item \href{#method-TweetHideResponseData-fromJSONString}{\code{TweetHideResponseData$fromJSONString()}}
\item \href{#method-TweetHideResponseData-validateJSON}{\code{TweetHideResponseData$validateJSON()}}
\item \href{#method-TweetHideResponseData-toString}{\code{TweetHideResponseData$toString()}}
\item \href{#method-TweetHideResponseData-isValid}{\code{TweetHideResponseData$isValid()}}
\item \href{#method-TweetHideResponseData-getInvalidFields}{\code{TweetHideResponseData$getInvalidFields()}}
\item \href{#method-TweetHideResponseData-clone}{\code{TweetHideResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetHideResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponseData$new(hidden = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hidden}}{hidden}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetHideResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetHideResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetHideResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetHideResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetHideResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetHideResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetHideResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetHideResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetHideResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetHideResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetHideResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
