% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_non_public_metrics.R
\docType{class}
\name{TweetNonPublicMetrics}
\alias{TweetNonPublicMetrics}
\title{TweetNonPublicMetrics}
\format{
An \code{R6Class} generator object
}
\description{
TweetNonPublicMetrics Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{impression_count}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetNonPublicMetrics-new}{\code{TweetNonPublicMetrics$new()}}
\item \href{#method-TweetNonPublicMetrics-toJSON}{\code{TweetNonPublicMetrics$toJSON()}}
\item \href{#method-TweetNonPublicMetrics-fromJSON}{\code{TweetNonPublicMetrics$fromJSON()}}
\item \href{#method-TweetNonPublicMetrics-toJSONString}{\code{TweetNonPublicMetrics$toJSONString()}}
\item \href{#method-TweetNonPublicMetrics-fromJSONString}{\code{TweetNonPublicMetrics$fromJSONString()}}
\item \href{#method-TweetNonPublicMetrics-validateJSON}{\code{TweetNonPublicMetrics$validateJSON()}}
\item \href{#method-TweetNonPublicMetrics-toString}{\code{TweetNonPublicMetrics$toString()}}
\item \href{#method-TweetNonPublicMetrics-isValid}{\code{TweetNonPublicMetrics$isValid()}}
\item \href{#method-TweetNonPublicMetrics-getInvalidFields}{\code{TweetNonPublicMetrics$getInvalidFields()}}
\item \href{#method-TweetNonPublicMetrics-clone}{\code{TweetNonPublicMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetNonPublicMetrics-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetNonPublicMetrics-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetNonPublicMetrics class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetNonPublicMetrics$new(impression_count = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{impression_count}}{Number of times this Tweet has been viewed.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetNonPublicMetrics-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetNonPublicMetrics-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetNonPublicMetrics$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetNonPublicMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetNonPublicMetrics-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetNonPublicMetrics-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetNonPublicMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetNonPublicMetrics$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetNonPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetNonPublicMetrics-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetNonPublicMetrics-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetNonPublicMetrics$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetNonPublicMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetNonPublicMetrics-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetNonPublicMetrics-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetNonPublicMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetNonPublicMetrics$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetNonPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetNonPublicMetrics-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetNonPublicMetrics-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetNonPublicMetrics and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetNonPublicMetrics$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetNonPublicMetrics-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetNonPublicMetrics-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetNonPublicMetrics$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetNonPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetNonPublicMetrics-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetNonPublicMetrics-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetNonPublicMetrics$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetNonPublicMetrics-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetNonPublicMetrics-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetNonPublicMetrics$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetNonPublicMetrics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetNonPublicMetrics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetNonPublicMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
