% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_organic_metrics.R
\docType{class}
\name{TweetOrganicMetrics}
\alias{TweetOrganicMetrics}
\title{TweetOrganicMetrics}
\format{
An \code{R6Class} generator object
}
\description{
TweetOrganicMetrics Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{impression_count}}{integer}

\item{\code{like_count}}{integer}

\item{\code{reply_count}}{integer}

\item{\code{retweet_count}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetOrganicMetrics-new}{\code{TweetOrganicMetrics$new()}}
\item \href{#method-TweetOrganicMetrics-toJSON}{\code{TweetOrganicMetrics$toJSON()}}
\item \href{#method-TweetOrganicMetrics-fromJSON}{\code{TweetOrganicMetrics$fromJSON()}}
\item \href{#method-TweetOrganicMetrics-toJSONString}{\code{TweetOrganicMetrics$toJSONString()}}
\item \href{#method-TweetOrganicMetrics-fromJSONString}{\code{TweetOrganicMetrics$fromJSONString()}}
\item \href{#method-TweetOrganicMetrics-validateJSON}{\code{TweetOrganicMetrics$validateJSON()}}
\item \href{#method-TweetOrganicMetrics-toString}{\code{TweetOrganicMetrics$toString()}}
\item \href{#method-TweetOrganicMetrics-isValid}{\code{TweetOrganicMetrics$isValid()}}
\item \href{#method-TweetOrganicMetrics-getInvalidFields}{\code{TweetOrganicMetrics$getInvalidFields()}}
\item \href{#method-TweetOrganicMetrics-clone}{\code{TweetOrganicMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetOrganicMetrics-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetOrganicMetrics-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetOrganicMetrics class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetOrganicMetrics$new(
  impression_count,
  like_count,
  reply_count,
  retweet_count,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{impression_count}}{Number of times this Tweet has been viewed.}

\item{\code{like_count}}{Number of times this Tweet has been liked.}

\item{\code{reply_count}}{Number of times this Tweet has been replied to.}

\item{\code{retweet_count}}{Number of times this Tweet has been Retweeted.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetOrganicMetrics-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetOrganicMetrics-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetOrganicMetrics$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetOrganicMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetOrganicMetrics-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetOrganicMetrics-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetOrganicMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetOrganicMetrics$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetOrganicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetOrganicMetrics-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetOrganicMetrics-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetOrganicMetrics$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetOrganicMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetOrganicMetrics-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetOrganicMetrics-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetOrganicMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetOrganicMetrics$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetOrganicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetOrganicMetrics-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetOrganicMetrics-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetOrganicMetrics and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetOrganicMetrics$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetOrganicMetrics-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetOrganicMetrics-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetOrganicMetrics$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetOrganicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetOrganicMetrics-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetOrganicMetrics-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetOrganicMetrics$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetOrganicMetrics-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetOrganicMetrics-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetOrganicMetrics$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetOrganicMetrics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetOrganicMetrics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetOrganicMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
