% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_promoted_metrics.R
\docType{class}
\name{TweetPromotedMetrics}
\alias{TweetPromotedMetrics}
\title{TweetPromotedMetrics}
\format{
An \code{R6Class} generator object
}
\description{
TweetPromotedMetrics Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{impression_count}}{integer [optional]}

\item{\code{like_count}}{integer [optional]}

\item{\code{reply_count}}{integer [optional]}

\item{\code{retweet_count}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetPromotedMetrics-new}{\code{TweetPromotedMetrics$new()}}
\item \href{#method-TweetPromotedMetrics-toJSON}{\code{TweetPromotedMetrics$toJSON()}}
\item \href{#method-TweetPromotedMetrics-fromJSON}{\code{TweetPromotedMetrics$fromJSON()}}
\item \href{#method-TweetPromotedMetrics-toJSONString}{\code{TweetPromotedMetrics$toJSONString()}}
\item \href{#method-TweetPromotedMetrics-fromJSONString}{\code{TweetPromotedMetrics$fromJSONString()}}
\item \href{#method-TweetPromotedMetrics-validateJSON}{\code{TweetPromotedMetrics$validateJSON()}}
\item \href{#method-TweetPromotedMetrics-toString}{\code{TweetPromotedMetrics$toString()}}
\item \href{#method-TweetPromotedMetrics-isValid}{\code{TweetPromotedMetrics$isValid()}}
\item \href{#method-TweetPromotedMetrics-getInvalidFields}{\code{TweetPromotedMetrics$getInvalidFields()}}
\item \href{#method-TweetPromotedMetrics-clone}{\code{TweetPromotedMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPromotedMetrics-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPromotedMetrics-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetPromotedMetrics class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPromotedMetrics$new(
  impression_count = NULL,
  like_count = NULL,
  reply_count = NULL,
  retweet_count = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{impression_count}}{Number of times this Tweet has been viewed.}

\item{\code{like_count}}{Number of times this Tweet has been liked.}

\item{\code{reply_count}}{Number of times this Tweet has been replied to.}

\item{\code{retweet_count}}{Number of times this Tweet has been Retweeted.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPromotedMetrics-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPromotedMetrics-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPromotedMetrics$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetPromotedMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPromotedMetrics-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPromotedMetrics-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetPromotedMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPromotedMetrics$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetPromotedMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPromotedMetrics-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPromotedMetrics-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPromotedMetrics$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetPromotedMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPromotedMetrics-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPromotedMetrics-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetPromotedMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPromotedMetrics$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetPromotedMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPromotedMetrics-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPromotedMetrics-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetPromotedMetrics and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPromotedMetrics$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPromotedMetrics-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPromotedMetrics-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPromotedMetrics$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetPromotedMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPromotedMetrics-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPromotedMetrics-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPromotedMetrics$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPromotedMetrics-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPromotedMetrics-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPromotedMetrics$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPromotedMetrics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPromotedMetrics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPromotedMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
