% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_public_metrics.R
\docType{class}
\name{TweetPublicMetrics}
\alias{TweetPublicMetrics}
\title{TweetPublicMetrics}
\format{
An \code{R6Class} generator object
}
\description{
TweetPublicMetrics Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{like_count}}{integer}

\item{\code{quote_count}}{integer [optional]}

\item{\code{reply_count}}{integer}

\item{\code{retweet_count}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetPublicMetrics-new}{\code{TweetPublicMetrics$new()}}
\item \href{#method-TweetPublicMetrics-toJSON}{\code{TweetPublicMetrics$toJSON()}}
\item \href{#method-TweetPublicMetrics-fromJSON}{\code{TweetPublicMetrics$fromJSON()}}
\item \href{#method-TweetPublicMetrics-toJSONString}{\code{TweetPublicMetrics$toJSONString()}}
\item \href{#method-TweetPublicMetrics-fromJSONString}{\code{TweetPublicMetrics$fromJSONString()}}
\item \href{#method-TweetPublicMetrics-validateJSON}{\code{TweetPublicMetrics$validateJSON()}}
\item \href{#method-TweetPublicMetrics-toString}{\code{TweetPublicMetrics$toString()}}
\item \href{#method-TweetPublicMetrics-isValid}{\code{TweetPublicMetrics$isValid()}}
\item \href{#method-TweetPublicMetrics-getInvalidFields}{\code{TweetPublicMetrics$getInvalidFields()}}
\item \href{#method-TweetPublicMetrics-clone}{\code{TweetPublicMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPublicMetrics-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPublicMetrics-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetPublicMetrics class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPublicMetrics$new(
  like_count,
  reply_count,
  retweet_count,
  quote_count = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{like_count}}{Number of times this Tweet has been liked.}

\item{\code{reply_count}}{Number of times this Tweet has been replied to.}

\item{\code{retweet_count}}{Number of times this Tweet has been Retweeted.}

\item{\code{quote_count}}{Number of times this Tweet has been quoted.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPublicMetrics-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPublicMetrics-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPublicMetrics$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetPublicMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPublicMetrics-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPublicMetrics-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetPublicMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPublicMetrics$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPublicMetrics-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPublicMetrics-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPublicMetrics$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetPublicMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPublicMetrics-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPublicMetrics-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetPublicMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPublicMetrics$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPublicMetrics-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPublicMetrics-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetPublicMetrics and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPublicMetrics$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPublicMetrics-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPublicMetrics-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPublicMetrics$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPublicMetrics-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPublicMetrics-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPublicMetrics$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPublicMetrics-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPublicMetrics-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPublicMetrics$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetPublicMetrics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetPublicMetrics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetPublicMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
