% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_referenced_tweets_inner.R
\docType{class}
\name{TweetReferencedTweetsInner}
\alias{TweetReferencedTweetsInner}
\title{TweetReferencedTweetsInner}
\format{
An \code{R6Class} generator object
}
\description{
TweetReferencedTweetsInner Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character}

\item{\code{type}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetReferencedTweetsInner-new}{\code{TweetReferencedTweetsInner$new()}}
\item \href{#method-TweetReferencedTweetsInner-toJSON}{\code{TweetReferencedTweetsInner$toJSON()}}
\item \href{#method-TweetReferencedTweetsInner-fromJSON}{\code{TweetReferencedTweetsInner$fromJSON()}}
\item \href{#method-TweetReferencedTweetsInner-toJSONString}{\code{TweetReferencedTweetsInner$toJSONString()}}
\item \href{#method-TweetReferencedTweetsInner-fromJSONString}{\code{TweetReferencedTweetsInner$fromJSONString()}}
\item \href{#method-TweetReferencedTweetsInner-validateJSON}{\code{TweetReferencedTweetsInner$validateJSON()}}
\item \href{#method-TweetReferencedTweetsInner-toString}{\code{TweetReferencedTweetsInner$toString()}}
\item \href{#method-TweetReferencedTweetsInner-isValid}{\code{TweetReferencedTweetsInner$isValid()}}
\item \href{#method-TweetReferencedTweetsInner-getInvalidFields}{\code{TweetReferencedTweetsInner$getInvalidFields()}}
\item \href{#method-TweetReferencedTweetsInner-clone}{\code{TweetReferencedTweetsInner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetReferencedTweetsInner-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetReferencedTweetsInner-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetReferencedTweetsInner class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetReferencedTweetsInner$new(id, type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Unique identifier of this Tweet. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.}

\item{\code{type}}{type}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetReferencedTweetsInner-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetReferencedTweetsInner-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetReferencedTweetsInner$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetReferencedTweetsInner in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetReferencedTweetsInner-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetReferencedTweetsInner-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetReferencedTweetsInner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetReferencedTweetsInner$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetReferencedTweetsInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetReferencedTweetsInner-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetReferencedTweetsInner-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetReferencedTweetsInner$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetReferencedTweetsInner in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetReferencedTweetsInner-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetReferencedTweetsInner-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetReferencedTweetsInner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetReferencedTweetsInner$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetReferencedTweetsInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetReferencedTweetsInner-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetReferencedTweetsInner-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetReferencedTweetsInner and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetReferencedTweetsInner$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetReferencedTweetsInner-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetReferencedTweetsInner-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetReferencedTweetsInner$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetReferencedTweetsInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetReferencedTweetsInner-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetReferencedTweetsInner-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetReferencedTweetsInner$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetReferencedTweetsInner-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetReferencedTweetsInner-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetReferencedTweetsInner$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetReferencedTweetsInner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetReferencedTweetsInner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetReferencedTweetsInner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
