% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_takedown_compliance_schema.R
\docType{class}
\name{TweetTakedownComplianceSchema}
\alias{TweetTakedownComplianceSchema}
\title{TweetTakedownComplianceSchema}
\format{
An \code{R6Class} generator object
}
\description{
TweetTakedownComplianceSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{event_at}}{character}

\item{\code{quote_tweet_id}}{character [optional]}

\item{\code{tweet}}{\link{TweetComplianceSchemaTweet}}

\item{\code{withheld_in_countries}}{list( character )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetTakedownComplianceSchema-new}{\code{TweetTakedownComplianceSchema$new()}}
\item \href{#method-TweetTakedownComplianceSchema-toJSON}{\code{TweetTakedownComplianceSchema$toJSON()}}
\item \href{#method-TweetTakedownComplianceSchema-fromJSON}{\code{TweetTakedownComplianceSchema$fromJSON()}}
\item \href{#method-TweetTakedownComplianceSchema-toJSONString}{\code{TweetTakedownComplianceSchema$toJSONString()}}
\item \href{#method-TweetTakedownComplianceSchema-fromJSONString}{\code{TweetTakedownComplianceSchema$fromJSONString()}}
\item \href{#method-TweetTakedownComplianceSchema-validateJSON}{\code{TweetTakedownComplianceSchema$validateJSON()}}
\item \href{#method-TweetTakedownComplianceSchema-toString}{\code{TweetTakedownComplianceSchema$toString()}}
\item \href{#method-TweetTakedownComplianceSchema-isValid}{\code{TweetTakedownComplianceSchema$isValid()}}
\item \href{#method-TweetTakedownComplianceSchema-getInvalidFields}{\code{TweetTakedownComplianceSchema$getInvalidFields()}}
\item \href{#method-TweetTakedownComplianceSchema-clone}{\code{TweetTakedownComplianceSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetTakedownComplianceSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetTakedownComplianceSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetTakedownComplianceSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetTakedownComplianceSchema$new(
  event_at,
  tweet,
  withheld_in_countries,
  quote_tweet_id = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_at}}{Event time.}

\item{\code{tweet}}{tweet}

\item{\code{withheld_in_countries}}{withheld_in_countries}

\item{\code{quote_tweet_id}}{Unique identifier of this Tweet. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetTakedownComplianceSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetTakedownComplianceSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetTakedownComplianceSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetTakedownComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetTakedownComplianceSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetTakedownComplianceSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetTakedownComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetTakedownComplianceSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetTakedownComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetTakedownComplianceSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetTakedownComplianceSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetTakedownComplianceSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetTakedownComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetTakedownComplianceSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetTakedownComplianceSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetTakedownComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetTakedownComplianceSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetTakedownComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetTakedownComplianceSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetTakedownComplianceSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetTakedownComplianceSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetTakedownComplianceSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetTakedownComplianceSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetTakedownComplianceSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetTakedownComplianceSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetTakedownComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetTakedownComplianceSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetTakedownComplianceSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetTakedownComplianceSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetTakedownComplianceSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetTakedownComplianceSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetTakedownComplianceSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetTakedownComplianceSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetTakedownComplianceSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetTakedownComplianceSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
