% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_undrop_compliance_schema.R
\docType{class}
\name{TweetUndropComplianceSchema}
\alias{TweetUndropComplianceSchema}
\title{TweetUndropComplianceSchema}
\format{
An \code{R6Class} generator object
}
\description{
TweetUndropComplianceSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{undrop}}{\link{TweetComplianceSchema}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TweetUndropComplianceSchema-new}{\code{TweetUndropComplianceSchema$new()}}
\item \href{#method-TweetUndropComplianceSchema-toJSON}{\code{TweetUndropComplianceSchema$toJSON()}}
\item \href{#method-TweetUndropComplianceSchema-fromJSON}{\code{TweetUndropComplianceSchema$fromJSON()}}
\item \href{#method-TweetUndropComplianceSchema-toJSONString}{\code{TweetUndropComplianceSchema$toJSONString()}}
\item \href{#method-TweetUndropComplianceSchema-fromJSONString}{\code{TweetUndropComplianceSchema$fromJSONString()}}
\item \href{#method-TweetUndropComplianceSchema-validateJSON}{\code{TweetUndropComplianceSchema$validateJSON()}}
\item \href{#method-TweetUndropComplianceSchema-toString}{\code{TweetUndropComplianceSchema$toString()}}
\item \href{#method-TweetUndropComplianceSchema-isValid}{\code{TweetUndropComplianceSchema$isValid()}}
\item \href{#method-TweetUndropComplianceSchema-getInvalidFields}{\code{TweetUndropComplianceSchema$getInvalidFields()}}
\item \href{#method-TweetUndropComplianceSchema-clone}{\code{TweetUndropComplianceSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetUndropComplianceSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-TweetUndropComplianceSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new TweetUndropComplianceSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetUndropComplianceSchema$new(undrop, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{undrop}}{undrop}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetUndropComplianceSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetUndropComplianceSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetUndropComplianceSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetUndropComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetUndropComplianceSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetUndropComplianceSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of TweetUndropComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetUndropComplianceSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetUndropComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetUndropComplianceSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetUndropComplianceSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetUndropComplianceSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TweetUndropComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetUndropComplianceSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetUndropComplianceSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of TweetUndropComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetUndropComplianceSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of TweetUndropComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetUndropComplianceSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TweetUndropComplianceSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to TweetUndropComplianceSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetUndropComplianceSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetUndropComplianceSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-TweetUndropComplianceSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetUndropComplianceSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of TweetUndropComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetUndropComplianceSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-TweetUndropComplianceSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetUndropComplianceSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetUndropComplianceSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-TweetUndropComplianceSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetUndropComplianceSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TweetUndropComplianceSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TweetUndropComplianceSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetUndropComplianceSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
