% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupported_authentication_problem.R
\docType{class}
\name{UnsupportedAuthenticationProblem}
\alias{UnsupportedAuthenticationProblem}
\title{UnsupportedAuthenticationProblem}
\format{
An \code{R6Class} generator object
}
\description{
UnsupportedAuthenticationProblem Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Super class}{
\code{\link[twitter:Problem]{twitter::Problem}} -> \code{UnsupportedAuthenticationProblem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detail}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character}

\item{\code{type}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UnsupportedAuthenticationProblem-new}{\code{UnsupportedAuthenticationProblem$new()}}
\item \href{#method-UnsupportedAuthenticationProblem-toJSON}{\code{UnsupportedAuthenticationProblem$toJSON()}}
\item \href{#method-UnsupportedAuthenticationProblem-fromJSON}{\code{UnsupportedAuthenticationProblem$fromJSON()}}
\item \href{#method-UnsupportedAuthenticationProblem-toJSONString}{\code{UnsupportedAuthenticationProblem$toJSONString()}}
\item \href{#method-UnsupportedAuthenticationProblem-fromJSONString}{\code{UnsupportedAuthenticationProblem$fromJSONString()}}
\item \href{#method-UnsupportedAuthenticationProblem-validateJSON}{\code{UnsupportedAuthenticationProblem$validateJSON()}}
\item \href{#method-UnsupportedAuthenticationProblem-toString}{\code{UnsupportedAuthenticationProblem$toString()}}
\item \href{#method-UnsupportedAuthenticationProblem-isValid}{\code{UnsupportedAuthenticationProblem$isValid()}}
\item \href{#method-UnsupportedAuthenticationProblem-getInvalidFields}{\code{UnsupportedAuthenticationProblem$getInvalidFields()}}
\item \href{#method-UnsupportedAuthenticationProblem-clone}{\code{UnsupportedAuthenticationProblem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnsupportedAuthenticationProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-UnsupportedAuthenticationProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UnsupportedAuthenticationProblem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnsupportedAuthenticationProblem$new(
  title,
  type,
  detail = NULL,
  status = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{type}}{type}

\item{\code{detail}}{detail}

\item{\code{status}}{status}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnsupportedAuthenticationProblem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UnsupportedAuthenticationProblem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnsupportedAuthenticationProblem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UnsupportedAuthenticationProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnsupportedAuthenticationProblem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UnsupportedAuthenticationProblem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UnsupportedAuthenticationProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnsupportedAuthenticationProblem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UnsupportedAuthenticationProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnsupportedAuthenticationProblem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UnsupportedAuthenticationProblem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnsupportedAuthenticationProblem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UnsupportedAuthenticationProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnsupportedAuthenticationProblem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UnsupportedAuthenticationProblem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UnsupportedAuthenticationProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnsupportedAuthenticationProblem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UnsupportedAuthenticationProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnsupportedAuthenticationProblem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UnsupportedAuthenticationProblem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UnsupportedAuthenticationProblem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnsupportedAuthenticationProblem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnsupportedAuthenticationProblem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UnsupportedAuthenticationProblem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnsupportedAuthenticationProblem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UnsupportedAuthenticationProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnsupportedAuthenticationProblem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UnsupportedAuthenticationProblem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnsupportedAuthenticationProblem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnsupportedAuthenticationProblem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UnsupportedAuthenticationProblem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnsupportedAuthenticationProblem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnsupportedAuthenticationProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UnsupportedAuthenticationProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnsupportedAuthenticationProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
