% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_entity.R
\docType{class}
\name{UrlEntity}
\alias{UrlEntity}
\title{UrlEntity}
\format{
An \code{R6Class} generator object
}
\description{
UrlEntity Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{end}}{integer}

\item{\code{start}}{integer}

\item{\code{description}}{character [optional]}

\item{\code{display_url}}{character [optional]}

\item{\code{expanded_url}}{character [optional]}

\item{\code{images}}{list( \link{UrlImage} ) [optional]}

\item{\code{media_key}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character [optional]}

\item{\code{unwound_url}}{character [optional]}

\item{\code{url}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UrlEntity-new}{\code{UrlEntity$new()}}
\item \href{#method-UrlEntity-toJSON}{\code{UrlEntity$toJSON()}}
\item \href{#method-UrlEntity-fromJSON}{\code{UrlEntity$fromJSON()}}
\item \href{#method-UrlEntity-toJSONString}{\code{UrlEntity$toJSONString()}}
\item \href{#method-UrlEntity-fromJSONString}{\code{UrlEntity$fromJSONString()}}
\item \href{#method-UrlEntity-validateJSON}{\code{UrlEntity$validateJSON()}}
\item \href{#method-UrlEntity-toString}{\code{UrlEntity$toString()}}
\item \href{#method-UrlEntity-isValid}{\code{UrlEntity$isValid()}}
\item \href{#method-UrlEntity-getInvalidFields}{\code{UrlEntity$getInvalidFields()}}
\item \href{#method-UrlEntity-clone}{\code{UrlEntity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlEntity-new"></a>}}
\if{latex}{\out{\hypertarget{method-UrlEntity-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UrlEntity class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlEntity$new(
  end,
  start,
  url,
  description = NULL,
  display_url = NULL,
  expanded_url = NULL,
  images = NULL,
  media_key = NULL,
  status = NULL,
  title = NULL,
  unwound_url = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{end}}{Index (zero-based) at which position this entity ends.  The index is exclusive.}

\item{\code{start}}{Index (zero-based) at which position this entity starts.  The index is inclusive.}

\item{\code{url}}{A validly formatted URL.}

\item{\code{description}}{Description of the URL landing page.}

\item{\code{display_url}}{The URL as displayed in the Twitter client.}

\item{\code{expanded_url}}{A validly formatted URL.}

\item{\code{images}}{images}

\item{\code{media_key}}{The Media Key identifier for this attachment.}

\item{\code{status}}{HTTP Status Code.}

\item{\code{title}}{Title of the page the URL points to.}

\item{\code{unwound_url}}{Fully resolved url.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlEntity-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UrlEntity-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlEntity$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UrlEntity in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlEntity-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UrlEntity-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UrlEntity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlEntity$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UrlEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlEntity-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UrlEntity-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlEntity$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UrlEntity in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlEntity-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UrlEntity-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UrlEntity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlEntity$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UrlEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlEntity-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UrlEntity-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UrlEntity and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlEntity$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlEntity-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UrlEntity-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlEntity$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UrlEntity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlEntity-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UrlEntity-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlEntity$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlEntity-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UrlEntity-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlEntity$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlEntity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UrlEntity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlEntity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
