% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_fields.R
\docType{class}
\name{UrlFields}
\alias{UrlFields}
\title{UrlFields}
\format{
An \code{R6Class} generator object
}
\description{
UrlFields Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character [optional]}

\item{\code{display_url}}{character [optional]}

\item{\code{expanded_url}}{character [optional]}

\item{\code{images}}{list( \link{UrlImage} ) [optional]}

\item{\code{media_key}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character [optional]}

\item{\code{unwound_url}}{character [optional]}

\item{\code{url}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UrlFields-new}{\code{UrlFields$new()}}
\item \href{#method-UrlFields-toJSON}{\code{UrlFields$toJSON()}}
\item \href{#method-UrlFields-fromJSON}{\code{UrlFields$fromJSON()}}
\item \href{#method-UrlFields-toJSONString}{\code{UrlFields$toJSONString()}}
\item \href{#method-UrlFields-fromJSONString}{\code{UrlFields$fromJSONString()}}
\item \href{#method-UrlFields-validateJSON}{\code{UrlFields$validateJSON()}}
\item \href{#method-UrlFields-toString}{\code{UrlFields$toString()}}
\item \href{#method-UrlFields-isValid}{\code{UrlFields$isValid()}}
\item \href{#method-UrlFields-getInvalidFields}{\code{UrlFields$getInvalidFields()}}
\item \href{#method-UrlFields-clone}{\code{UrlFields$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlFields-new"></a>}}
\if{latex}{\out{\hypertarget{method-UrlFields-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UrlFields class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlFields$new(
  url,
  description = NULL,
  display_url = NULL,
  expanded_url = NULL,
  images = NULL,
  media_key = NULL,
  status = NULL,
  title = NULL,
  unwound_url = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{A validly formatted URL.}

\item{\code{description}}{Description of the URL landing page.}

\item{\code{display_url}}{The URL as displayed in the Twitter client.}

\item{\code{expanded_url}}{A validly formatted URL.}

\item{\code{images}}{images}

\item{\code{media_key}}{The Media Key identifier for this attachment.}

\item{\code{status}}{HTTP Status Code.}

\item{\code{title}}{Title of the page the URL points to.}

\item{\code{unwound_url}}{Fully resolved url.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlFields-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UrlFields-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlFields$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UrlFields in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlFields-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UrlFields-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UrlFields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlFields$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UrlFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlFields-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UrlFields-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlFields$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UrlFields in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlFields-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UrlFields-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UrlFields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlFields$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UrlFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlFields-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UrlFields-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UrlFields and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlFields$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlFields-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UrlFields-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlFields$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UrlFields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlFields-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UrlFields-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlFields$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlFields-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UrlFields-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlFields$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrlFields-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UrlFields-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UrlFields$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
