% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_cap_exceeded_problem.R
\docType{class}
\name{UsageCapExceededProblem}
\alias{UsageCapExceededProblem}
\title{UsageCapExceededProblem}
\format{
An \code{R6Class} generator object
}
\description{
UsageCapExceededProblem Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Super class}{
\code{\link[twitter:Problem]{twitter::Problem}} -> \code{UsageCapExceededProblem}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{detail}}{character [optional]}

\item{\code{status}}{integer [optional]}

\item{\code{title}}{character}

\item{\code{type}}{character}

\item{\code{period}}{character [optional]}

\item{\code{scope}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsageCapExceededProblem-new}{\code{UsageCapExceededProblem$new()}}
\item \href{#method-UsageCapExceededProblem-toJSON}{\code{UsageCapExceededProblem$toJSON()}}
\item \href{#method-UsageCapExceededProblem-fromJSON}{\code{UsageCapExceededProblem$fromJSON()}}
\item \href{#method-UsageCapExceededProblem-toJSONString}{\code{UsageCapExceededProblem$toJSONString()}}
\item \href{#method-UsageCapExceededProblem-fromJSONString}{\code{UsageCapExceededProblem$fromJSONString()}}
\item \href{#method-UsageCapExceededProblem-validateJSON}{\code{UsageCapExceededProblem$validateJSON()}}
\item \href{#method-UsageCapExceededProblem-toString}{\code{UsageCapExceededProblem$toString()}}
\item \href{#method-UsageCapExceededProblem-isValid}{\code{UsageCapExceededProblem$isValid()}}
\item \href{#method-UsageCapExceededProblem-getInvalidFields}{\code{UsageCapExceededProblem$getInvalidFields()}}
\item \href{#method-UsageCapExceededProblem-clone}{\code{UsageCapExceededProblem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UsageCapExceededProblem class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblem$new(
  title,
  type,
  detail = NULL,
  status = NULL,
  period = NULL,
  scope = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title}

\item{\code{type}}{type}

\item{\code{detail}}{detail}

\item{\code{status}}{status}

\item{\code{period}}{period}

\item{\code{scope}}{scope}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblem-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblem-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblem$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsageCapExceededProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblem-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblem-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UsageCapExceededProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblem$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsageCapExceededProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblem-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblem-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblem$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsageCapExceededProblem in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblem-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblem-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UsageCapExceededProblem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblem$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsageCapExceededProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblem-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblem-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UsageCapExceededProblem and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblem$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblem-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblem-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblem$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UsageCapExceededProblem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblem-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblem-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblem$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblem-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblem-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblem$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
