% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_cap_exceeded_problem_all_of.R
\docType{class}
\name{UsageCapExceededProblemAllOf}
\alias{UsageCapExceededProblemAllOf}
\title{UsageCapExceededProblemAllOf}
\format{
An \code{R6Class} generator object
}
\description{
UsageCapExceededProblemAllOf Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{period}}{character [optional]}

\item{\code{scope}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsageCapExceededProblemAllOf-new}{\code{UsageCapExceededProblemAllOf$new()}}
\item \href{#method-UsageCapExceededProblemAllOf-toJSON}{\code{UsageCapExceededProblemAllOf$toJSON()}}
\item \href{#method-UsageCapExceededProblemAllOf-fromJSON}{\code{UsageCapExceededProblemAllOf$fromJSON()}}
\item \href{#method-UsageCapExceededProblemAllOf-toJSONString}{\code{UsageCapExceededProblemAllOf$toJSONString()}}
\item \href{#method-UsageCapExceededProblemAllOf-fromJSONString}{\code{UsageCapExceededProblemAllOf$fromJSONString()}}
\item \href{#method-UsageCapExceededProblemAllOf-validateJSON}{\code{UsageCapExceededProblemAllOf$validateJSON()}}
\item \href{#method-UsageCapExceededProblemAllOf-toString}{\code{UsageCapExceededProblemAllOf$toString()}}
\item \href{#method-UsageCapExceededProblemAllOf-isValid}{\code{UsageCapExceededProblemAllOf$isValid()}}
\item \href{#method-UsageCapExceededProblemAllOf-getInvalidFields}{\code{UsageCapExceededProblemAllOf$getInvalidFields()}}
\item \href{#method-UsageCapExceededProblemAllOf-clone}{\code{UsageCapExceededProblemAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblemAllOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblemAllOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UsageCapExceededProblemAllOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblemAllOf$new(period = NULL, scope = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{period}}{period}

\item{\code{scope}}{scope}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblemAllOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblemAllOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblemAllOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsageCapExceededProblemAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblemAllOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblemAllOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UsageCapExceededProblemAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblemAllOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsageCapExceededProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblemAllOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblemAllOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblemAllOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsageCapExceededProblemAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblemAllOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblemAllOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UsageCapExceededProblemAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblemAllOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsageCapExceededProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblemAllOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblemAllOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UsageCapExceededProblemAllOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblemAllOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblemAllOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblemAllOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblemAllOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UsageCapExceededProblemAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblemAllOf-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblemAllOf-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblemAllOf$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblemAllOf-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblemAllOf-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblemAllOf$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsageCapExceededProblemAllOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsageCapExceededProblemAllOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsageCapExceededProblemAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
