% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_compliance_data.R
\docType{class}
\name{UserComplianceData}
\alias{UserComplianceData}
\title{UserComplianceData}
\format{
An \code{R6Class} generator object
}
\description{
UserComplianceData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of object types defined in the oneOf schema.
Initialize a new UserComplianceData.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserComplianceData-new}{\code{UserComplianceData$new()}}
\item \href{#method-UserComplianceData-fromJSON}{\code{UserComplianceData$fromJSON()}}
\item \href{#method-UserComplianceData-toJSONString}{\code{UserComplianceData$toJSONString()}}
\item \href{#method-UserComplianceData-toJSON}{\code{UserComplianceData$toJSON()}}
\item \href{#method-UserComplianceData-validateJSON}{\code{UserComplianceData$validateJSON()}}
\item \href{#method-UserComplianceData-toString}{\code{UserComplianceData$toString()}}
\item \href{#method-UserComplianceData-clone}{\code{UserComplianceData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceData-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserComplianceData.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceData$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "UserDeleteComplianceSchema", "UserProfileModificationComplianceSchema", "UserProtectComplianceSchema", "UserScrubGeoSchema", "UserSuspendComplianceSchema", "UserUndeleteComplianceSchema", "UserUnprotectComplianceSchema", "UserUnsuspendComplianceSchema", "UserWithheldComplianceSchema"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserComplianceData.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceData$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of UserComplianceData.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize UserComplianceData to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the UserComplianceData.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize UserComplianceData to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the UserComplianceData.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to UserComplianceData and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceData-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
