% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_compliance_schema.R
\docType{class}
\name{UserComplianceSchema}
\alias{UserComplianceSchema}
\title{UserComplianceSchema}
\format{
An \code{R6Class} generator object
}
\description{
UserComplianceSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{event_at}}{character}

\item{\code{user}}{\link{UserComplianceSchemaUser}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserComplianceSchema-new}{\code{UserComplianceSchema$new()}}
\item \href{#method-UserComplianceSchema-toJSON}{\code{UserComplianceSchema$toJSON()}}
\item \href{#method-UserComplianceSchema-fromJSON}{\code{UserComplianceSchema$fromJSON()}}
\item \href{#method-UserComplianceSchema-toJSONString}{\code{UserComplianceSchema$toJSONString()}}
\item \href{#method-UserComplianceSchema-fromJSONString}{\code{UserComplianceSchema$fromJSONString()}}
\item \href{#method-UserComplianceSchema-validateJSON}{\code{UserComplianceSchema$validateJSON()}}
\item \href{#method-UserComplianceSchema-toString}{\code{UserComplianceSchema$toString()}}
\item \href{#method-UserComplianceSchema-isValid}{\code{UserComplianceSchema$isValid()}}
\item \href{#method-UserComplianceSchema-getInvalidFields}{\code{UserComplianceSchema$getInvalidFields()}}
\item \href{#method-UserComplianceSchema-clone}{\code{UserComplianceSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserComplianceSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchema$new(event_at, user, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_at}}{Event time.}

\item{\code{user}}{user}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserComplianceSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
