% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_compliance_schema_user.R
\docType{class}
\name{UserComplianceSchemaUser}
\alias{UserComplianceSchemaUser}
\title{UserComplianceSchemaUser}
\format{
An \code{R6Class} generator object
}
\description{
UserComplianceSchemaUser Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserComplianceSchemaUser-new}{\code{UserComplianceSchemaUser$new()}}
\item \href{#method-UserComplianceSchemaUser-toJSON}{\code{UserComplianceSchemaUser$toJSON()}}
\item \href{#method-UserComplianceSchemaUser-fromJSON}{\code{UserComplianceSchemaUser$fromJSON()}}
\item \href{#method-UserComplianceSchemaUser-toJSONString}{\code{UserComplianceSchemaUser$toJSONString()}}
\item \href{#method-UserComplianceSchemaUser-fromJSONString}{\code{UserComplianceSchemaUser$fromJSONString()}}
\item \href{#method-UserComplianceSchemaUser-validateJSON}{\code{UserComplianceSchemaUser$validateJSON()}}
\item \href{#method-UserComplianceSchemaUser-toString}{\code{UserComplianceSchemaUser$toString()}}
\item \href{#method-UserComplianceSchemaUser-isValid}{\code{UserComplianceSchemaUser$isValid()}}
\item \href{#method-UserComplianceSchemaUser-getInvalidFields}{\code{UserComplianceSchemaUser$getInvalidFields()}}
\item \href{#method-UserComplianceSchemaUser-clone}{\code{UserComplianceSchemaUser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchemaUser-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchemaUser-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserComplianceSchemaUser class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchemaUser$new(id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Unique identifier of this User. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchemaUser-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchemaUser-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchemaUser$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserComplianceSchemaUser in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchemaUser-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchemaUser-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserComplianceSchemaUser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchemaUser$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserComplianceSchemaUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchemaUser-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchemaUser-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchemaUser$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserComplianceSchemaUser in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchemaUser-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchemaUser-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserComplianceSchemaUser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchemaUser$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserComplianceSchemaUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchemaUser-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchemaUser-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserComplianceSchemaUser and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchemaUser$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchemaUser-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchemaUser-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchemaUser$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserComplianceSchemaUser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchemaUser-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchemaUser-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchemaUser$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchemaUser-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchemaUser-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchemaUser$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceSchemaUser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceSchemaUser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceSchemaUser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
