% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_compliance_stream_response.R
\docType{class}
\name{UserComplianceStreamResponse}
\alias{UserComplianceStreamResponse}
\title{UserComplianceStreamResponse}
\format{
An \code{R6Class} generator object
}
\description{
UserComplianceStreamResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of object types defined in the oneOf schema.
Initialize a new UserComplianceStreamResponse.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserComplianceStreamResponse-new}{\code{UserComplianceStreamResponse$new()}}
\item \href{#method-UserComplianceStreamResponse-fromJSON}{\code{UserComplianceStreamResponse$fromJSON()}}
\item \href{#method-UserComplianceStreamResponse-toJSONString}{\code{UserComplianceStreamResponse$toJSONString()}}
\item \href{#method-UserComplianceStreamResponse-toJSON}{\code{UserComplianceStreamResponse$toJSON()}}
\item \href{#method-UserComplianceStreamResponse-validateJSON}{\code{UserComplianceStreamResponse$validateJSON()}}
\item \href{#method-UserComplianceStreamResponse-toString}{\code{UserComplianceStreamResponse$toString()}}
\item \href{#method-UserComplianceStreamResponse-clone}{\code{UserComplianceStreamResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserComplianceStreamResponse.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponse$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "TweetComplianceStreamResponseOneOf1", "UserComplianceStreamResponseOneOf"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserComplianceStreamResponse.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponse$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of UserComplianceStreamResponse.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize UserComplianceStreamResponse to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the UserComplianceStreamResponse.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize UserComplianceStreamResponse to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the UserComplianceStreamResponse.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to UserComplianceStreamResponse and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponse-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
