% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_compliance_stream_response_one_of.R
\docType{class}
\name{UserComplianceStreamResponseOneOf}
\alias{UserComplianceStreamResponseOneOf}
\title{UserComplianceStreamResponseOneOf}
\format{
An \code{R6Class} generator object
}
\description{
UserComplianceStreamResponseOneOf Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{UserComplianceData}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserComplianceStreamResponseOneOf-new}{\code{UserComplianceStreamResponseOneOf$new()}}
\item \href{#method-UserComplianceStreamResponseOneOf-toJSON}{\code{UserComplianceStreamResponseOneOf$toJSON()}}
\item \href{#method-UserComplianceStreamResponseOneOf-fromJSON}{\code{UserComplianceStreamResponseOneOf$fromJSON()}}
\item \href{#method-UserComplianceStreamResponseOneOf-toJSONString}{\code{UserComplianceStreamResponseOneOf$toJSONString()}}
\item \href{#method-UserComplianceStreamResponseOneOf-fromJSONString}{\code{UserComplianceStreamResponseOneOf$fromJSONString()}}
\item \href{#method-UserComplianceStreamResponseOneOf-validateJSON}{\code{UserComplianceStreamResponseOneOf$validateJSON()}}
\item \href{#method-UserComplianceStreamResponseOneOf-toString}{\code{UserComplianceStreamResponseOneOf$toString()}}
\item \href{#method-UserComplianceStreamResponseOneOf-isValid}{\code{UserComplianceStreamResponseOneOf$isValid()}}
\item \href{#method-UserComplianceStreamResponseOneOf-getInvalidFields}{\code{UserComplianceStreamResponseOneOf$getInvalidFields()}}
\item \href{#method-UserComplianceStreamResponseOneOf-clone}{\code{UserComplianceStreamResponseOneOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponseOneOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponseOneOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserComplianceStreamResponseOneOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponseOneOf$new(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponseOneOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponseOneOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponseOneOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserComplianceStreamResponseOneOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponseOneOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponseOneOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserComplianceStreamResponseOneOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponseOneOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserComplianceStreamResponseOneOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponseOneOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponseOneOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponseOneOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserComplianceStreamResponseOneOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponseOneOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponseOneOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserComplianceStreamResponseOneOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponseOneOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserComplianceStreamResponseOneOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponseOneOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponseOneOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserComplianceStreamResponseOneOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponseOneOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponseOneOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponseOneOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponseOneOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserComplianceStreamResponseOneOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponseOneOf-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponseOneOf-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponseOneOf$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponseOneOf-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponseOneOf-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponseOneOf$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserComplianceStreamResponseOneOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserComplianceStreamResponseOneOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserComplianceStreamResponseOneOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
