% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_delete_compliance_schema.R
\docType{class}
\name{UserDeleteComplianceSchema}
\alias{UserDeleteComplianceSchema}
\title{UserDeleteComplianceSchema}
\format{
An \code{R6Class} generator object
}
\description{
UserDeleteComplianceSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user_delete}}{\link{UserComplianceSchema}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserDeleteComplianceSchema-new}{\code{UserDeleteComplianceSchema$new()}}
\item \href{#method-UserDeleteComplianceSchema-toJSON}{\code{UserDeleteComplianceSchema$toJSON()}}
\item \href{#method-UserDeleteComplianceSchema-fromJSON}{\code{UserDeleteComplianceSchema$fromJSON()}}
\item \href{#method-UserDeleteComplianceSchema-toJSONString}{\code{UserDeleteComplianceSchema$toJSONString()}}
\item \href{#method-UserDeleteComplianceSchema-fromJSONString}{\code{UserDeleteComplianceSchema$fromJSONString()}}
\item \href{#method-UserDeleteComplianceSchema-validateJSON}{\code{UserDeleteComplianceSchema$validateJSON()}}
\item \href{#method-UserDeleteComplianceSchema-toString}{\code{UserDeleteComplianceSchema$toString()}}
\item \href{#method-UserDeleteComplianceSchema-isValid}{\code{UserDeleteComplianceSchema$isValid()}}
\item \href{#method-UserDeleteComplianceSchema-getInvalidFields}{\code{UserDeleteComplianceSchema$getInvalidFields()}}
\item \href{#method-UserDeleteComplianceSchema-clone}{\code{UserDeleteComplianceSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserDeleteComplianceSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserDeleteComplianceSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserDeleteComplianceSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDeleteComplianceSchema$new(user_delete, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_delete}}{user_delete}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserDeleteComplianceSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserDeleteComplianceSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDeleteComplianceSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserDeleteComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserDeleteComplianceSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserDeleteComplianceSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserDeleteComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDeleteComplianceSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserDeleteComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserDeleteComplianceSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserDeleteComplianceSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDeleteComplianceSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserDeleteComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserDeleteComplianceSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserDeleteComplianceSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserDeleteComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDeleteComplianceSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserDeleteComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserDeleteComplianceSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserDeleteComplianceSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserDeleteComplianceSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDeleteComplianceSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserDeleteComplianceSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserDeleteComplianceSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDeleteComplianceSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserDeleteComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserDeleteComplianceSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserDeleteComplianceSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDeleteComplianceSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserDeleteComplianceSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserDeleteComplianceSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDeleteComplianceSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserDeleteComplianceSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserDeleteComplianceSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDeleteComplianceSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
