% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_entities.R
\docType{class}
\name{UserEntities}
\alias{UserEntities}
\title{UserEntities}
\format{
An \code{R6Class} generator object
}
\description{
UserEntities Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{\link{FullTextEntities} [optional]}

\item{\code{url}}{\link{UserEntitiesUrl} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserEntities-new}{\code{UserEntities$new()}}
\item \href{#method-UserEntities-toJSON}{\code{UserEntities$toJSON()}}
\item \href{#method-UserEntities-fromJSON}{\code{UserEntities$fromJSON()}}
\item \href{#method-UserEntities-toJSONString}{\code{UserEntities$toJSONString()}}
\item \href{#method-UserEntities-fromJSONString}{\code{UserEntities$fromJSONString()}}
\item \href{#method-UserEntities-validateJSON}{\code{UserEntities$validateJSON()}}
\item \href{#method-UserEntities-toString}{\code{UserEntities$toString()}}
\item \href{#method-UserEntities-isValid}{\code{UserEntities$isValid()}}
\item \href{#method-UserEntities-getInvalidFields}{\code{UserEntities$getInvalidFields()}}
\item \href{#method-UserEntities-clone}{\code{UserEntities$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntities-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntities-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserEntities class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntities$new(description = NULL, url = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{description}

\item{\code{url}}{url}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntities-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntities-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntities$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserEntities in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntities-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntities-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserEntities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntities$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserEntities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntities-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntities-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntities$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserEntities in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntities-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntities-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserEntities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntities$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserEntities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntities-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntities-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserEntities and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntities$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntities-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntities-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntities$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserEntities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntities-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntities-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntities$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntities-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntities-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntities$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntities-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntities-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntities$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
