% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_entities_url.R
\docType{class}
\name{UserEntitiesUrl}
\alias{UserEntitiesUrl}
\title{UserEntitiesUrl}
\format{
An \code{R6Class} generator object
}
\description{
UserEntitiesUrl Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{urls}}{list( \link{UrlEntity} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserEntitiesUrl-new}{\code{UserEntitiesUrl$new()}}
\item \href{#method-UserEntitiesUrl-toJSON}{\code{UserEntitiesUrl$toJSON()}}
\item \href{#method-UserEntitiesUrl-fromJSON}{\code{UserEntitiesUrl$fromJSON()}}
\item \href{#method-UserEntitiesUrl-toJSONString}{\code{UserEntitiesUrl$toJSONString()}}
\item \href{#method-UserEntitiesUrl-fromJSONString}{\code{UserEntitiesUrl$fromJSONString()}}
\item \href{#method-UserEntitiesUrl-validateJSON}{\code{UserEntitiesUrl$validateJSON()}}
\item \href{#method-UserEntitiesUrl-toString}{\code{UserEntitiesUrl$toString()}}
\item \href{#method-UserEntitiesUrl-isValid}{\code{UserEntitiesUrl$isValid()}}
\item \href{#method-UserEntitiesUrl-getInvalidFields}{\code{UserEntitiesUrl$getInvalidFields()}}
\item \href{#method-UserEntitiesUrl-clone}{\code{UserEntitiesUrl$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntitiesUrl-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntitiesUrl-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserEntitiesUrl class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntitiesUrl$new(urls = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{urls}}{urls}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntitiesUrl-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntitiesUrl-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntitiesUrl$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserEntitiesUrl in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntitiesUrl-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntitiesUrl-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserEntitiesUrl
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntitiesUrl$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserEntitiesUrl
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntitiesUrl-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntitiesUrl-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntitiesUrl$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserEntitiesUrl in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntitiesUrl-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntitiesUrl-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserEntitiesUrl
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntitiesUrl$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserEntitiesUrl
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntitiesUrl-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntitiesUrl-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserEntitiesUrl and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntitiesUrl$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntitiesUrl-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntitiesUrl-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntitiesUrl$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserEntitiesUrl
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntitiesUrl-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntitiesUrl-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntitiesUrl$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntitiesUrl-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntitiesUrl-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntitiesUrl$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserEntitiesUrl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserEntitiesUrl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserEntitiesUrl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
