% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_profile_modification_compliance_schema.R
\docType{class}
\name{UserProfileModificationComplianceSchema}
\alias{UserProfileModificationComplianceSchema}
\title{UserProfileModificationComplianceSchema}
\format{
An \code{R6Class} generator object
}
\description{
UserProfileModificationComplianceSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user_profile_modification}}{\link{UserProfileModificationObjectSchema}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserProfileModificationComplianceSchema-new}{\code{UserProfileModificationComplianceSchema$new()}}
\item \href{#method-UserProfileModificationComplianceSchema-toJSON}{\code{UserProfileModificationComplianceSchema$toJSON()}}
\item \href{#method-UserProfileModificationComplianceSchema-fromJSON}{\code{UserProfileModificationComplianceSchema$fromJSON()}}
\item \href{#method-UserProfileModificationComplianceSchema-toJSONString}{\code{UserProfileModificationComplianceSchema$toJSONString()}}
\item \href{#method-UserProfileModificationComplianceSchema-fromJSONString}{\code{UserProfileModificationComplianceSchema$fromJSONString()}}
\item \href{#method-UserProfileModificationComplianceSchema-validateJSON}{\code{UserProfileModificationComplianceSchema$validateJSON()}}
\item \href{#method-UserProfileModificationComplianceSchema-toString}{\code{UserProfileModificationComplianceSchema$toString()}}
\item \href{#method-UserProfileModificationComplianceSchema-isValid}{\code{UserProfileModificationComplianceSchema$isValid()}}
\item \href{#method-UserProfileModificationComplianceSchema-getInvalidFields}{\code{UserProfileModificationComplianceSchema$getInvalidFields()}}
\item \href{#method-UserProfileModificationComplianceSchema-clone}{\code{UserProfileModificationComplianceSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserProfileModificationComplianceSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserProfileModificationComplianceSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserProfileModificationComplianceSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserProfileModificationComplianceSchema$new(user_profile_modification, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_profile_modification}}{user_profile_modification}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserProfileModificationComplianceSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserProfileModificationComplianceSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserProfileModificationComplianceSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserProfileModificationComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserProfileModificationComplianceSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserProfileModificationComplianceSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserProfileModificationComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserProfileModificationComplianceSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserProfileModificationComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserProfileModificationComplianceSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserProfileModificationComplianceSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserProfileModificationComplianceSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserProfileModificationComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserProfileModificationComplianceSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserProfileModificationComplianceSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserProfileModificationComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserProfileModificationComplianceSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserProfileModificationComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserProfileModificationComplianceSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserProfileModificationComplianceSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserProfileModificationComplianceSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserProfileModificationComplianceSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserProfileModificationComplianceSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserProfileModificationComplianceSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserProfileModificationComplianceSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserProfileModificationComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserProfileModificationComplianceSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserProfileModificationComplianceSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserProfileModificationComplianceSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserProfileModificationComplianceSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserProfileModificationComplianceSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserProfileModificationComplianceSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserProfileModificationComplianceSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserProfileModificationComplianceSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserProfileModificationComplianceSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
