% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_public_metrics.R
\docType{class}
\name{UserPublicMetrics}
\alias{UserPublicMetrics}
\title{UserPublicMetrics}
\format{
An \code{R6Class} generator object
}
\description{
UserPublicMetrics Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{followers_count}}{integer}

\item{\code{following_count}}{integer}

\item{\code{listed_count}}{integer}

\item{\code{tweet_count}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserPublicMetrics-new}{\code{UserPublicMetrics$new()}}
\item \href{#method-UserPublicMetrics-toJSON}{\code{UserPublicMetrics$toJSON()}}
\item \href{#method-UserPublicMetrics-fromJSON}{\code{UserPublicMetrics$fromJSON()}}
\item \href{#method-UserPublicMetrics-toJSONString}{\code{UserPublicMetrics$toJSONString()}}
\item \href{#method-UserPublicMetrics-fromJSONString}{\code{UserPublicMetrics$fromJSONString()}}
\item \href{#method-UserPublicMetrics-validateJSON}{\code{UserPublicMetrics$validateJSON()}}
\item \href{#method-UserPublicMetrics-toString}{\code{UserPublicMetrics$toString()}}
\item \href{#method-UserPublicMetrics-isValid}{\code{UserPublicMetrics$isValid()}}
\item \href{#method-UserPublicMetrics-getInvalidFields}{\code{UserPublicMetrics$getInvalidFields()}}
\item \href{#method-UserPublicMetrics-clone}{\code{UserPublicMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserPublicMetrics-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserPublicMetrics-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserPublicMetrics class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserPublicMetrics$new(
  followers_count,
  following_count,
  listed_count,
  tweet_count,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{followers_count}}{Number of Users who are following this User.}

\item{\code{following_count}}{Number of Users this User is following.}

\item{\code{listed_count}}{The number of lists that include this User.}

\item{\code{tweet_count}}{The number of Tweets (including Retweets) posted by this User.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserPublicMetrics-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserPublicMetrics-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserPublicMetrics$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserPublicMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserPublicMetrics-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserPublicMetrics-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserPublicMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserPublicMetrics$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserPublicMetrics-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserPublicMetrics-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserPublicMetrics$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserPublicMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserPublicMetrics-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserPublicMetrics-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserPublicMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserPublicMetrics$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserPublicMetrics-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserPublicMetrics-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserPublicMetrics and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserPublicMetrics$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserPublicMetrics-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserPublicMetrics-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserPublicMetrics$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserPublicMetrics-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserPublicMetrics-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserPublicMetrics$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserPublicMetrics-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserPublicMetrics-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserPublicMetrics$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserPublicMetrics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserPublicMetrics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserPublicMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
