% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_scrub_geo_object_schema.R
\docType{class}
\name{UserScrubGeoObjectSchema}
\alias{UserScrubGeoObjectSchema}
\title{UserScrubGeoObjectSchema}
\format{
An \code{R6Class} generator object
}
\description{
UserScrubGeoObjectSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{event_at}}{character}

\item{\code{up_to_tweet_id}}{character}

\item{\code{user}}{\link{UserComplianceSchemaUser}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserScrubGeoObjectSchema-new}{\code{UserScrubGeoObjectSchema$new()}}
\item \href{#method-UserScrubGeoObjectSchema-toJSON}{\code{UserScrubGeoObjectSchema$toJSON()}}
\item \href{#method-UserScrubGeoObjectSchema-fromJSON}{\code{UserScrubGeoObjectSchema$fromJSON()}}
\item \href{#method-UserScrubGeoObjectSchema-toJSONString}{\code{UserScrubGeoObjectSchema$toJSONString()}}
\item \href{#method-UserScrubGeoObjectSchema-fromJSONString}{\code{UserScrubGeoObjectSchema$fromJSONString()}}
\item \href{#method-UserScrubGeoObjectSchema-validateJSON}{\code{UserScrubGeoObjectSchema$validateJSON()}}
\item \href{#method-UserScrubGeoObjectSchema-toString}{\code{UserScrubGeoObjectSchema$toString()}}
\item \href{#method-UserScrubGeoObjectSchema-isValid}{\code{UserScrubGeoObjectSchema$isValid()}}
\item \href{#method-UserScrubGeoObjectSchema-getInvalidFields}{\code{UserScrubGeoObjectSchema$getInvalidFields()}}
\item \href{#method-UserScrubGeoObjectSchema-clone}{\code{UserScrubGeoObjectSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoObjectSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoObjectSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserScrubGeoObjectSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoObjectSchema$new(event_at, up_to_tweet_id, user, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_at}}{Event time.}

\item{\code{up_to_tweet_id}}{Unique identifier of this Tweet. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.}

\item{\code{user}}{user}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoObjectSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoObjectSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoObjectSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserScrubGeoObjectSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoObjectSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoObjectSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserScrubGeoObjectSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoObjectSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserScrubGeoObjectSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoObjectSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoObjectSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoObjectSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserScrubGeoObjectSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoObjectSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoObjectSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserScrubGeoObjectSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoObjectSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserScrubGeoObjectSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoObjectSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoObjectSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserScrubGeoObjectSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoObjectSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoObjectSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoObjectSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoObjectSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserScrubGeoObjectSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoObjectSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoObjectSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoObjectSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoObjectSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoObjectSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoObjectSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoObjectSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoObjectSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoObjectSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
