% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_scrub_geo_schema.R
\docType{class}
\name{UserScrubGeoSchema}
\alias{UserScrubGeoSchema}
\title{UserScrubGeoSchema}
\format{
An \code{R6Class} generator object
}
\description{
UserScrubGeoSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{scrub_geo}}{\link{UserScrubGeoObjectSchema}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserScrubGeoSchema-new}{\code{UserScrubGeoSchema$new()}}
\item \href{#method-UserScrubGeoSchema-toJSON}{\code{UserScrubGeoSchema$toJSON()}}
\item \href{#method-UserScrubGeoSchema-fromJSON}{\code{UserScrubGeoSchema$fromJSON()}}
\item \href{#method-UserScrubGeoSchema-toJSONString}{\code{UserScrubGeoSchema$toJSONString()}}
\item \href{#method-UserScrubGeoSchema-fromJSONString}{\code{UserScrubGeoSchema$fromJSONString()}}
\item \href{#method-UserScrubGeoSchema-validateJSON}{\code{UserScrubGeoSchema$validateJSON()}}
\item \href{#method-UserScrubGeoSchema-toString}{\code{UserScrubGeoSchema$toString()}}
\item \href{#method-UserScrubGeoSchema-isValid}{\code{UserScrubGeoSchema$isValid()}}
\item \href{#method-UserScrubGeoSchema-getInvalidFields}{\code{UserScrubGeoSchema$getInvalidFields()}}
\item \href{#method-UserScrubGeoSchema-clone}{\code{UserScrubGeoSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserScrubGeoSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoSchema$new(scrub_geo, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scrub_geo}}{scrub_geo}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserScrubGeoSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserScrubGeoSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserScrubGeoSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserScrubGeoSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserScrubGeoSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserScrubGeoSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserScrubGeoSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserScrubGeoSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserScrubGeoSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserScrubGeoSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserScrubGeoSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
