% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_following_create_request.R
\docType{class}
\name{UsersFollowingCreateRequest}
\alias{UsersFollowingCreateRequest}
\title{UsersFollowingCreateRequest}
\format{
An \code{R6Class} generator object
}
\description{
UsersFollowingCreateRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{target_user_id}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsersFollowingCreateRequest-new}{\code{UsersFollowingCreateRequest$new()}}
\item \href{#method-UsersFollowingCreateRequest-toJSON}{\code{UsersFollowingCreateRequest$toJSON()}}
\item \href{#method-UsersFollowingCreateRequest-fromJSON}{\code{UsersFollowingCreateRequest$fromJSON()}}
\item \href{#method-UsersFollowingCreateRequest-toJSONString}{\code{UsersFollowingCreateRequest$toJSONString()}}
\item \href{#method-UsersFollowingCreateRequest-fromJSONString}{\code{UsersFollowingCreateRequest$fromJSONString()}}
\item \href{#method-UsersFollowingCreateRequest-validateJSON}{\code{UsersFollowingCreateRequest$validateJSON()}}
\item \href{#method-UsersFollowingCreateRequest-toString}{\code{UsersFollowingCreateRequest$toString()}}
\item \href{#method-UsersFollowingCreateRequest-isValid}{\code{UsersFollowingCreateRequest$isValid()}}
\item \href{#method-UsersFollowingCreateRequest-getInvalidFields}{\code{UsersFollowingCreateRequest$getInvalidFields()}}
\item \href{#method-UsersFollowingCreateRequest-clone}{\code{UsersFollowingCreateRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UsersFollowingCreateRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateRequest$new(target_user_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_user_id}}{Unique identifier of this User. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersFollowingCreateRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UsersFollowingCreateRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersFollowingCreateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersFollowingCreateRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UsersFollowingCreateRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersFollowingCreateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UsersFollowingCreateRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UsersFollowingCreateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
