% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_following_create_response.R
\docType{class}
\name{UsersFollowingCreateResponse}
\alias{UsersFollowingCreateResponse}
\title{UsersFollowingCreateResponse}
\format{
An \code{R6Class} generator object
}
\description{
UsersFollowingCreateResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{UsersFollowingCreateResponseData} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsersFollowingCreateResponse-new}{\code{UsersFollowingCreateResponse$new()}}
\item \href{#method-UsersFollowingCreateResponse-toJSON}{\code{UsersFollowingCreateResponse$toJSON()}}
\item \href{#method-UsersFollowingCreateResponse-fromJSON}{\code{UsersFollowingCreateResponse$fromJSON()}}
\item \href{#method-UsersFollowingCreateResponse-toJSONString}{\code{UsersFollowingCreateResponse$toJSONString()}}
\item \href{#method-UsersFollowingCreateResponse-fromJSONString}{\code{UsersFollowingCreateResponse$fromJSONString()}}
\item \href{#method-UsersFollowingCreateResponse-validateJSON}{\code{UsersFollowingCreateResponse$validateJSON()}}
\item \href{#method-UsersFollowingCreateResponse-toString}{\code{UsersFollowingCreateResponse$toString()}}
\item \href{#method-UsersFollowingCreateResponse-isValid}{\code{UsersFollowingCreateResponse$isValid()}}
\item \href{#method-UsersFollowingCreateResponse-getInvalidFields}{\code{UsersFollowingCreateResponse$getInvalidFields()}}
\item \href{#method-UsersFollowingCreateResponse-clone}{\code{UsersFollowingCreateResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UsersFollowingCreateResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersFollowingCreateResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UsersFollowingCreateResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersFollowingCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersFollowingCreateResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UsersFollowingCreateResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersFollowingCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UsersFollowingCreateResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UsersFollowingCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
