% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_following_create_response_data.R
\docType{class}
\name{UsersFollowingCreateResponseData}
\alias{UsersFollowingCreateResponseData}
\title{UsersFollowingCreateResponseData}
\format{
An \code{R6Class} generator object
}
\description{
UsersFollowingCreateResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{following}}{character [optional]}

\item{\code{pending_follow}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsersFollowingCreateResponseData-new}{\code{UsersFollowingCreateResponseData$new()}}
\item \href{#method-UsersFollowingCreateResponseData-toJSON}{\code{UsersFollowingCreateResponseData$toJSON()}}
\item \href{#method-UsersFollowingCreateResponseData-fromJSON}{\code{UsersFollowingCreateResponseData$fromJSON()}}
\item \href{#method-UsersFollowingCreateResponseData-toJSONString}{\code{UsersFollowingCreateResponseData$toJSONString()}}
\item \href{#method-UsersFollowingCreateResponseData-fromJSONString}{\code{UsersFollowingCreateResponseData$fromJSONString()}}
\item \href{#method-UsersFollowingCreateResponseData-validateJSON}{\code{UsersFollowingCreateResponseData$validateJSON()}}
\item \href{#method-UsersFollowingCreateResponseData-toString}{\code{UsersFollowingCreateResponseData$toString()}}
\item \href{#method-UsersFollowingCreateResponseData-isValid}{\code{UsersFollowingCreateResponseData$isValid()}}
\item \href{#method-UsersFollowingCreateResponseData-getInvalidFields}{\code{UsersFollowingCreateResponseData$getInvalidFields()}}
\item \href{#method-UsersFollowingCreateResponseData-clone}{\code{UsersFollowingCreateResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UsersFollowingCreateResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponseData$new(
  following = NULL,
  pending_follow = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{following}}{following}

\item{\code{pending_follow}}{pending_follow}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersFollowingCreateResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UsersFollowingCreateResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersFollowingCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersFollowingCreateResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UsersFollowingCreateResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersFollowingCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UsersFollowingCreateResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UsersFollowingCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingCreateResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingCreateResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingCreateResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
