% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_following_delete_response.R
\docType{class}
\name{UsersFollowingDeleteResponse}
\alias{UsersFollowingDeleteResponse}
\title{UsersFollowingDeleteResponse}
\format{
An \code{R6Class} generator object
}
\description{
UsersFollowingDeleteResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{ListFollowedResponseData} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsersFollowingDeleteResponse-new}{\code{UsersFollowingDeleteResponse$new()}}
\item \href{#method-UsersFollowingDeleteResponse-toJSON}{\code{UsersFollowingDeleteResponse$toJSON()}}
\item \href{#method-UsersFollowingDeleteResponse-fromJSON}{\code{UsersFollowingDeleteResponse$fromJSON()}}
\item \href{#method-UsersFollowingDeleteResponse-toJSONString}{\code{UsersFollowingDeleteResponse$toJSONString()}}
\item \href{#method-UsersFollowingDeleteResponse-fromJSONString}{\code{UsersFollowingDeleteResponse$fromJSONString()}}
\item \href{#method-UsersFollowingDeleteResponse-validateJSON}{\code{UsersFollowingDeleteResponse$validateJSON()}}
\item \href{#method-UsersFollowingDeleteResponse-toString}{\code{UsersFollowingDeleteResponse$toString()}}
\item \href{#method-UsersFollowingDeleteResponse-isValid}{\code{UsersFollowingDeleteResponse$isValid()}}
\item \href{#method-UsersFollowingDeleteResponse-getInvalidFields}{\code{UsersFollowingDeleteResponse$getInvalidFields()}}
\item \href{#method-UsersFollowingDeleteResponse-clone}{\code{UsersFollowingDeleteResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingDeleteResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingDeleteResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UsersFollowingDeleteResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingDeleteResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingDeleteResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingDeleteResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingDeleteResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersFollowingDeleteResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingDeleteResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingDeleteResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UsersFollowingDeleteResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingDeleteResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersFollowingDeleteResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingDeleteResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingDeleteResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingDeleteResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersFollowingDeleteResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingDeleteResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingDeleteResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UsersFollowingDeleteResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingDeleteResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersFollowingDeleteResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingDeleteResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingDeleteResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UsersFollowingDeleteResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingDeleteResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingDeleteResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingDeleteResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingDeleteResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UsersFollowingDeleteResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingDeleteResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingDeleteResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingDeleteResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingDeleteResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingDeleteResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingDeleteResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersFollowingDeleteResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsersFollowingDeleteResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersFollowingDeleteResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
