% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_likes_create_request.R
\docType{class}
\name{UsersLikesCreateRequest}
\alias{UsersLikesCreateRequest}
\title{UsersLikesCreateRequest}
\format{
An \code{R6Class} generator object
}
\description{
UsersLikesCreateRequest Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tweet_id}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsersLikesCreateRequest-new}{\code{UsersLikesCreateRequest$new()}}
\item \href{#method-UsersLikesCreateRequest-toJSON}{\code{UsersLikesCreateRequest$toJSON()}}
\item \href{#method-UsersLikesCreateRequest-fromJSON}{\code{UsersLikesCreateRequest$fromJSON()}}
\item \href{#method-UsersLikesCreateRequest-toJSONString}{\code{UsersLikesCreateRequest$toJSONString()}}
\item \href{#method-UsersLikesCreateRequest-fromJSONString}{\code{UsersLikesCreateRequest$fromJSONString()}}
\item \href{#method-UsersLikesCreateRequest-validateJSON}{\code{UsersLikesCreateRequest$validateJSON()}}
\item \href{#method-UsersLikesCreateRequest-toString}{\code{UsersLikesCreateRequest$toString()}}
\item \href{#method-UsersLikesCreateRequest-isValid}{\code{UsersLikesCreateRequest$isValid()}}
\item \href{#method-UsersLikesCreateRequest-getInvalidFields}{\code{UsersLikesCreateRequest$getInvalidFields()}}
\item \href{#method-UsersLikesCreateRequest-clone}{\code{UsersLikesCreateRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UsersLikesCreateRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateRequest$new(tweet_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tweet_id}}{Unique identifier of this Tweet. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersLikesCreateRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UsersLikesCreateRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersLikesCreateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersLikesCreateRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UsersLikesCreateRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersLikesCreateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UsersLikesCreateRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UsersLikesCreateRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateRequest-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateRequest-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateRequest$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateRequest-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateRequest-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateRequest$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
