% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_likes_create_response.R
\docType{class}
\name{UsersLikesCreateResponse}
\alias{UsersLikesCreateResponse}
\title{UsersLikesCreateResponse}
\format{
An \code{R6Class} generator object
}
\description{
UsersLikesCreateResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{UsersLikesCreateResponseData} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsersLikesCreateResponse-new}{\code{UsersLikesCreateResponse$new()}}
\item \href{#method-UsersLikesCreateResponse-toJSON}{\code{UsersLikesCreateResponse$toJSON()}}
\item \href{#method-UsersLikesCreateResponse-fromJSON}{\code{UsersLikesCreateResponse$fromJSON()}}
\item \href{#method-UsersLikesCreateResponse-toJSONString}{\code{UsersLikesCreateResponse$toJSONString()}}
\item \href{#method-UsersLikesCreateResponse-fromJSONString}{\code{UsersLikesCreateResponse$fromJSONString()}}
\item \href{#method-UsersLikesCreateResponse-validateJSON}{\code{UsersLikesCreateResponse$validateJSON()}}
\item \href{#method-UsersLikesCreateResponse-toString}{\code{UsersLikesCreateResponse$toString()}}
\item \href{#method-UsersLikesCreateResponse-isValid}{\code{UsersLikesCreateResponse$isValid()}}
\item \href{#method-UsersLikesCreateResponse-getInvalidFields}{\code{UsersLikesCreateResponse$getInvalidFields()}}
\item \href{#method-UsersLikesCreateResponse-clone}{\code{UsersLikesCreateResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UsersLikesCreateResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersLikesCreateResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UsersLikesCreateResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersLikesCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersLikesCreateResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UsersLikesCreateResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersLikesCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UsersLikesCreateResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UsersLikesCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
