% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_likes_create_response_data.R
\docType{class}
\name{UsersLikesCreateResponseData}
\alias{UsersLikesCreateResponseData}
\title{UsersLikesCreateResponseData}
\format{
An \code{R6Class} generator object
}
\description{
UsersLikesCreateResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{liked}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsersLikesCreateResponseData-new}{\code{UsersLikesCreateResponseData$new()}}
\item \href{#method-UsersLikesCreateResponseData-toJSON}{\code{UsersLikesCreateResponseData$toJSON()}}
\item \href{#method-UsersLikesCreateResponseData-fromJSON}{\code{UsersLikesCreateResponseData$fromJSON()}}
\item \href{#method-UsersLikesCreateResponseData-toJSONString}{\code{UsersLikesCreateResponseData$toJSONString()}}
\item \href{#method-UsersLikesCreateResponseData-fromJSONString}{\code{UsersLikesCreateResponseData$fromJSONString()}}
\item \href{#method-UsersLikesCreateResponseData-validateJSON}{\code{UsersLikesCreateResponseData$validateJSON()}}
\item \href{#method-UsersLikesCreateResponseData-toString}{\code{UsersLikesCreateResponseData$toString()}}
\item \href{#method-UsersLikesCreateResponseData-isValid}{\code{UsersLikesCreateResponseData$isValid()}}
\item \href{#method-UsersLikesCreateResponseData-getInvalidFields}{\code{UsersLikesCreateResponseData$getInvalidFields()}}
\item \href{#method-UsersLikesCreateResponseData-clone}{\code{UsersLikesCreateResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UsersLikesCreateResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponseData$new(liked = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{liked}}{liked}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersLikesCreateResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UsersLikesCreateResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersLikesCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersLikesCreateResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UsersLikesCreateResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersLikesCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UsersLikesCreateResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UsersLikesCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesCreateResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesCreateResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesCreateResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
