% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_likes_delete_response.R
\docType{class}
\name{UsersLikesDeleteResponse}
\alias{UsersLikesDeleteResponse}
\title{UsersLikesDeleteResponse}
\format{
An \code{R6Class} generator object
}
\description{
UsersLikesDeleteResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{UsersLikesCreateResponseData} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsersLikesDeleteResponse-new}{\code{UsersLikesDeleteResponse$new()}}
\item \href{#method-UsersLikesDeleteResponse-toJSON}{\code{UsersLikesDeleteResponse$toJSON()}}
\item \href{#method-UsersLikesDeleteResponse-fromJSON}{\code{UsersLikesDeleteResponse$fromJSON()}}
\item \href{#method-UsersLikesDeleteResponse-toJSONString}{\code{UsersLikesDeleteResponse$toJSONString()}}
\item \href{#method-UsersLikesDeleteResponse-fromJSONString}{\code{UsersLikesDeleteResponse$fromJSONString()}}
\item \href{#method-UsersLikesDeleteResponse-validateJSON}{\code{UsersLikesDeleteResponse$validateJSON()}}
\item \href{#method-UsersLikesDeleteResponse-toString}{\code{UsersLikesDeleteResponse$toString()}}
\item \href{#method-UsersLikesDeleteResponse-isValid}{\code{UsersLikesDeleteResponse$isValid()}}
\item \href{#method-UsersLikesDeleteResponse-getInvalidFields}{\code{UsersLikesDeleteResponse$getInvalidFields()}}
\item \href{#method-UsersLikesDeleteResponse-clone}{\code{UsersLikesDeleteResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesDeleteResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesDeleteResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UsersLikesDeleteResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesDeleteResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesDeleteResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesDeleteResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesDeleteResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersLikesDeleteResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesDeleteResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesDeleteResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UsersLikesDeleteResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesDeleteResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersLikesDeleteResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesDeleteResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesDeleteResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesDeleteResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersLikesDeleteResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesDeleteResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesDeleteResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UsersLikesDeleteResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesDeleteResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersLikesDeleteResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesDeleteResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesDeleteResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UsersLikesDeleteResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesDeleteResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesDeleteResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesDeleteResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesDeleteResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UsersLikesDeleteResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesDeleteResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesDeleteResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesDeleteResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesDeleteResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesDeleteResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesDeleteResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersLikesDeleteResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsersLikesDeleteResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersLikesDeleteResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
