% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_retweets_create_response.R
\docType{class}
\name{UsersRetweetsCreateResponse}
\alias{UsersRetweetsCreateResponse}
\title{UsersRetweetsCreateResponse}
\format{
An \code{R6Class} generator object
}
\description{
UsersRetweetsCreateResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{UsersRetweetsCreateResponseData} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsersRetweetsCreateResponse-new}{\code{UsersRetweetsCreateResponse$new()}}
\item \href{#method-UsersRetweetsCreateResponse-toJSON}{\code{UsersRetweetsCreateResponse$toJSON()}}
\item \href{#method-UsersRetweetsCreateResponse-fromJSON}{\code{UsersRetweetsCreateResponse$fromJSON()}}
\item \href{#method-UsersRetweetsCreateResponse-toJSONString}{\code{UsersRetweetsCreateResponse$toJSONString()}}
\item \href{#method-UsersRetweetsCreateResponse-fromJSONString}{\code{UsersRetweetsCreateResponse$fromJSONString()}}
\item \href{#method-UsersRetweetsCreateResponse-validateJSON}{\code{UsersRetweetsCreateResponse$validateJSON()}}
\item \href{#method-UsersRetweetsCreateResponse-toString}{\code{UsersRetweetsCreateResponse$toString()}}
\item \href{#method-UsersRetweetsCreateResponse-isValid}{\code{UsersRetweetsCreateResponse$isValid()}}
\item \href{#method-UsersRetweetsCreateResponse-getInvalidFields}{\code{UsersRetweetsCreateResponse$getInvalidFields()}}
\item \href{#method-UsersRetweetsCreateResponse-clone}{\code{UsersRetweetsCreateResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UsersRetweetsCreateResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersRetweetsCreateResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UsersRetweetsCreateResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersRetweetsCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersRetweetsCreateResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UsersRetweetsCreateResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersRetweetsCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UsersRetweetsCreateResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UsersRetweetsCreateResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
