% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_retweets_create_response_data.R
\docType{class}
\name{UsersRetweetsCreateResponseData}
\alias{UsersRetweetsCreateResponseData}
\title{UsersRetweetsCreateResponseData}
\format{
An \code{R6Class} generator object
}
\description{
UsersRetweetsCreateResponseData Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{retweeted}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsersRetweetsCreateResponseData-new}{\code{UsersRetweetsCreateResponseData$new()}}
\item \href{#method-UsersRetweetsCreateResponseData-toJSON}{\code{UsersRetweetsCreateResponseData$toJSON()}}
\item \href{#method-UsersRetweetsCreateResponseData-fromJSON}{\code{UsersRetweetsCreateResponseData$fromJSON()}}
\item \href{#method-UsersRetweetsCreateResponseData-toJSONString}{\code{UsersRetweetsCreateResponseData$toJSONString()}}
\item \href{#method-UsersRetweetsCreateResponseData-fromJSONString}{\code{UsersRetweetsCreateResponseData$fromJSONString()}}
\item \href{#method-UsersRetweetsCreateResponseData-validateJSON}{\code{UsersRetweetsCreateResponseData$validateJSON()}}
\item \href{#method-UsersRetweetsCreateResponseData-toString}{\code{UsersRetweetsCreateResponseData$toString()}}
\item \href{#method-UsersRetweetsCreateResponseData-isValid}{\code{UsersRetweetsCreateResponseData$isValid()}}
\item \href{#method-UsersRetweetsCreateResponseData-getInvalidFields}{\code{UsersRetweetsCreateResponseData$getInvalidFields()}}
\item \href{#method-UsersRetweetsCreateResponseData-clone}{\code{UsersRetweetsCreateResponseData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponseData-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponseData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UsersRetweetsCreateResponseData class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponseData$new(retweeted = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{retweeted}}{retweeted}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponseData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponseData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponseData$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersRetweetsCreateResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponseData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponseData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UsersRetweetsCreateResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponseData$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersRetweetsCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponseData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponseData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponseData$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersRetweetsCreateResponseData in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponseData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponseData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UsersRetweetsCreateResponseData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponseData$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersRetweetsCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponseData-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponseData-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UsersRetweetsCreateResponseData and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponseData$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponseData-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponseData-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponseData$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UsersRetweetsCreateResponseData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponseData-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponseData-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponseData$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponseData-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponseData-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponseData$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsCreateResponseData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsCreateResponseData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsCreateResponseData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
