% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_retweets_delete_response.R
\docType{class}
\name{UsersRetweetsDeleteResponse}
\alias{UsersRetweetsDeleteResponse}
\title{UsersRetweetsDeleteResponse}
\format{
An \code{R6Class} generator object
}
\description{
UsersRetweetsDeleteResponse Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{UsersRetweetsCreateResponseData} [optional]}

\item{\code{errors}}{list( \link{Problem} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UsersRetweetsDeleteResponse-new}{\code{UsersRetweetsDeleteResponse$new()}}
\item \href{#method-UsersRetweetsDeleteResponse-toJSON}{\code{UsersRetweetsDeleteResponse$toJSON()}}
\item \href{#method-UsersRetweetsDeleteResponse-fromJSON}{\code{UsersRetweetsDeleteResponse$fromJSON()}}
\item \href{#method-UsersRetweetsDeleteResponse-toJSONString}{\code{UsersRetweetsDeleteResponse$toJSONString()}}
\item \href{#method-UsersRetweetsDeleteResponse-fromJSONString}{\code{UsersRetweetsDeleteResponse$fromJSONString()}}
\item \href{#method-UsersRetweetsDeleteResponse-validateJSON}{\code{UsersRetweetsDeleteResponse$validateJSON()}}
\item \href{#method-UsersRetweetsDeleteResponse-toString}{\code{UsersRetweetsDeleteResponse$toString()}}
\item \href{#method-UsersRetweetsDeleteResponse-isValid}{\code{UsersRetweetsDeleteResponse$isValid()}}
\item \href{#method-UsersRetweetsDeleteResponse-getInvalidFields}{\code{UsersRetweetsDeleteResponse$getInvalidFields()}}
\item \href{#method-UsersRetweetsDeleteResponse-clone}{\code{UsersRetweetsDeleteResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsDeleteResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsDeleteResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UsersRetweetsDeleteResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsDeleteResponse$new(data = NULL, errors = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsDeleteResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsDeleteResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsDeleteResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersRetweetsDeleteResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsDeleteResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsDeleteResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UsersRetweetsDeleteResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsDeleteResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersRetweetsDeleteResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsDeleteResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsDeleteResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsDeleteResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UsersRetweetsDeleteResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsDeleteResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsDeleteResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UsersRetweetsDeleteResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsDeleteResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UsersRetweetsDeleteResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsDeleteResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsDeleteResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UsersRetweetsDeleteResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsDeleteResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsDeleteResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsDeleteResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsDeleteResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UsersRetweetsDeleteResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsDeleteResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsDeleteResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsDeleteResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsDeleteResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsDeleteResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsDeleteResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UsersRetweetsDeleteResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UsersRetweetsDeleteResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UsersRetweetsDeleteResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
