% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_suspend_compliance_schema.R
\docType{class}
\name{UserSuspendComplianceSchema}
\alias{UserSuspendComplianceSchema}
\title{UserSuspendComplianceSchema}
\format{
An \code{R6Class} generator object
}
\description{
UserSuspendComplianceSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user_suspend}}{\link{UserComplianceSchema}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserSuspendComplianceSchema-new}{\code{UserSuspendComplianceSchema$new()}}
\item \href{#method-UserSuspendComplianceSchema-toJSON}{\code{UserSuspendComplianceSchema$toJSON()}}
\item \href{#method-UserSuspendComplianceSchema-fromJSON}{\code{UserSuspendComplianceSchema$fromJSON()}}
\item \href{#method-UserSuspendComplianceSchema-toJSONString}{\code{UserSuspendComplianceSchema$toJSONString()}}
\item \href{#method-UserSuspendComplianceSchema-fromJSONString}{\code{UserSuspendComplianceSchema$fromJSONString()}}
\item \href{#method-UserSuspendComplianceSchema-validateJSON}{\code{UserSuspendComplianceSchema$validateJSON()}}
\item \href{#method-UserSuspendComplianceSchema-toString}{\code{UserSuspendComplianceSchema$toString()}}
\item \href{#method-UserSuspendComplianceSchema-isValid}{\code{UserSuspendComplianceSchema$isValid()}}
\item \href{#method-UserSuspendComplianceSchema-getInvalidFields}{\code{UserSuspendComplianceSchema$getInvalidFields()}}
\item \href{#method-UserSuspendComplianceSchema-clone}{\code{UserSuspendComplianceSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserSuspendComplianceSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserSuspendComplianceSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserSuspendComplianceSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserSuspendComplianceSchema$new(user_suspend, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_suspend}}{user_suspend}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserSuspendComplianceSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserSuspendComplianceSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserSuspendComplianceSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserSuspendComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserSuspendComplianceSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserSuspendComplianceSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserSuspendComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserSuspendComplianceSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserSuspendComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserSuspendComplianceSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserSuspendComplianceSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserSuspendComplianceSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserSuspendComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserSuspendComplianceSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserSuspendComplianceSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserSuspendComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserSuspendComplianceSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserSuspendComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserSuspendComplianceSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserSuspendComplianceSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserSuspendComplianceSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserSuspendComplianceSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserSuspendComplianceSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserSuspendComplianceSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserSuspendComplianceSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserSuspendComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserSuspendComplianceSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserSuspendComplianceSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserSuspendComplianceSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserSuspendComplianceSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserSuspendComplianceSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserSuspendComplianceSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserSuspendComplianceSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserSuspendComplianceSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserSuspendComplianceSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
