% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_undelete_compliance_schema.R
\docType{class}
\name{UserUndeleteComplianceSchema}
\alias{UserUndeleteComplianceSchema}
\title{UserUndeleteComplianceSchema}
\format{
An \code{R6Class} generator object
}
\description{
UserUndeleteComplianceSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user_undelete}}{\link{UserComplianceSchema}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserUndeleteComplianceSchema-new}{\code{UserUndeleteComplianceSchema$new()}}
\item \href{#method-UserUndeleteComplianceSchema-toJSON}{\code{UserUndeleteComplianceSchema$toJSON()}}
\item \href{#method-UserUndeleteComplianceSchema-fromJSON}{\code{UserUndeleteComplianceSchema$fromJSON()}}
\item \href{#method-UserUndeleteComplianceSchema-toJSONString}{\code{UserUndeleteComplianceSchema$toJSONString()}}
\item \href{#method-UserUndeleteComplianceSchema-fromJSONString}{\code{UserUndeleteComplianceSchema$fromJSONString()}}
\item \href{#method-UserUndeleteComplianceSchema-validateJSON}{\code{UserUndeleteComplianceSchema$validateJSON()}}
\item \href{#method-UserUndeleteComplianceSchema-toString}{\code{UserUndeleteComplianceSchema$toString()}}
\item \href{#method-UserUndeleteComplianceSchema-isValid}{\code{UserUndeleteComplianceSchema$isValid()}}
\item \href{#method-UserUndeleteComplianceSchema-getInvalidFields}{\code{UserUndeleteComplianceSchema$getInvalidFields()}}
\item \href{#method-UserUndeleteComplianceSchema-clone}{\code{UserUndeleteComplianceSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUndeleteComplianceSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserUndeleteComplianceSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserUndeleteComplianceSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUndeleteComplianceSchema$new(user_undelete, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_undelete}}{user_undelete}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUndeleteComplianceSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserUndeleteComplianceSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUndeleteComplianceSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserUndeleteComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUndeleteComplianceSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserUndeleteComplianceSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserUndeleteComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUndeleteComplianceSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserUndeleteComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUndeleteComplianceSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserUndeleteComplianceSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUndeleteComplianceSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserUndeleteComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUndeleteComplianceSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserUndeleteComplianceSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserUndeleteComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUndeleteComplianceSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserUndeleteComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUndeleteComplianceSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserUndeleteComplianceSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserUndeleteComplianceSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUndeleteComplianceSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUndeleteComplianceSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserUndeleteComplianceSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUndeleteComplianceSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserUndeleteComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUndeleteComplianceSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserUndeleteComplianceSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUndeleteComplianceSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUndeleteComplianceSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserUndeleteComplianceSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUndeleteComplianceSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUndeleteComplianceSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserUndeleteComplianceSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUndeleteComplianceSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
