% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_unprotect_compliance_schema.R
\docType{class}
\name{UserUnprotectComplianceSchema}
\alias{UserUnprotectComplianceSchema}
\title{UserUnprotectComplianceSchema}
\format{
An \code{R6Class} generator object
}
\description{
UserUnprotectComplianceSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user_unprotect}}{\link{UserComplianceSchema}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserUnprotectComplianceSchema-new}{\code{UserUnprotectComplianceSchema$new()}}
\item \href{#method-UserUnprotectComplianceSchema-toJSON}{\code{UserUnprotectComplianceSchema$toJSON()}}
\item \href{#method-UserUnprotectComplianceSchema-fromJSON}{\code{UserUnprotectComplianceSchema$fromJSON()}}
\item \href{#method-UserUnprotectComplianceSchema-toJSONString}{\code{UserUnprotectComplianceSchema$toJSONString()}}
\item \href{#method-UserUnprotectComplianceSchema-fromJSONString}{\code{UserUnprotectComplianceSchema$fromJSONString()}}
\item \href{#method-UserUnprotectComplianceSchema-validateJSON}{\code{UserUnprotectComplianceSchema$validateJSON()}}
\item \href{#method-UserUnprotectComplianceSchema-toString}{\code{UserUnprotectComplianceSchema$toString()}}
\item \href{#method-UserUnprotectComplianceSchema-isValid}{\code{UserUnprotectComplianceSchema$isValid()}}
\item \href{#method-UserUnprotectComplianceSchema-getInvalidFields}{\code{UserUnprotectComplianceSchema$getInvalidFields()}}
\item \href{#method-UserUnprotectComplianceSchema-clone}{\code{UserUnprotectComplianceSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnprotectComplianceSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnprotectComplianceSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserUnprotectComplianceSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnprotectComplianceSchema$new(user_unprotect, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_unprotect}}{user_unprotect}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnprotectComplianceSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnprotectComplianceSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnprotectComplianceSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserUnprotectComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnprotectComplianceSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnprotectComplianceSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserUnprotectComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnprotectComplianceSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserUnprotectComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnprotectComplianceSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnprotectComplianceSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnprotectComplianceSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserUnprotectComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnprotectComplianceSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnprotectComplianceSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserUnprotectComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnprotectComplianceSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserUnprotectComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnprotectComplianceSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnprotectComplianceSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserUnprotectComplianceSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnprotectComplianceSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnprotectComplianceSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnprotectComplianceSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnprotectComplianceSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserUnprotectComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnprotectComplianceSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnprotectComplianceSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnprotectComplianceSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnprotectComplianceSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnprotectComplianceSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnprotectComplianceSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnprotectComplianceSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnprotectComplianceSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnprotectComplianceSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
