% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_unsuspend_compliance_schema.R
\docType{class}
\name{UserUnsuspendComplianceSchema}
\alias{UserUnsuspendComplianceSchema}
\title{UserUnsuspendComplianceSchema}
\format{
An \code{R6Class} generator object
}
\description{
UserUnsuspendComplianceSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user_unsuspend}}{\link{UserComplianceSchema}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserUnsuspendComplianceSchema-new}{\code{UserUnsuspendComplianceSchema$new()}}
\item \href{#method-UserUnsuspendComplianceSchema-toJSON}{\code{UserUnsuspendComplianceSchema$toJSON()}}
\item \href{#method-UserUnsuspendComplianceSchema-fromJSON}{\code{UserUnsuspendComplianceSchema$fromJSON()}}
\item \href{#method-UserUnsuspendComplianceSchema-toJSONString}{\code{UserUnsuspendComplianceSchema$toJSONString()}}
\item \href{#method-UserUnsuspendComplianceSchema-fromJSONString}{\code{UserUnsuspendComplianceSchema$fromJSONString()}}
\item \href{#method-UserUnsuspendComplianceSchema-validateJSON}{\code{UserUnsuspendComplianceSchema$validateJSON()}}
\item \href{#method-UserUnsuspendComplianceSchema-toString}{\code{UserUnsuspendComplianceSchema$toString()}}
\item \href{#method-UserUnsuspendComplianceSchema-isValid}{\code{UserUnsuspendComplianceSchema$isValid()}}
\item \href{#method-UserUnsuspendComplianceSchema-getInvalidFields}{\code{UserUnsuspendComplianceSchema$getInvalidFields()}}
\item \href{#method-UserUnsuspendComplianceSchema-clone}{\code{UserUnsuspendComplianceSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnsuspendComplianceSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnsuspendComplianceSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserUnsuspendComplianceSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnsuspendComplianceSchema$new(user_unsuspend, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_unsuspend}}{user_unsuspend}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnsuspendComplianceSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnsuspendComplianceSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnsuspendComplianceSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserUnsuspendComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnsuspendComplianceSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnsuspendComplianceSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserUnsuspendComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnsuspendComplianceSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserUnsuspendComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnsuspendComplianceSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnsuspendComplianceSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnsuspendComplianceSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserUnsuspendComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnsuspendComplianceSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnsuspendComplianceSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserUnsuspendComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnsuspendComplianceSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserUnsuspendComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnsuspendComplianceSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnsuspendComplianceSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserUnsuspendComplianceSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnsuspendComplianceSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnsuspendComplianceSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnsuspendComplianceSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnsuspendComplianceSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserUnsuspendComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnsuspendComplianceSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnsuspendComplianceSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnsuspendComplianceSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnsuspendComplianceSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnsuspendComplianceSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnsuspendComplianceSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserUnsuspendComplianceSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserUnsuspendComplianceSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserUnsuspendComplianceSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
