% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_withheld.R
\docType{class}
\name{UserWithheld}
\alias{UserWithheld}
\title{UserWithheld}
\format{
An \code{R6Class} generator object
}
\description{
UserWithheld Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{country_codes}}{list( character )}

\item{\code{scope}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserWithheld-new}{\code{UserWithheld$new()}}
\item \href{#method-UserWithheld-toJSON}{\code{UserWithheld$toJSON()}}
\item \href{#method-UserWithheld-fromJSON}{\code{UserWithheld$fromJSON()}}
\item \href{#method-UserWithheld-toJSONString}{\code{UserWithheld$toJSONString()}}
\item \href{#method-UserWithheld-fromJSONString}{\code{UserWithheld$fromJSONString()}}
\item \href{#method-UserWithheld-validateJSON}{\code{UserWithheld$validateJSON()}}
\item \href{#method-UserWithheld-toString}{\code{UserWithheld$toString()}}
\item \href{#method-UserWithheld-isValid}{\code{UserWithheld$isValid()}}
\item \href{#method-UserWithheld-getInvalidFields}{\code{UserWithheld$getInvalidFields()}}
\item \href{#method-UserWithheld-clone}{\code{UserWithheld$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheld-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheld-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserWithheld class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheld$new(country_codes, scope = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{country_codes}}{Provides a list of countries where this content is not available.}

\item{\code{scope}}{Indicates that the content being withheld is a `user`.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheld-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheld-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheld$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserWithheld in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheld-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheld-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserWithheld
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheld$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserWithheld
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheld-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheld-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheld$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserWithheld in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheld-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheld-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserWithheld
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheld$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserWithheld
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheld-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheld-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserWithheld and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheld$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheld-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheld-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheld$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserWithheld
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheld-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheld-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheld$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheld-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheld-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheld$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheld-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheld-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheld$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
