% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_withheld_compliance_schema.R
\docType{class}
\name{UserWithheldComplianceSchema}
\alias{UserWithheldComplianceSchema}
\title{UserWithheldComplianceSchema}
\format{
An \code{R6Class} generator object
}
\description{
UserWithheldComplianceSchema Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user_withheld}}{\link{UserTakedownComplianceSchema}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserWithheldComplianceSchema-new}{\code{UserWithheldComplianceSchema$new()}}
\item \href{#method-UserWithheldComplianceSchema-toJSON}{\code{UserWithheldComplianceSchema$toJSON()}}
\item \href{#method-UserWithheldComplianceSchema-fromJSON}{\code{UserWithheldComplianceSchema$fromJSON()}}
\item \href{#method-UserWithheldComplianceSchema-toJSONString}{\code{UserWithheldComplianceSchema$toJSONString()}}
\item \href{#method-UserWithheldComplianceSchema-fromJSONString}{\code{UserWithheldComplianceSchema$fromJSONString()}}
\item \href{#method-UserWithheldComplianceSchema-validateJSON}{\code{UserWithheldComplianceSchema$validateJSON()}}
\item \href{#method-UserWithheldComplianceSchema-toString}{\code{UserWithheldComplianceSchema$toString()}}
\item \href{#method-UserWithheldComplianceSchema-isValid}{\code{UserWithheldComplianceSchema$isValid()}}
\item \href{#method-UserWithheldComplianceSchema-getInvalidFields}{\code{UserWithheldComplianceSchema$getInvalidFields()}}
\item \href{#method-UserWithheldComplianceSchema-clone}{\code{UserWithheldComplianceSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheldComplianceSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheldComplianceSchema-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UserWithheldComplianceSchema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheldComplianceSchema$new(user_withheld, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_withheld}}{user_withheld}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheldComplianceSchema-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheldComplianceSchema-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheldComplianceSchema$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserWithheldComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheldComplianceSchema-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheldComplianceSchema-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UserWithheldComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheldComplianceSchema$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserWithheldComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheldComplianceSchema-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheldComplianceSchema-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheldComplianceSchema$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UserWithheldComplianceSchema in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheldComplianceSchema-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheldComplianceSchema-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UserWithheldComplianceSchema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheldComplianceSchema$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UserWithheldComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheldComplianceSchema-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheldComplianceSchema-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UserWithheldComplianceSchema and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheldComplianceSchema$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheldComplianceSchema-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheldComplianceSchema-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheldComplianceSchema$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UserWithheldComplianceSchema
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheldComplianceSchema-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheldComplianceSchema-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheldComplianceSchema$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheldComplianceSchema-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheldComplianceSchema-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheldComplianceSchema$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserWithheldComplianceSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserWithheldComplianceSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserWithheldComplianceSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
