% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_all_of.R
\docType{class}
\name{VideoAllOf}
\alias{VideoAllOf}
\title{VideoAllOf}
\format{
An \code{R6Class} generator object
}
\description{
VideoAllOf Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{duration_ms}}{integer [optional]}

\item{\code{non_public_metrics}}{\link{VideoAllOfNonPublicMetrics} [optional]}

\item{\code{organic_metrics}}{\link{VideoAllOfOrganicMetrics} [optional]}

\item{\code{preview_image_url}}{character [optional]}

\item{\code{promoted_metrics}}{\link{VideoAllOfPromotedMetrics} [optional]}

\item{\code{public_metrics}}{\link{VideoAllOfPublicMetrics} [optional]}

\item{\code{variants}}{list( \link{Variant} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VideoAllOf-new}{\code{VideoAllOf$new()}}
\item \href{#method-VideoAllOf-toJSON}{\code{VideoAllOf$toJSON()}}
\item \href{#method-VideoAllOf-fromJSON}{\code{VideoAllOf$fromJSON()}}
\item \href{#method-VideoAllOf-toJSONString}{\code{VideoAllOf$toJSONString()}}
\item \href{#method-VideoAllOf-fromJSONString}{\code{VideoAllOf$fromJSONString()}}
\item \href{#method-VideoAllOf-validateJSON}{\code{VideoAllOf$validateJSON()}}
\item \href{#method-VideoAllOf-toString}{\code{VideoAllOf$toString()}}
\item \href{#method-VideoAllOf-isValid}{\code{VideoAllOf$isValid()}}
\item \href{#method-VideoAllOf-getInvalidFields}{\code{VideoAllOf$getInvalidFields()}}
\item \href{#method-VideoAllOf-clone}{\code{VideoAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new VideoAllOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOf$new(
  duration_ms = NULL,
  non_public_metrics = NULL,
  organic_metrics = NULL,
  preview_image_url = NULL,
  promoted_metrics = NULL,
  public_metrics = NULL,
  variants = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{duration_ms}}{duration_ms}

\item{\code{non_public_metrics}}{non_public_metrics}

\item{\code{organic_metrics}}{organic_metrics}

\item{\code{preview_image_url}}{preview_image_url}

\item{\code{promoted_metrics}}{promoted_metrics}

\item{\code{public_metrics}}{public_metrics}

\item{\code{variants}}{An array of all available variants of the media.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
VideoAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of VideoAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of VideoAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
VideoAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of VideoAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of VideoAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to VideoAllOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of VideoAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOf-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOf-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOf$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOf-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOf-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOf$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
