% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_all_of_non_public_metrics.R
\docType{class}
\name{VideoAllOfNonPublicMetrics}
\alias{VideoAllOfNonPublicMetrics}
\title{VideoAllOfNonPublicMetrics}
\format{
An \code{R6Class} generator object
}
\description{
VideoAllOfNonPublicMetrics Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{playback_0_count}}{integer [optional]}

\item{\code{playback_100_count}}{integer [optional]}

\item{\code{playback_25_count}}{integer [optional]}

\item{\code{playback_50_count}}{integer [optional]}

\item{\code{playback_75_count}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VideoAllOfNonPublicMetrics-new}{\code{VideoAllOfNonPublicMetrics$new()}}
\item \href{#method-VideoAllOfNonPublicMetrics-toJSON}{\code{VideoAllOfNonPublicMetrics$toJSON()}}
\item \href{#method-VideoAllOfNonPublicMetrics-fromJSON}{\code{VideoAllOfNonPublicMetrics$fromJSON()}}
\item \href{#method-VideoAllOfNonPublicMetrics-toJSONString}{\code{VideoAllOfNonPublicMetrics$toJSONString()}}
\item \href{#method-VideoAllOfNonPublicMetrics-fromJSONString}{\code{VideoAllOfNonPublicMetrics$fromJSONString()}}
\item \href{#method-VideoAllOfNonPublicMetrics-validateJSON}{\code{VideoAllOfNonPublicMetrics$validateJSON()}}
\item \href{#method-VideoAllOfNonPublicMetrics-toString}{\code{VideoAllOfNonPublicMetrics$toString()}}
\item \href{#method-VideoAllOfNonPublicMetrics-isValid}{\code{VideoAllOfNonPublicMetrics$isValid()}}
\item \href{#method-VideoAllOfNonPublicMetrics-getInvalidFields}{\code{VideoAllOfNonPublicMetrics$getInvalidFields()}}
\item \href{#method-VideoAllOfNonPublicMetrics-clone}{\code{VideoAllOfNonPublicMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfNonPublicMetrics-new"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfNonPublicMetrics-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new VideoAllOfNonPublicMetrics class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfNonPublicMetrics$new(
  playback_0_count = NULL,
  playback_100_count = NULL,
  playback_25_count = NULL,
  playback_50_count = NULL,
  playback_75_count = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{playback_0_count}}{Number of users who made it through 0\% of the video.}

\item{\code{playback_100_count}}{Number of users who made it through 100\% of the video.}

\item{\code{playback_25_count}}{Number of users who made it through 25\% of the video.}

\item{\code{playback_50_count}}{Number of users who made it through 50\% of the video.}

\item{\code{playback_75_count}}{Number of users who made it through 75\% of the video.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfNonPublicMetrics-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfNonPublicMetrics-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfNonPublicMetrics$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
VideoAllOfNonPublicMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfNonPublicMetrics-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfNonPublicMetrics-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of VideoAllOfNonPublicMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfNonPublicMetrics$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of VideoAllOfNonPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfNonPublicMetrics-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfNonPublicMetrics-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfNonPublicMetrics$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
VideoAllOfNonPublicMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfNonPublicMetrics-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfNonPublicMetrics-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of VideoAllOfNonPublicMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfNonPublicMetrics$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of VideoAllOfNonPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfNonPublicMetrics-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfNonPublicMetrics-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to VideoAllOfNonPublicMetrics and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfNonPublicMetrics$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfNonPublicMetrics-toString"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfNonPublicMetrics-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfNonPublicMetrics$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of VideoAllOfNonPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfNonPublicMetrics-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfNonPublicMetrics-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfNonPublicMetrics$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfNonPublicMetrics-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfNonPublicMetrics-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfNonPublicMetrics$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfNonPublicMetrics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfNonPublicMetrics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfNonPublicMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
