% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_all_of_organic_metrics.R
\docType{class}
\name{VideoAllOfOrganicMetrics}
\alias{VideoAllOfOrganicMetrics}
\title{VideoAllOfOrganicMetrics}
\format{
An \code{R6Class} generator object
}
\description{
VideoAllOfOrganicMetrics Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{playback_0_count}}{integer [optional]}

\item{\code{playback_100_count}}{integer [optional]}

\item{\code{playback_25_count}}{integer [optional]}

\item{\code{playback_50_count}}{integer [optional]}

\item{\code{playback_75_count}}{integer [optional]}

\item{\code{view_count}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VideoAllOfOrganicMetrics-new}{\code{VideoAllOfOrganicMetrics$new()}}
\item \href{#method-VideoAllOfOrganicMetrics-toJSON}{\code{VideoAllOfOrganicMetrics$toJSON()}}
\item \href{#method-VideoAllOfOrganicMetrics-fromJSON}{\code{VideoAllOfOrganicMetrics$fromJSON()}}
\item \href{#method-VideoAllOfOrganicMetrics-toJSONString}{\code{VideoAllOfOrganicMetrics$toJSONString()}}
\item \href{#method-VideoAllOfOrganicMetrics-fromJSONString}{\code{VideoAllOfOrganicMetrics$fromJSONString()}}
\item \href{#method-VideoAllOfOrganicMetrics-validateJSON}{\code{VideoAllOfOrganicMetrics$validateJSON()}}
\item \href{#method-VideoAllOfOrganicMetrics-toString}{\code{VideoAllOfOrganicMetrics$toString()}}
\item \href{#method-VideoAllOfOrganicMetrics-isValid}{\code{VideoAllOfOrganicMetrics$isValid()}}
\item \href{#method-VideoAllOfOrganicMetrics-getInvalidFields}{\code{VideoAllOfOrganicMetrics$getInvalidFields()}}
\item \href{#method-VideoAllOfOrganicMetrics-clone}{\code{VideoAllOfOrganicMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfOrganicMetrics-new"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfOrganicMetrics-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new VideoAllOfOrganicMetrics class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfOrganicMetrics$new(
  playback_0_count = NULL,
  playback_100_count = NULL,
  playback_25_count = NULL,
  playback_50_count = NULL,
  playback_75_count = NULL,
  view_count = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{playback_0_count}}{Number of users who made it through 0\% of the video.}

\item{\code{playback_100_count}}{Number of users who made it through 100\% of the video.}

\item{\code{playback_25_count}}{Number of users who made it through 25\% of the video.}

\item{\code{playback_50_count}}{Number of users who made it through 50\% of the video.}

\item{\code{playback_75_count}}{Number of users who made it through 75\% of the video.}

\item{\code{view_count}}{Number of times this video has been viewed.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfOrganicMetrics-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfOrganicMetrics-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfOrganicMetrics$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
VideoAllOfOrganicMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfOrganicMetrics-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfOrganicMetrics-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of VideoAllOfOrganicMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfOrganicMetrics$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of VideoAllOfOrganicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfOrganicMetrics-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfOrganicMetrics-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfOrganicMetrics$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
VideoAllOfOrganicMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfOrganicMetrics-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfOrganicMetrics-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of VideoAllOfOrganicMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfOrganicMetrics$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of VideoAllOfOrganicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfOrganicMetrics-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfOrganicMetrics-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to VideoAllOfOrganicMetrics and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfOrganicMetrics$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfOrganicMetrics-toString"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfOrganicMetrics-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfOrganicMetrics$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of VideoAllOfOrganicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfOrganicMetrics-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfOrganicMetrics-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfOrganicMetrics$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfOrganicMetrics-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfOrganicMetrics-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfOrganicMetrics$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfOrganicMetrics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfOrganicMetrics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfOrganicMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
