% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_all_of_public_metrics.R
\docType{class}
\name{VideoAllOfPublicMetrics}
\alias{VideoAllOfPublicMetrics}
\title{VideoAllOfPublicMetrics}
\format{
An \code{R6Class} generator object
}
\description{
VideoAllOfPublicMetrics Class
}
\details{
Twitter API v2

Twitter API v2 available endpoints

The version of the OpenAPI document: 2.49
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{view_count}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VideoAllOfPublicMetrics-new}{\code{VideoAllOfPublicMetrics$new()}}
\item \href{#method-VideoAllOfPublicMetrics-toJSON}{\code{VideoAllOfPublicMetrics$toJSON()}}
\item \href{#method-VideoAllOfPublicMetrics-fromJSON}{\code{VideoAllOfPublicMetrics$fromJSON()}}
\item \href{#method-VideoAllOfPublicMetrics-toJSONString}{\code{VideoAllOfPublicMetrics$toJSONString()}}
\item \href{#method-VideoAllOfPublicMetrics-fromJSONString}{\code{VideoAllOfPublicMetrics$fromJSONString()}}
\item \href{#method-VideoAllOfPublicMetrics-validateJSON}{\code{VideoAllOfPublicMetrics$validateJSON()}}
\item \href{#method-VideoAllOfPublicMetrics-toString}{\code{VideoAllOfPublicMetrics$toString()}}
\item \href{#method-VideoAllOfPublicMetrics-isValid}{\code{VideoAllOfPublicMetrics$isValid()}}
\item \href{#method-VideoAllOfPublicMetrics-getInvalidFields}{\code{VideoAllOfPublicMetrics$getInvalidFields()}}
\item \href{#method-VideoAllOfPublicMetrics-clone}{\code{VideoAllOfPublicMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfPublicMetrics-new"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfPublicMetrics-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new VideoAllOfPublicMetrics class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfPublicMetrics$new(view_count = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{view_count}}{Number of times this video has been viewed.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfPublicMetrics-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfPublicMetrics-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfPublicMetrics$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
VideoAllOfPublicMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfPublicMetrics-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfPublicMetrics-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of VideoAllOfPublicMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfPublicMetrics$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of VideoAllOfPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfPublicMetrics-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfPublicMetrics-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfPublicMetrics$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
VideoAllOfPublicMetrics in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfPublicMetrics-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfPublicMetrics-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of VideoAllOfPublicMetrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfPublicMetrics$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of VideoAllOfPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfPublicMetrics-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfPublicMetrics-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to VideoAllOfPublicMetrics and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfPublicMetrics$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfPublicMetrics-toString"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfPublicMetrics-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfPublicMetrics$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of VideoAllOfPublicMetrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfPublicMetrics-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfPublicMetrics-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfPublicMetrics$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfPublicMetrics-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfPublicMetrics-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfPublicMetrics$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VideoAllOfPublicMetrics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VideoAllOfPublicMetrics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoAllOfPublicMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
